--- backend-v2/public/assets/index-aa718ec1.js
@@ -1,51 +1,0 @@
-(function(){const e=document.createElement("link").relList;if(e&&e.supports&&e.supports("modulepreload"))return;for(const i of document.querySelectorAll('link[rel="modulepreload"]'))r(i);new MutationObserver(i=>{for(const s of i)if(s.type==="childList")for(const o of s.addedNodes)o.tagName==="LINK"&&o.rel==="modulepreload"&&r(o)}).observe(document,{childList:!0,subtree:!0});function n(i){const s={};return i.integrity&&(s.integrity=i.integrity),i.referrerPolicy&&(s.referrerPolicy=i.referrerPolicy),i.crossOrigin==="use-credentials"?s.credentials="include":i.crossOrigin==="anonymous"?s.credentials="omit":s.credentials="same-origin",s}function r(i){if(i.ep)return;i.ep=!0;const s=n(i);fetch(i.href,s)}})();function Q(){}const ri=t=>t;function Ei(t,e){for(const n in e)t[n]=e[n];return t}function Oo(t){return t()}function hs(){return Object.create(null)}function je(t){t.forEach(Oo)}function wn(t){return typeof t=="function"}function ve(t,e){return t!=t?e==e:t!==e||t&&typeof t=="object"||typeof t=="function"}let Ar;function qn(t,e){return Ar||(Ar=document.createElement("a")),Ar.href=e,t===Ar.href}function Gl(t){return Object.keys(t).length===0}function Ql(t,...e){if(t==null)return Q;const n=t.subscribe(...e);return n.unsubscribe?()=>n.unsubscribe():n}function lt(t,e,n){t.$$.on_destroy.push(Ql(e,n))}function Yl(t,e,n,r){if(t){const i=Mo(t,e,n,r);return t[0](i)}}function Mo(t,e,n,r){return t[1]&&r?Ei(n.ctx.slice(),t[1](r(e))):n.ctx}function Xl(t,e,n,r){if(t[2]&&r){const i=t[2](r(n));if(e.dirty===void 0)return i;if(typeof i=="object"){const s=[],o=Math.max(e.dirty.length,i.length);for(let l=0;l<o;l+=1)s[l]=e.dirty[l]|i[l];return s}return e.dirty|i}return e.dirty}function Kl(t,e,n,r,i,s){if(i){const o=Mo(e,n,r,s);t.p(o,i)}}function Jl(t){if(t.ctx.length>32){const e=[],n=t.ctx.length/32;for(let r=0;r<n;r++)e[r]=-1;return e}return-1}function ps(t,e,n){return t.set(n),e}function Ro(t){return t&&wn(t.destroy)?t.destroy:Q}const No=typeof window<"u";let Zi=No?()=>window.performance.now():()=>Date.now(),ji=No?t=>requestAnimationFrame(t):Q;const In=new Set;function qo(t){In.forEach(e=>{e.c(t)||(In.delete(e),e.f())}),In.size!==0&&ji(qo)}function Hi(t){let e;return In.size===0&&ji(qo),{promise:new Promise(n=>{In.add(e={c:t,f:n})}),abort(){In.delete(e)}}}function g(t,e){t.appendChild(e)}function Fo(t){if(!t)return document;const e=t.getRootNode?t.getRootNode():t.ownerDocument;return e&&e.host?e:t.ownerDocument}function ea(t){const e=D("style");return ta(Fo(t),e),e.sheet}function ta(t,e){return g(t.head||t,e),e.sheet}function W(t,e,n){t.insertBefore(e,n||null)}function V(t){t.parentNode&&t.parentNode.removeChild(t)}function Kn(t,e){for(let n=0;n<t.length;n+=1)t[n]&&t[n].d(e)}function D(t){return document.createElement(t)}function ce(t){return document.createElementNS("http://www.w3.org/2000/svg",t)}function te(t){return document.createTextNode(t)}function U(){return te(" ")}function Lt(){return te("")}function J(t,e,n,r){return t.addEventListener(e,n,r),()=>t.removeEventListener(e,n,r)}function m(t,e,n){n==null?t.removeAttribute(e):t.getAttribute(e)!==n&&t.setAttribute(e,n)}function na(t){let e;return{p(...n){e=n,e.forEach(r=>t.push(r))},r(){e.forEach(n=>t.splice(t.indexOf(n),1))}}}function Rt(t){return t===""?null:+t}function ra(t){return Array.from(t.childNodes)}function Ie(t,e){e=""+e,t.data!==e&&(t.data=e)}function Ze(t,e){t.value=e??""}function Ai(t,e,n,r){n===null?t.style.removeProperty(e):t.style.setProperty(e,n,r?"important":"")}function Fn(t,e,n){for(let r=0;r<t.options.length;r+=1){const i=t.options[r];if(i.__value===e){i.selected=!0;return}}(!n||e!==void 0)&&(t.selectedIndex=-1)}function Bi(t){const e=t.querySelector(":checked");return e&&e.__value}function Or(t,e,n){t.classList[n?"add":"remove"](e)}function ia(t,e,{bubbles:n=!1,cancelable:r=!1}={}){const i=document.createEvent("CustomEvent");return i.initCustomEvent(t,n,r,e),i}const Mr=new Map;let Rr=0;function sa(t){let e=5381,n=t.length;for(;n--;)e=(e<<5)-e^t.charCodeAt(n);return e>>>0}function oa(t,e){const n={stylesheet:ea(e),rules:{}};return Mr.set(t,n),n}function Ti(t,e,n,r,i,s,o,l=0){const a=16.666/r;let f=`{
-`;for(let k=0;k<=1;k+=a){const _=e+(n-e)*s(k);f+=k*100+`%{${o(_,1-_)}}
-`}const u=f+`100% {${o(n,1-n)}}
-}`,d=`__svelte_${sa(u)}_${l}`,c=Fo(t),{stylesheet:h,rules:p}=Mr.get(c)||oa(c,t);p[d]||(p[d]=!0,h.insertRule(`@keyframes ${d} ${u}`,h.cssRules.length));const y=t.style.animation||"";return t.style.animation=`${y?`${y}, `:""}${d} ${r}ms linear ${i}ms 1 both`,Rr+=1,d}function Lo(t,e){const n=(t.style.animation||"").split(", "),r=n.filter(e?s=>s.indexOf(e)<0:s=>s.indexOf("__svelte")===-1),i=n.length-r.length;i&&(t.style.animation=r.join(", "),Rr-=i,Rr||la())}function la(){ji(()=>{Rr||(Mr.forEach(t=>{const{ownerNode:e}=t.stylesheet;e&&V(e)}),Mr.clear())})}function mr(t,e,n,r){if(!e)return Q;const i=t.getBoundingClientRect();if(e.left===i.left&&e.right===i.right&&e.top===i.top&&e.bottom===i.bottom)return Q;const{delay:s=0,duration:o=300,easing:l=ri,start:a=Zi()+s,end:f=a+o,tick:u=Q,css:d}=n(t,{from:e,to:i},r);let c=!0,h=!1,p;function y(){d&&(p=Ti(t,0,1,o,s,l,d)),s||(h=!0)}function k(){d&&Lo(t,p),c=!1}return Hi(_=>{if(!h&&_>=a&&(h=!0),h&&_>=f&&(u(1,0),k()),!c)return!1;if(h){const C=_-a,T=0+1*l(C/o);u(T,1-T)}return!0}),y(),u(0,1),k}function gr(t){const e=getComputedStyle(t);if(e.position!=="absolute"&&e.position!=="fixed"){const{width:n,height:r}=e,i=t.getBoundingClientRect();t.style.position="absolute",t.style.width=n,t.style.height=r,aa(t,i)}}function aa(t,e){const n=t.getBoundingClientRect();if(e.left!==n.left||e.top!==n.top){const r=getComputedStyle(t),i=r.transform==="none"?"":r.transform;t.style.transform=`${i} translate(${e.left-n.left}px, ${e.top-n.top}px)`}}let Jn;function Xn(t){Jn=t}function Po(){if(!Jn)throw new Error("Function called outside component initialization");return Jn}function Uo(t){Po().$$.on_mount.push(t)}function ua(t){Po().$$.on_destroy.push(t)}function ca(t,e){const n=t.$$.callbacks[e.type];n&&n.slice().forEach(r=>r.call(this,e))}const Tn=[],Ke=[];let $n=[];const Ci=[],fa=Promise.resolve();let xi=!1;function da(){xi||(xi=!0,fa.then(zo))}function Le(t){$n.push(t)}function at(t){Ci.push(t)}const _i=new Set;let En=0;function zo(){if(En!==0)return;const t=Jn;do{try{for(;En<Tn.length;){const e=Tn[En];En++,Xn(e),ha(e.$$)}}catch(e){throw Tn.length=0,En=0,e}for(Xn(null),Tn.length=0,En=0;Ke.length;)Ke.pop()();for(let e=0;e<$n.length;e+=1){const n=$n[e];_i.has(n)||(_i.add(n),n())}$n.length=0}while(Tn.length);for(;Ci.length;)Ci.pop()();xi=!1,_i.clear(),Xn(t)}function ha(t){if(t.fragment!==null){t.update(),je(t.before_update);const e=t.dirty;t.dirty=[-1],t.fragment&&t.fragment.p(t.ctx,e),t.after_update.forEach(Le)}}function pa(t){const e=[],n=[];$n.forEach(r=>t.indexOf(r)===-1?e.push(r):n.push(r)),n.forEach(r=>r()),$n=e}let Gn;function ma(){return Gn||(Gn=Promise.resolve(),Gn.then(()=>{Gn=null})),Gn}function vi(t,e,n){t.dispatchEvent(ia(`${e?"intro":"outro"}${n}`))}const Cr=new Set;let Ut;function qe(){Ut={r:0,c:[],p:Ut}}function Fe(){Ut.r||je(Ut.c),Ut=Ut.p}function B(t,e){t&&t.i&&(Cr.delete(t),t.i(e))}function G(t,e,n,r){if(t&&t.o){if(Cr.has(t))return;Cr.add(t),Ut.c.push(()=>{Cr.delete(t),r&&(n&&t.d(1),r())}),t.o(e)}else r&&r()}const ga={duration:0};function Ae(t,e,n,r){const i={direction:"both"};let s=e(t,n,i),o=r?0:1,l=null,a=null,f=null;function u(){f&&Lo(t,f)}function d(h,p){const y=h.b-o;return p*=Math.abs(y),{a:o,b:h.b,d:y,duration:p,start:h.start,end:h.start+p,group:h.group}}function c(h){const{delay:p=0,duration:y=300,easing:k=ri,tick:_=Q,css:C}=s||ga,T={start:Zi()+p,b:h};h||(T.group=Ut,Ut.r+=1),l||a?a=T:(C&&(u(),f=Ti(t,o,h,y,p,k,C)),h&&_(0,1),l=d(T,y),Le(()=>vi(t,h,"start")),Hi(S=>{if(a&&S>a.start&&(l=d(a,y),a=null,vi(t,l.b,"start"),C&&(u(),f=Ti(t,o,l.b,l.duration,0,k,s.css))),l){if(S>=l.end)_(o=l.b,1-o),vi(t,l.b,"end"),a||(l.b?u():--l.group.r||je(l.group.c)),l=null;else if(S>=l.start){const $=S-l.start;o=l.a+l.d*k($/l.duration),_(o,1-o)}}return!!(l||a)}))}return{run(h){wn(s)?ma().then(()=>{s=s(i),c(h)}):c(h)},end(){u(),l=a=null}}}function _a(t,e){t.d(1),e.delete(t.key)}function va(t,e){G(t,1,1,()=>{e.delete(t.key)})}function ya(t,e){t.f(),_a(t,e)}function ii(t,e){t.f(),va(t,e)}function _r(t,e,n,r,i,s,o,l,a,f,u,d){let c=t.length,h=s.length,p=c;const y={};for(;p--;)y[t[p].key]=p;const k=[],_=new Map,C=new Map,T=[];for(p=h;p--;){const x=d(i,s,p),v=n(x);let b=o.get(v);b?r&&T.push(()=>b.p(x,e)):(b=f(v,x),b.c()),_.set(v,k[p]=b),v in y&&C.set(v,Math.abs(p-y[v]))}const S=new Set,$=new Set;function w(x){B(x,1),x.m(l,u),o.set(x.key,x),u=x.first,h--}for(;c&&h;){const x=k[h-1],v=t[c-1],b=x.key,I=v.key;x===v?(u=x.first,c--,h--):_.has(I)?!o.has(b)||S.has(b)?w(x):$.has(I)?c--:C.get(b)>C.get(I)?($.add(b),w(x)):(S.add(I),c--):(a(v,o),c--)}for(;c--;){const x=t[c];_.has(x.key)||a(x,o)}for(;h;)w(k[h-1]);return je(T),k}const ba=["allowfullscreen","allowpaymentrequest","async","autofocus","autoplay","checked","controls","default","defer","disabled","formnovalidate","hidden","inert","ismap","loop","multiple","muted","nomodule","novalidate","open","playsinline","readonly","required","reversed","selected"];[...ba];function ut(t,e,n){const r=t.$$.props[e];r!==void 0&&(t.$$.bound[r]=n,n(t.$$.ctx[r]))}function le(t){t&&t.c()}function se(t,e,n,r){const{fragment:i,after_update:s}=t.$$;i&&i.m(e,n),r||Le(()=>{const o=t.$$.on_mount.map(Oo).filter(wn);t.$$.on_destroy?t.$$.on_destroy.push(...o):je(o),t.$$.on_mount=[]}),s.forEach(Le)}function oe(t,e){const n=t.$$;n.fragment!==null&&(pa(n.after_update),je(n.on_destroy),n.fragment&&n.fragment.d(e),n.on_destroy=n.fragment=null,n.ctx=[])}function wa(t,e){t.$$.dirty[0]===-1&&(Tn.push(t),da(),t.$$.dirty.fill(0)),t.$$.dirty[e/31|0]|=1<<e%31}function ye(t,e,n,r,i,s,o,l=[-1]){const a=Jn;Xn(t);const f=t.$$={fragment:null,ctx:[],props:s,update:Q,not_equal:i,bound:hs(),on_mount:[],on_destroy:[],on_disconnect:[],before_update:[],after_update:[],context:new Map(e.context||(a?a.$$.context:[])),callbacks:hs(),dirty:l,skip_bound:!1,root:e.target||a.$$.root};o&&o(f.root);let u=!1;if(f.ctx=n?n(t,e.props||{},(d,c,...h)=>{const p=h.length?h[0]:c;return f.ctx&&i(f.ctx[d],f.ctx[d]=p)&&(!f.skip_bound&&f.bound[d]&&f.bound[d](p),u&&wa(t,d)),c}):[],f.update(),u=!0,je(f.before_update),f.fragment=r?r(f.ctx):!1,e.target){if(e.hydrate){const d=ra(e.target);f.fragment&&f.fragment.l(d),d.forEach(V)}else f.fragment&&f.fragment.c();e.intro&&B(t.$$.fragment),se(t,e.target,e.anchor,e.customElement),zo()}Xn(a)}class be{$destroy(){oe(this,1),this.$destroy=Q}$on(e,n){if(!wn(n))return Q;const r=this.$$.callbacks[e]||(this.$$.callbacks[e]=[]);return r.push(n),()=>{const i=r.indexOf(n);i!==-1&&r.splice(i,1)}}$set(e){this.$$set&&!Gl(e)&&(this.$$.skip_bound=!0,this.$$set(e),this.$$.skip_bound=!1)}}const An=[];function Bt(t,e=Q){let n;const r=new Set;function i(l){if(ve(t,l)&&(t=l,n)){const a=!An.length;for(const f of r)f[1](),An.push(f,t);if(a){for(let f=0;f<An.length;f+=2)An[f][0](An[f+1]);An.length=0}}}function s(l){i(l(t))}function o(l,a=Q){const f=[l,a];return r.add(f),r.size===1&&(n=e(i)||Q),l(t),()=>{r.delete(f),r.size===0&&n&&(n(),n=null)}}return{set:i,update:s,subscribe:o}}var Yn=(t=>(t[t.Home=0]="Home",t[t.Connect=1]="Connect",t[t.Create=2]="Create",t[t.Game=3]="Game",t))(Yn||{});const vr=Bt({ty:0});function yr(){vr.set({ty:0})}function ka(){vr.set({ty:2})}function Ea(){vr.set({ty:1})}function Zo(t){vr.set({ty:3,gameData:t})}function jo(t){const e=t-1;return e*e*e+1}function Nt(t,{delay:e=0,duration:n=400,easing:r=ri}={}){const i=+getComputedStyle(t).opacity;return{delay:e,duration:n,easing:r,css:s=>`opacity: ${s*i}`}}function Me(t,{delay:e=0,duration:n=400,easing:r=jo,axis:i="y"}={}){const s=getComputedStyle(t),o=+s.opacity,l=i==="y"?"height":"width",a=parseFloat(s[l]),f=i==="y"?["top","bottom"]:["left","right"],u=f.map(_=>`${_[0].toUpperCase()}${_.slice(1)}`),d=parseFloat(s[`padding${u[0]}`]),c=parseFloat(s[`padding${u[1]}`]),h=parseFloat(s[`margin${u[0]}`]),p=parseFloat(s[`margin${u[1]}`]),y=parseFloat(s[`border${u[0]}Width`]),k=parseFloat(s[`border${u[1]}Width`]);return{delay:e,duration:n,easing:r,css:_=>`overflow: hidden;opacity: ${Math.min(_*20,1)*o};${l}: ${_*a}px;padding-${f[0]}: ${_*d}px;padding-${f[1]}: ${_*c}px;margin-${f[0]}: ${_*h}px;margin-${f[1]}: ${_*p}px;border-${f[0]}-width: ${_*y}px;border-${f[1]}-width: ${_*k}px;`}}var Ho=(t=>(t[t.Error=0]="Error",t[t.Confirm=1]="Confirm",t))(Ho||{});const si=Bt(null);function yi(t){si.update(e=>(e!==null&&e.ty===1&&e.callback(t),null))}function Vi(t,e){return new Promise(n=>{si.set({ty:1,title:t,message:e,callback:n})})}function Qt(t,e){si.set({ty:0,title:t,message:e})}function ms(t){let e,n,r,i=t[0].title+"",s,o,l,a=t[0].message+"",f,u,d,c,h,p;function y(C,T){return C[0].ty===Ho.Error?Ta:Aa}let k=y(t),_=k(t);return{c(){e=D("div"),n=D("div"),r=D("h1"),s=te(i),o=U(),l=D("p"),f=te(a),u=U(),d=D("div"),_.c(),m(r,"class","dialog__title svelte-f5490n"),m(l,"class","dialog__msg svelte-f5490n"),m(d,"class","btn-row btn-row--fill"),m(n,"class","dialog svelte-f5490n"),m(e,"class","floating-wrapper")},m(C,T){W(C,e,T),g(e,n),g(n,r),g(r,s),g(n,o),g(n,l),g(l,f),g(n,u),g(n,d),_.m(d,null),p=!0},p(C,T){(!p||T&1)&&i!==(i=C[0].title+"")&&Ie(s,i),(!p||T&1)&&a!==(a=C[0].message+"")&&Ie(f,a),k===(k=y(C))&&_?_.p(C,T):(_.d(1),_=k(C),_&&(_.c(),_.m(d,null)))},i(C){p||(Le(()=>{p&&(c||(c=Ae(n,Me,{duration:200},!0)),c.run(1))}),Le(()=>{p&&(h||(h=Ae(e,Nt,{duration:200},!0)),h.run(1))}),p=!0)},o(C){c||(c=Ae(n,Me,{duration:200},!1)),c.run(0),h||(h=Ae(e,Nt,{duration:200},!1)),h.run(0),p=!1},d(C){C&&V(e),_.d(),C&&c&&c.end(),C&&h&&h.end()}}}function Aa(t){let e,n,r,i,s;return{c(){e=D("button"),e.textContent="Confirm",n=U(),r=D("button"),r.textContent="Cancel",m(e,"class","btn btn--surface"),m(r,"class","btn btn--surface")},m(o,l){W(o,e,l),W(o,n,l),W(o,r,l),i||(s=[J(e,"click",t[2]),J(r,"click",t[3])],i=!0)},p:Q,d(o){o&&V(e),o&&V(n),o&&V(r),i=!1,je(s)}}}function Ta(t){let e,n,r;return{c(){e=D("button"),e.textContent="Close",m(e,"class","btn btn--surface")},m(i,s){W(i,e,s),n||(r=J(e,"click",t[1]),n=!0)},p:Q,d(i){i&&V(e),n=!1,r()}}}function Ca(t){let e,n,r=t[0]&&ms(t);return{c(){r&&r.c(),e=Lt()},m(i,s){r&&r.m(i,s),W(i,e,s),n=!0},p(i,[s]){i[0]?r?(r.p(i,s),s&1&&B(r,1)):(r=ms(i),r.c(),B(r,1),r.m(e.parentNode,e)):r&&(qe(),G(r,1,1,()=>{r=null}),Fe())},i(i){n||(B(r),n=!0)},o(i){G(r),n=!1},d(i){r&&r.d(i),i&&V(e)}}}function xa(t,e,n){let r;return lt(t,si,l=>n(0,r=l)),[r,()=>yi(!0),()=>yi(!0),()=>yi(!1)]}class Da extends be{constructor(e){super(),ye(this,e,xa,Ca,ve,{})}}var Oe;(function(t){t.assertEqual=i=>i;function e(i){}t.assertIs=e;function n(i){throw new Error}t.assertNever=n,t.arrayToEnum=i=>{const s={};for(const o of i)s[o]=o;return s},t.getValidEnumValues=i=>{const s=t.objectKeys(i).filter(l=>typeof i[i[l]]!="number"),o={};for(const l of s)o[l]=i[l];return t.objectValues(o)},t.objectValues=i=>t.objectKeys(i).map(function(s){return i[s]}),t.objectKeys=typeof Object.keys=="function"?i=>Object.keys(i):i=>{const s=[];for(const o in i)Object.prototype.hasOwnProperty.call(i,o)&&s.push(o);return s},t.find=(i,s)=>{for(const o of i)if(s(o))return o},t.isInteger=typeof Number.isInteger=="function"?i=>Number.isInteger(i):i=>typeof i=="number"&&isFinite(i)&&Math.floor(i)===i;function r(i,s=" | "){return i.map(o=>typeof o=="string"?`'${o}'`:o).join(s)}t.joinValues=r,t.jsonStringifyReplacer=(i,s)=>typeof s=="bigint"?s.toString():s})(Oe||(Oe={}));var Di;(function(t){t.mergeShapes=(e,n)=>({...e,...n})})(Di||(Di={}));const ee=Oe.arrayToEnum(["string","nan","number","integer","float","boolean","date","bigint","symbol","function","undefined","null","array","object","unknown","promise","void","never","map","set"]),Yt=t=>{switch(typeof t){case"undefined":return ee.undefined;case"string":return ee.string;case"number":return isNaN(t)?ee.nan:ee.number;case"boolean":return ee.boolean;case"function":return ee.function;case"bigint":return ee.bigint;case"symbol":return ee.symbol;case"object":return Array.isArray(t)?ee.array:t===null?ee.null:t.then&&typeof t.then=="function"&&t.catch&&typeof t.catch=="function"?ee.promise:typeof Map<"u"&&t instanceof Map?ee.map:typeof Set<"u"&&t instanceof Set?ee.set:typeof Date<"u"&&t instanceof Date?ee.date:ee.object;default:return ee.unknown}},X=Oe.arrayToEnum(["invalid_type","invalid_literal","custom","invalid_union","invalid_union_discriminator","invalid_enum_value","unrecognized_keys","invalid_arguments","invalid_return_type","invalid_date","invalid_string","too_small","too_big","invalid_intersection_types","not_multiple_of","not_finite"]),Sa=t=>JSON.stringify(t,null,2).replace(/"([^"]+)":/g,"$1:");class xt extends Error{constructor(e){super(),this.issues=[],this.addIssue=r=>{this.issues=[...this.issues,r]},this.addIssues=(r=[])=>{this.issues=[...this.issues,...r]};const n=new.target.prototype;Object.setPrototypeOf?Object.setPrototypeOf(this,n):this.__proto__=n,this.name="ZodError",this.issues=e}get errors(){return this.issues}format(e){const n=e||function(s){return s.message},r={_errors:[]},i=s=>{for(const o of s.issues)if(o.code==="invalid_union")o.unionErrors.map(i);else if(o.code==="invalid_return_type")i(o.returnTypeError);else if(o.code==="invalid_arguments")i(o.argumentsError);else if(o.path.length===0)r._errors.push(n(o));else{let l=r,a=0;for(;a<o.path.length;){const f=o.path[a];a===o.path.length-1?(l[f]=l[f]||{_errors:[]},l[f]._errors.push(n(o))):l[f]=l[f]||{_errors:[]},l=l[f],a++}}};return i(this),r}toString(){return this.message}get message(){return JSON.stringify(this.issues,Oe.jsonStringifyReplacer,2)}get isEmpty(){return this.issues.length===0}flatten(e=n=>n.message){const n={},r=[];for(const i of this.issues)i.path.length>0?(n[i.path[0]]=n[i.path[0]]||[],n[i.path[0]].push(e(i))):r.push(e(i));return{formErrors:r,fieldErrors:n}}get formErrors(){return this.flatten()}}xt.create=t=>new xt(t);const er=(t,e)=>{let n;switch(t.code){case X.invalid_type:t.received===ee.undefined?n="Required":n=`Expected ${t.expected}, received ${t.received}`;break;case X.invalid_literal:n=`Invalid literal value, expected ${JSON.stringify(t.expected,Oe.jsonStringifyReplacer)}`;break;case X.unrecognized_keys:n=`Unrecognized key(s) in object: ${Oe.joinValues(t.keys,", ")}`;break;case X.invalid_union:n="Invalid input";break;case X.invalid_union_discriminator:n=`Invalid discriminator value. Expected ${Oe.joinValues(t.options)}`;break;case X.invalid_enum_value:n=`Invalid enum value. Expected ${Oe.joinValues(t.options)}, received '${t.received}'`;break;case X.invalid_arguments:n="Invalid function arguments";break;case X.invalid_return_type:n="Invalid function return type";break;case X.invalid_date:n="Invalid date";break;case X.invalid_string:typeof t.validation=="object"?"includes"in t.validation?(n=`Invalid input: must include "${t.validation.includes}"`,typeof t.validation.position=="number"&&(n=`${n} at one or more positions greater than or equal to ${t.validation.position}`)):"startsWith"in t.validation?n=`Invalid input: must start with "${t.validation.startsWith}"`:"endsWith"in t.validation?n=`Invalid input: must end with "${t.validation.endsWith}"`:Oe.assertNever(t.validation):t.validation!=="regex"?n=`Invalid ${t.validation}`:n="Invalid";break;case X.too_small:t.type==="array"?n=`Array must contain ${t.exact?"exactly":t.inclusive?"at least":"more than"} ${t.minimum} element(s)`:t.type==="string"?n=`String must contain ${t.exact?"exactly":t.inclusive?"at least":"over"} ${t.minimum} character(s)`:t.type==="number"?n=`Number must be ${t.exact?"exactly equal to ":t.inclusive?"greater than or equal to ":"greater than "}${t.minimum}`:t.type==="date"?n=`Date must be ${t.exact?"exactly equal to ":t.inclusive?"greater than or equal to ":"greater than "}${new Date(Number(t.minimum))}`:n="Invalid input";break;case X.too_big:t.type==="array"?n=`Array must contain ${t.exact?"exactly":t.inclusive?"at most":"less than"} ${t.maximum} element(s)`:t.type==="string"?n=`String must contain ${t.exact?"exactly":t.inclusive?"at most":"under"} ${t.maximum} character(s)`:t.type==="number"?n=`Number must be ${t.exact?"exactly":t.inclusive?"less than or equal to":"less than"} ${t.maximum}`:t.type==="bigint"?n=`BigInt must be ${t.exact?"exactly":t.inclusive?"less than or equal to":"less than"} ${t.maximum}`:t.type==="date"?n=`Date must be ${t.exact?"exactly":t.inclusive?"smaller than or equal to":"smaller than"} ${new Date(Number(t.maximum))}`:n="Invalid input";break;case X.custom:n="Invalid input";break;case X.invalid_intersection_types:n="Intersection results could not be merged";break;case X.not_multiple_of:n=`Number must be a multiple of ${t.multipleOf}`;break;case X.not_finite:n="Number must be finite";break;default:n=e.defaultError,Oe.assertNever(t)}return{message:n}};let Bo=er;function Ia(t){Bo=t}function Nr(){return Bo}const qr=t=>{const{data:e,path:n,errorMaps:r,issueData:i}=t,s=[...n,...i.path||[]],o={...i,path:s};let l="";const a=r.filter(f=>!!f).slice().reverse();for(const f of a)l=f(o,{data:e,defaultError:l}).message;return{...i,path:s,message:i.message||l}},$a=[];function re(t,e){const n=qr({issueData:e,data:t.data,path:t.path,errorMaps:[t.common.contextualErrorMap,t.schemaErrorMap,Nr(),er].filter(r=>!!r)});t.common.issues.push(n)}class ct{constructor(){this.value="valid"}dirty(){this.value==="valid"&&(this.value="dirty")}abort(){this.value!=="aborted"&&(this.value="aborted")}static mergeArray(e,n){const r=[];for(const i of n){if(i.status==="aborted")return _e;i.status==="dirty"&&e.dirty(),r.push(i.value)}return{status:e.value,value:r}}static async mergeObjectAsync(e,n){const r=[];for(const i of n)r.push({key:await i.key,value:await i.value});return ct.mergeObjectSync(e,r)}static mergeObjectSync(e,n){const r={};for(const i of n){const{key:s,value:o}=i;if(s.status==="aborted"||o.status==="aborted")return _e;s.status==="dirty"&&e.dirty(),o.status==="dirty"&&e.dirty(),(typeof o.value<"u"||i.alwaysSet)&&(r[s.value]=o.value)}return{status:e.value,value:r}}}const _e=Object.freeze({status:"aborted"}),Vo=t=>({status:"dirty",value:t}),pt=t=>({status:"valid",value:t}),Si=t=>t.status==="aborted",Ii=t=>t.status==="dirty",Fr=t=>t.status==="valid",Lr=t=>typeof Promise<"u"&&t instanceof Promise;var ue;(function(t){t.errToObj=e=>typeof e=="string"?{message:e}:e||{},t.toString=e=>typeof e=="string"?e:e==null?void 0:e.message})(ue||(ue={}));class qt{constructor(e,n,r,i){this._cachedPath=[],this.parent=e,this.data=n,this._path=r,this._key=i}get path(){return this._cachedPath.length||(this._key instanceof Array?this._cachedPath.push(...this._path,...this._key):this._cachedPath.push(...this._path,this._key)),this._cachedPath}}const gs=(t,e)=>{if(Fr(e))return{success:!0,data:e.value};if(!t.common.issues.length)throw new Error("Validation failed but no issues detected.");return{success:!1,get error(){if(this._error)return this._error;const n=new xt(t.common.issues);return this._error=n,this._error}}};function we(t){if(!t)return{};const{errorMap:e,invalid_type_error:n,required_error:r,description:i}=t;if(e&&(n||r))throw new Error(`Can't use "invalid_type_error" or "required_error" in conjunction with custom error map.`);return e?{errorMap:e,description:i}:{errorMap:(o,l)=>o.code!=="invalid_type"?{message:l.defaultError}:typeof l.data>"u"?{message:r??l.defaultError}:{message:n??l.defaultError},description:i}}class Ce{constructor(e){this.spa=this.safeParseAsync,this._def=e,this.parse=this.parse.bind(this),this.safeParse=this.safeParse.bind(this),this.parseAsync=this.parseAsync.bind(this),this.safeParseAsync=this.safeParseAsync.bind(this),this.spa=this.spa.bind(this),this.refine=this.refine.bind(this),this.refinement=this.refinement.bind(this),this.superRefine=this.superRefine.bind(this),this.optional=this.optional.bind(this),this.nullable=this.nullable.bind(this),this.nullish=this.nullish.bind(this),this.array=this.array.bind(this),this.promise=this.promise.bind(this),this.or=this.or.bind(this),this.and=this.and.bind(this),this.transform=this.transform.bind(this),this.brand=this.brand.bind(this),this.default=this.default.bind(this),this.catch=this.catch.bind(this),this.describe=this.describe.bind(this),this.pipe=this.pipe.bind(this),this.isNullable=this.isNullable.bind(this),this.isOptional=this.isOptional.bind(this)}get description(){return this._def.description}_getType(e){return Yt(e.data)}_getOrReturnCtx(e,n){return n||{common:e.parent.common,data:e.data,parsedType:Yt(e.data),schemaErrorMap:this._def.errorMap,path:e.path,parent:e.parent}}_processInputParams(e){return{status:new ct,ctx:{common:e.parent.common,data:e.data,parsedType:Yt(e.data),schemaErrorMap:this._def.errorMap,path:e.path,parent:e.parent}}}_parseSync(e){const n=this._parse(e);if(Lr(n))throw new Error("Synchronous parse encountered promise.");return n}_parseAsync(e){const n=this._parse(e);return Promise.resolve(n)}parse(e,n){const r=this.safeParse(e,n);if(r.success)return r.data;throw r.error}safeParse(e,n){var r;const i={common:{issues:[],async:(r=n==null?void 0:n.async)!==null&&r!==void 0?r:!1,contextualErrorMap:n==null?void 0:n.errorMap},path:(n==null?void 0:n.path)||[],schemaErrorMap:this._def.errorMap,parent:null,data:e,parsedType:Yt(e)},s=this._parseSync({data:e,path:i.path,parent:i});return gs(i,s)}async parseAsync(e,n){const r=await this.safeParseAsync(e,n);if(r.success)return r.data;throw r.error}async safeParseAsync(e,n){const r={common:{issues:[],contextualErrorMap:n==null?void 0:n.errorMap,async:!0},path:(n==null?void 0:n.path)||[],schemaErrorMap:this._def.errorMap,parent:null,data:e,parsedType:Yt(e)},i=this._parse({data:e,path:r.path,parent:r}),s=await(Lr(i)?i:Promise.resolve(i));return gs(r,s)}refine(e,n){const r=i=>typeof n=="string"||typeof n>"u"?{message:n}:typeof n=="function"?n(i):n;return this._refinement((i,s)=>{const o=e(i),l=()=>s.addIssue({code:X.custom,...r(i)});return typeof Promise<"u"&&o instanceof Promise?o.then(a=>a?!0:(l(),!1)):o?!0:(l(),!1)})}refinement(e,n){return this._refinement((r,i)=>e(r)?!0:(i.addIssue(typeof n=="function"?n(r,i):n),!1))}_refinement(e){return new St({schema:this,typeName:he.ZodEffects,effect:{type:"refinement",refinement:e}})}superRefine(e){return this._refinement(e)}optional(){return zt.create(this,this._def)}nullable(){return bn.create(this,this._def)}nullish(){return this.nullable().optional()}array(){return Dt.create(this,this._def)}promise(){return Pn.create(this,this._def)}or(e){return ir.create([this,e],this._def)}and(e){return sr.create(this,e,this._def)}transform(e){return new St({...we(this._def),schema:this,typeName:he.ZodEffects,effect:{type:"transform",transform:e}})}default(e){const n=typeof e=="function"?e:()=>e;return new cr({...we(this._def),innerType:this,defaultValue:n,typeName:he.ZodDefault})}brand(){return new Go({typeName:he.ZodBranded,type:this,...we(this._def)})}catch(e){const n=typeof e=="function"?e:()=>e;return new Zr({...we(this._def),innerType:this,catchValue:n,typeName:he.ZodCatch})}describe(e){const n=this.constructor;return new n({...this._def,description:e})}pipe(e){return br.create(this,e)}isOptional(){return this.safeParse(void 0).success}isNullable(){return this.safeParse(null).success}}const Oa=/^c[^\s-]{8,}$/i,Ma=/^[a-z][a-z0-9]*$/,Ra=/[0-9A-HJKMNP-TV-Z]{26}/,Na=/^([a-f0-9]{8}-[a-f0-9]{4}-[1-5][a-f0-9]{3}-[a-f0-9]{4}-[a-f0-9]{12}|00000000-0000-0000-0000-000000000000)$/i,qa=/^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[(((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))\.){3}((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))\])|(\[IPv6:(([a-f0-9]{1,4}:){7}|::([a-f0-9]{1,4}:){0,6}|([a-f0-9]{1,4}:){1}:([a-f0-9]{1,4}:){0,5}|([a-f0-9]{1,4}:){2}:([a-f0-9]{1,4}:){0,4}|([a-f0-9]{1,4}:){3}:([a-f0-9]{1,4}:){0,3}|([a-f0-9]{1,4}:){4}:([a-f0-9]{1,4}:){0,2}|([a-f0-9]{1,4}:){5}:([a-f0-9]{1,4}:){0,1})([a-f0-9]{1,4}|(((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))\.){3}((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2})))\])|([A-Za-z0-9]([A-Za-z0-9-]*[A-Za-z0-9])*(\.[A-Za-z]{2,})+))$/,Fa=/^(\p{Extended_Pictographic}|\p{Emoji_Component})+$/u,La=/^(((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))\.){3}((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))$/,Pa=/^(([a-f0-9]{1,4}:){7}|::([a-f0-9]{1,4}:){0,6}|([a-f0-9]{1,4}:){1}:([a-f0-9]{1,4}:){0,5}|([a-f0-9]{1,4}:){2}:([a-f0-9]{1,4}:){0,4}|([a-f0-9]{1,4}:){3}:([a-f0-9]{1,4}:){0,3}|([a-f0-9]{1,4}:){4}:([a-f0-9]{1,4}:){0,2}|([a-f0-9]{1,4}:){5}:([a-f0-9]{1,4}:){0,1})([a-f0-9]{1,4}|(((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))\.){3}((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2})))$/,Ua=t=>t.precision?t.offset?new RegExp(`^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{${t.precision}}(([+-]\\d{2}(:?\\d{2})?)|Z)$`):new RegExp(`^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{${t.precision}}Z$`):t.precision===0?t.offset?new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(([+-]\\d{2}(:?\\d{2})?)|Z)$"):new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"):t.offset?new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d+)?(([+-]\\d{2}(:?\\d{2})?)|Z)$"):new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d+)?Z$");function za(t,e){return!!((e==="v4"||!e)&&La.test(t)||(e==="v6"||!e)&&Pa.test(t))}class Tt extends Ce{constructor(){super(...arguments),this._regex=(e,n,r)=>this.refinement(i=>e.test(i),{validation:n,code:X.invalid_string,...ue.errToObj(r)}),this.nonempty=e=>this.min(1,ue.errToObj(e)),this.trim=()=>new Tt({...this._def,checks:[...this._def.checks,{kind:"trim"}]}),this.toLowerCase=()=>new Tt({...this._def,checks:[...this._def.checks,{kind:"toLowerCase"}]}),this.toUpperCase=()=>new Tt({...this._def,checks:[...this._def.checks,{kind:"toUpperCase"}]})}_parse(e){if(this._def.coerce&&(e.data=String(e.data)),this._getType(e)!==ee.string){const s=this._getOrReturnCtx(e);return re(s,{code:X.invalid_type,expected:ee.string,received:s.parsedType}),_e}const r=new ct;let i;for(const s of this._def.checks)if(s.kind==="min")e.data.length<s.value&&(i=this._getOrReturnCtx(e,i),re(i,{code:X.too_small,minimum:s.value,type:"string",inclusive:!0,exact:!1,message:s.message}),r.dirty());else if(s.kind==="max")e.data.length>s.value&&(i=this._getOrReturnCtx(e,i),re(i,{code:X.too_big,maximum:s.value,type:"string",inclusive:!0,exact:!1,message:s.message}),r.dirty());else if(s.kind==="length"){const o=e.data.length>s.value,l=e.data.length<s.value;(o||l)&&(i=this._getOrReturnCtx(e,i),o?re(i,{code:X.too_big,maximum:s.value,type:"string",inclusive:!0,exact:!0,message:s.message}):l&&re(i,{code:X.too_small,minimum:s.value,type:"string",inclusive:!0,exact:!0,message:s.message}),r.dirty())}else if(s.kind==="email")qa.test(e.data)||(i=this._getOrReturnCtx(e,i),re(i,{validation:"email",code:X.invalid_string,message:s.message}),r.dirty());else if(s.kind==="emoji")Fa.test(e.data)||(i=this._getOrReturnCtx(e,i),re(i,{validation:"emoji",code:X.invalid_string,message:s.message}),r.dirty());else if(s.kind==="uuid")Na.test(e.data)||(i=this._getOrReturnCtx(e,i),re(i,{validation:"uuid",code:X.invalid_string,message:s.message}),r.dirty());else if(s.kind==="cuid")Oa.test(e.data)||(i=this._getOrReturnCtx(e,i),re(i,{validation:"cuid",code:X.invalid_string,message:s.message}),r.dirty());else if(s.kind==="cuid2")Ma.test(e.data)||(i=this._getOrReturnCtx(e,i),re(i,{validation:"cuid2",code:X.invalid_string,message:s.message}),r.dirty());else if(s.kind==="ulid")Ra.test(e.data)||(i=this._getOrReturnCtx(e,i),re(i,{validation:"ulid",code:X.invalid_string,message:s.message}),r.dirty());else if(s.kind==="url")try{new URL(e.data)}catch{i=this._getOrReturnCtx(e,i),re(i,{validation:"url",code:X.invalid_string,message:s.message}),r.dirty()}else s.kind==="regex"?(s.regex.lastIndex=0,s.regex.test(e.data)||(i=this._getOrReturnCtx(e,i),re(i,{validation:"regex",code:X.invalid_string,message:s.message}),r.dirty())):s.kind==="trim"?e.data=e.data.trim():s.kind==="includes"?e.data.includes(s.value,s.position)||(i=this._getOrReturnCtx(e,i),re(i,{code:X.invalid_string,validation:{includes:s.value,position:s.position},message:s.message}),r.dirty()):s.kind==="toLowerCase"?e.data=e.data.toLowerCase():s.kind==="toUpperCase"?e.data=e.data.toUpperCase():s.kind==="startsWith"?e.data.startsWith(s.value)||(i=this._getOrReturnCtx(e,i),re(i,{code:X.invalid_string,validation:{startsWith:s.value},message:s.message}),r.dirty()):s.kind==="endsWith"?e.data.endsWith(s.value)||(i=this._getOrReturnCtx(e,i),re(i,{code:X.invalid_string,validation:{endsWith:s.value},message:s.message}),r.dirty()):s.kind==="datetime"?Ua(s).test(e.data)||(i=this._getOrReturnCtx(e,i),re(i,{code:X.invalid_string,validation:"datetime",message:s.message}),r.dirty()):s.kind==="ip"?za(e.data,s.version)||(i=this._getOrReturnCtx(e,i),re(i,{validation:"ip",code:X.invalid_string,message:s.message}),r.dirty()):Oe.assertNever(s);return{status:r.value,value:e.data}}_addCheck(e){return new Tt({...this._def,checks:[...this._def.checks,e]})}email(e){return this._addCheck({kind:"email",...ue.errToObj(e)})}url(e){return this._addCheck({kind:"url",...ue.errToObj(e)})}emoji(e){return this._addCheck({kind:"emoji",...ue.errToObj(e)})}uuid(e){return this._addCheck({kind:"uuid",...ue.errToObj(e)})}cuid(e){return this._addCheck({kind:"cuid",...ue.errToObj(e)})}cuid2(e){return this._addCheck({kind:"cuid2",...ue.errToObj(e)})}ulid(e){return this._addCheck({kind:"ulid",...ue.errToObj(e)})}ip(e){return this._addCheck({kind:"ip",...ue.errToObj(e)})}datetime(e){var n;return typeof e=="string"?this._addCheck({kind:"datetime",precision:null,offset:!1,message:e}):this._addCheck({kind:"datetime",precision:typeof(e==null?void 0:e.precision)>"u"?null:e==null?void 0:e.precision,offset:(n=e==null?void 0:e.offset)!==null&&n!==void 0?n:!1,...ue.errToObj(e==null?void 0:e.message)})}regex(e,n){return this._addCheck({kind:"regex",regex:e,...ue.errToObj(n)})}includes(e,n){return this._addCheck({kind:"includes",value:e,position:n==null?void 0:n.position,...ue.errToObj(n==null?void 0:n.message)})}startsWith(e,n){return this._addCheck({kind:"startsWith",value:e,...ue.errToObj(n)})}endsWith(e,n){return this._addCheck({kind:"endsWith",value:e,...ue.errToObj(n)})}min(e,n){return this._addCheck({kind:"min",value:e,...ue.errToObj(n)})}max(e,n){return this._addCheck({kind:"max",value:e,...ue.errToObj(n)})}length(e,n){return this._addCheck({kind:"length",value:e,...ue.errToObj(n)})}get isDatetime(){return!!this._def.checks.find(e=>e.kind==="datetime")}get isEmail(){return!!this._def.checks.find(e=>e.kind==="email")}get isURL(){return!!this._def.checks.find(e=>e.kind==="url")}get isEmoji(){return!!this._def.checks.find(e=>e.kind==="emoji")}get isUUID(){return!!this._def.checks.find(e=>e.kind==="uuid")}get isCUID(){return!!this._def.checks.find(e=>e.kind==="cuid")}get isCUID2(){return!!this._def.checks.find(e=>e.kind==="cuid2")}get isULID(){return!!this._def.checks.find(e=>e.kind==="ulid")}get isIP(){return!!this._def.checks.find(e=>e.kind==="ip")}get minLength(){let e=null;for(const n of this._def.checks)n.kind==="min"&&(e===null||n.value>e)&&(e=n.value);return e}get maxLength(){let e=null;for(const n of this._def.checks)n.kind==="max"&&(e===null||n.value<e)&&(e=n.value);return e}}Tt.create=t=>{var e;return new Tt({checks:[],typeName:he.ZodString,coerce:(e=t==null?void 0:t.coerce)!==null&&e!==void 0?e:!1,...we(t)})};function Za(t,e){const n=(t.toString().split(".")[1]||"").length,r=(e.toString().split(".")[1]||"").length,i=n>r?n:r,s=parseInt(t.toFixed(i).replace(".","")),o=parseInt(e.toFixed(i).replace(".",""));return s%o/Math.pow(10,i)}class Jt extends Ce{constructor(){super(...arguments),this.min=this.gte,this.max=this.lte,this.step=this.multipleOf}_parse(e){if(this._def.coerce&&(e.data=Number(e.data)),this._getType(e)!==ee.number){const s=this._getOrReturnCtx(e);return re(s,{code:X.invalid_type,expected:ee.number,received:s.parsedType}),_e}let r;const i=new ct;for(const s of this._def.checks)s.kind==="int"?Oe.isInteger(e.data)||(r=this._getOrReturnCtx(e,r),re(r,{code:X.invalid_type,expected:"integer",received:"float",message:s.message}),i.dirty()):s.kind==="min"?(s.inclusive?e.data<s.value:e.data<=s.value)&&(r=this._getOrReturnCtx(e,r),re(r,{code:X.too_small,minimum:s.value,type:"number",inclusive:s.inclusive,exact:!1,message:s.message}),i.dirty()):s.kind==="max"?(s.inclusive?e.data>s.value:e.data>=s.value)&&(r=this._getOrReturnCtx(e,r),re(r,{code:X.too_big,maximum:s.value,type:"number",inclusive:s.inclusive,exact:!1,message:s.message}),i.dirty()):s.kind==="multipleOf"?Za(e.data,s.value)!==0&&(r=this._getOrReturnCtx(e,r),re(r,{code:X.not_multiple_of,multipleOf:s.value,message:s.message}),i.dirty()):s.kind==="finite"?Number.isFinite(e.data)||(r=this._getOrReturnCtx(e,r),re(r,{code:X.not_finite,message:s.message}),i.dirty()):Oe.assertNever(s);return{status:i.value,value:e.data}}gte(e,n){return this.setLimit("min",e,!0,ue.toString(n))}gt(e,n){return this.setLimit("min",e,!1,ue.toString(n))}lte(e,n){return this.setLimit("max",e,!0,ue.toString(n))}lt(e,n){return this.setLimit("max",e,!1,ue.toString(n))}setLimit(e,n,r,i){return new Jt({...this._def,checks:[...this._def.checks,{kind:e,value:n,inclusive:r,message:ue.toString(i)}]})}_addCheck(e){return new Jt({...this._def,checks:[...this._def.checks,e]})}int(e){return this._addCheck({kind:"int",message:ue.toString(e)})}positive(e){return this._addCheck({kind:"min",value:0,inclusive:!1,message:ue.toString(e)})}negative(e){return this._addCheck({kind:"max",value:0,inclusive:!1,message:ue.toString(e)})}nonpositive(e){return this._addCheck({kind:"max",value:0,inclusive:!0,message:ue.toString(e)})}nonnegative(e){return this._addCheck({kind:"min",value:0,inclusive:!0,message:ue.toString(e)})}multipleOf(e,n){return this._addCheck({kind:"multipleOf",value:e,message:ue.toString(n)})}finite(e){return this._addCheck({kind:"finite",message:ue.toString(e)})}safe(e){return this._addCheck({kind:"min",inclusive:!0,value:Number.MIN_SAFE_INTEGER,message:ue.toString(e)})._addCheck({kind:"max",inclusive:!0,value:Number.MAX_SAFE_INTEGER,message:ue.toString(e)})}get minValue(){let e=null;for(const n of this._def.checks)n.kind==="min"&&(e===null||n.value>e)&&(e=n.value);return e}get maxValue(){let e=null;for(const n of this._def.checks)n.kind==="max"&&(e===null||n.value<e)&&(e=n.value);return e}get isInt(){return!!this._def.checks.find(e=>e.kind==="int"||e.kind==="multipleOf"&&Oe.isInteger(e.value))}get isFinite(){let e=null,n=null;for(const r of this._def.checks){if(r.kind==="finite"||r.kind==="int"||r.kind==="multipleOf")return!0;r.kind==="min"?(n===null||r.value>n)&&(n=r.value):r.kind==="max"&&(e===null||r.value<e)&&(e=r.value)}return Number.isFinite(n)&&Number.isFinite(e)}}Jt.create=t=>new Jt({checks:[],typeName:he.ZodNumber,coerce:(t==null?void 0:t.coerce)||!1,...we(t)});class en extends Ce{constructor(){super(...arguments),this.min=this.gte,this.max=this.lte}_parse(e){if(this._def.coerce&&(e.data=BigInt(e.data)),this._getType(e)!==ee.bigint){const s=this._getOrReturnCtx(e);return re(s,{code:X.invalid_type,expected:ee.bigint,received:s.parsedType}),_e}let r;const i=new ct;for(const s of this._def.checks)s.kind==="min"?(s.inclusive?e.data<s.value:e.data<=s.value)&&(r=this._getOrReturnCtx(e,r),re(r,{code:X.too_small,type:"bigint",minimum:s.value,inclusive:s.inclusive,message:s.message}),i.dirty()):s.kind==="max"?(s.inclusive?e.data>s.value:e.data>=s.value)&&(r=this._getOrReturnCtx(e,r),re(r,{code:X.too_big,type:"bigint",maximum:s.value,inclusive:s.inclusive,message:s.message}),i.dirty()):s.kind==="multipleOf"?e.data%s.value!==BigInt(0)&&(r=this._getOrReturnCtx(e,r),re(r,{code:X.not_multiple_of,multipleOf:s.value,message:s.message}),i.dirty()):Oe.assertNever(s);return{status:i.value,value:e.data}}gte(e,n){return this.setLimit("min",e,!0,ue.toString(n))}gt(e,n){return this.setLimit("min",e,!1,ue.toString(n))}lte(e,n){return this.setLimit("max",e,!0,ue.toString(n))}lt(e,n){return this.setLimit("max",e,!1,ue.toString(n))}setLimit(e,n,r,i){return new en({...this._def,checks:[...this._def.checks,{kind:e,value:n,inclusive:r,message:ue.toString(i)}]})}_addCheck(e){return new en({...this._def,checks:[...this._def.checks,e]})}positive(e){return this._addCheck({kind:"min",value:BigInt(0),inclusive:!1,message:ue.toString(e)})}negative(e){return this._addCheck({kind:"max",value:BigInt(0),inclusive:!1,message:ue.toString(e)})}nonpositive(e){return this._addCheck({kind:"max",value:BigInt(0),inclusive:!0,message:ue.toString(e)})}nonnegative(e){return this._addCheck({kind:"min",value:BigInt(0),inclusive:!0,message:ue.toString(e)})}multipleOf(e,n){return this._addCheck({kind:"multipleOf",value:e,message:ue.toString(n)})}get minValue(){let e=null;for(const n of this._def.checks)n.kind==="min"&&(e===null||n.value>e)&&(e=n.value);return e}get maxValue(){let e=null;for(const n of this._def.checks)n.kind==="max"&&(e===null||n.value<e)&&(e=n.value);return e}}en.create=t=>{var e;return new en({checks:[],typeName:he.ZodBigInt,coerce:(e=t==null?void 0:t.coerce)!==null&&e!==void 0?e:!1,...we(t)})};class tr extends Ce{_parse(e){if(this._def.coerce&&(e.data=!!e.data),this._getType(e)!==ee.boolean){const r=this._getOrReturnCtx(e);return re(r,{code:X.invalid_type,expected:ee.boolean,received:r.parsedType}),_e}return pt(e.data)}}tr.create=t=>new tr({typeName:he.ZodBoolean,coerce:(t==null?void 0:t.coerce)||!1,...we(t)});class vn extends Ce{_parse(e){if(this._def.coerce&&(e.data=new Date(e.data)),this._getType(e)!==ee.date){const s=this._getOrReturnCtx(e);return re(s,{code:X.invalid_type,expected:ee.date,received:s.parsedType}),_e}if(isNaN(e.data.getTime())){const s=this._getOrReturnCtx(e);return re(s,{code:X.invalid_date}),_e}const r=new ct;let i;for(const s of this._def.checks)s.kind==="min"?e.data.getTime()<s.value&&(i=this._getOrReturnCtx(e,i),re(i,{code:X.too_small,message:s.message,inclusive:!0,exact:!1,minimum:s.value,type:"date"}),r.dirty()):s.kind==="max"?e.data.getTime()>s.value&&(i=this._getOrReturnCtx(e,i),re(i,{code:X.too_big,message:s.message,inclusive:!0,exact:!1,maximum:s.value,type:"date"}),r.dirty()):Oe.assertNever(s);return{status:r.value,value:new Date(e.data.getTime())}}_addCheck(e){return new vn({...this._def,checks:[...this._def.checks,e]})}min(e,n){return this._addCheck({kind:"min",value:e.getTime(),message:ue.toString(n)})}max(e,n){return this._addCheck({kind:"max",value:e.getTime(),message:ue.toString(n)})}get minDate(){let e=null;for(const n of this._def.checks)n.kind==="min"&&(e===null||n.value>e)&&(e=n.value);return e!=null?new Date(e):null}get maxDate(){let e=null;for(const n of this._def.checks)n.kind==="max"&&(e===null||n.value<e)&&(e=n.value);return e!=null?new Date(e):null}}vn.create=t=>new vn({checks:[],coerce:(t==null?void 0:t.coerce)||!1,typeName:he.ZodDate,...we(t)});class Pr extends Ce{_parse(e){if(this._getType(e)!==ee.symbol){const r=this._getOrReturnCtx(e);return re(r,{code:X.invalid_type,expected:ee.symbol,received:r.parsedType}),_e}return pt(e.data)}}Pr.create=t=>new Pr({typeName:he.ZodSymbol,...we(t)});class nr extends Ce{_parse(e){if(this._getType(e)!==ee.undefined){const r=this._getOrReturnCtx(e);return re(r,{code:X.invalid_type,expected:ee.undefined,received:r.parsedType}),_e}return pt(e.data)}}nr.create=t=>new nr({typeName:he.ZodUndefined,...we(t)});class rr extends Ce{_parse(e){if(this._getType(e)!==ee.null){const r=this._getOrReturnCtx(e);return re(r,{code:X.invalid_type,expected:ee.null,received:r.parsedType}),_e}return pt(e.data)}}rr.create=t=>new rr({typeName:he.ZodNull,...we(t)});class Ln extends Ce{constructor(){super(...arguments),this._any=!0}_parse(e){return pt(e.data)}}Ln.create=t=>new Ln({typeName:he.ZodAny,...we(t)});class pn extends Ce{constructor(){super(...arguments),this._unknown=!0}_parse(e){return pt(e.data)}}pn.create=t=>new pn({typeName:he.ZodUnknown,...we(t)});class jt extends Ce{_parse(e){const n=this._getOrReturnCtx(e);return re(n,{code:X.invalid_type,expected:ee.never,received:n.parsedType}),_e}}jt.create=t=>new jt({typeName:he.ZodNever,...we(t)});class Ur extends Ce{_parse(e){if(this._getType(e)!==ee.undefined){const r=this._getOrReturnCtx(e);return re(r,{code:X.invalid_type,expected:ee.void,received:r.parsedType}),_e}return pt(e.data)}}Ur.create=t=>new Ur({typeName:he.ZodVoid,...we(t)});class Dt extends Ce{_parse(e){const{ctx:n,status:r}=this._processInputParams(e),i=this._def;if(n.parsedType!==ee.array)return re(n,{code:X.invalid_type,expected:ee.array,received:n.parsedType}),_e;if(i.exactLength!==null){const o=n.data.length>i.exactLength.value,l=n.data.length<i.exactLength.value;(o||l)&&(re(n,{code:o?X.too_big:X.too_small,minimum:l?i.exactLength.value:void 0,maximum:o?i.exactLength.value:void 0,type:"array",inclusive:!0,exact:!0,message:i.exactLength.message}),r.dirty())}if(i.minLength!==null&&n.data.length<i.minLength.value&&(re(n,{code:X.too_small,minimum:i.minLength.value,type:"array",inclusive:!0,exact:!1,message:i.minLength.message}),r.dirty()),i.maxLength!==null&&n.data.length>i.maxLength.value&&(re(n,{code:X.too_big,maximum:i.maxLength.value,type:"array",inclusive:!0,exact:!1,message:i.maxLength.message}),r.dirty()),n.common.async)return Promise.all([...n.data].map((o,l)=>i.type._parseAsync(new qt(n,o,n.path,l)))).then(o=>ct.mergeArray(r,o));const s=[...n.data].map((o,l)=>i.type._parseSync(new qt(n,o,n.path,l)));return ct.mergeArray(r,s)}get element(){return this._def.type}min(e,n){return new Dt({...this._def,minLength:{value:e,message:ue.toString(n)}})}max(e,n){return new Dt({...this._def,maxLength:{value:e,message:ue.toString(n)}})}length(e,n){return new Dt({...this._def,exactLength:{value:e,message:ue.toString(n)}})}nonempty(e){return this.min(1,e)}}Dt.create=(t,e)=>new Dt({type:t,minLength:null,maxLength:null,exactLength:null,typeName:he.ZodArray,...we(e)});function Cn(t){if(t instanceof ze){const e={};for(const n in t.shape){const r=t.shape[n];e[n]=zt.create(Cn(r))}return new ze({...t._def,shape:()=>e})}else return t instanceof Dt?new Dt({...t._def,type:Cn(t.element)}):t instanceof zt?zt.create(Cn(t.unwrap())):t instanceof bn?bn.create(Cn(t.unwrap())):t instanceof Ft?Ft.create(t.items.map(e=>Cn(e))):t}class ze extends Ce{constructor(){super(...arguments),this._cached=null,this.nonstrict=this.passthrough,this.augment=this.extend}_getCached(){if(this._cached!==null)return this._cached;const e=this._def.shape(),n=Oe.objectKeys(e);return this._cached={shape:e,keys:n}}_parse(e){if(this._getType(e)!==ee.object){const f=this._getOrReturnCtx(e);return re(f,{code:X.invalid_type,expected:ee.object,received:f.parsedType}),_e}const{status:r,ctx:i}=this._processInputParams(e),{shape:s,keys:o}=this._getCached(),l=[];if(!(this._def.catchall instanceof jt&&this._def.unknownKeys==="strip"))for(const f in i.data)o.includes(f)||l.push(f);const a=[];for(const f of o){const u=s[f],d=i.data[f];a.push({key:{status:"valid",value:f},value:u._parse(new qt(i,d,i.path,f)),alwaysSet:f in i.data})}if(this._def.catchall instanceof jt){const f=this._def.unknownKeys;if(f==="passthrough")for(const u of l)a.push({key:{status:"valid",value:u},value:{status:"valid",value:i.data[u]}});else if(f==="strict")l.length>0&&(re(i,{code:X.unrecognized_keys,keys:l}),r.dirty());else if(f!=="strip")throw new Error("Internal ZodObject error: invalid unknownKeys value.")}else{const f=this._def.catchall;for(const u of l){const d=i.data[u];a.push({key:{status:"valid",value:u},value:f._parse(new qt(i,d,i.path,u)),alwaysSet:u in i.data})}}return i.common.async?Promise.resolve().then(async()=>{const f=[];for(const u of a){const d=await u.key;f.push({key:d,value:await u.value,alwaysSet:u.alwaysSet})}return f}).then(f=>ct.mergeObjectSync(r,f)):ct.mergeObjectSync(r,a)}get shape(){return this._def.shape()}strict(e){return ue.errToObj,new ze({...this._def,unknownKeys:"strict",...e!==void 0?{errorMap:(n,r)=>{var i,s,o,l;const a=(o=(s=(i=this._def).errorMap)===null||s===void 0?void 0:s.call(i,n,r).message)!==null&&o!==void 0?o:r.defaultError;return n.code==="unrecognized_keys"?{message:(l=ue.errToObj(e).message)!==null&&l!==void 0?l:a}:{message:a}}}:{}})}strip(){return new ze({...this._def,unknownKeys:"strip"})}passthrough(){return new ze({...this._def,unknownKeys:"passthrough"})}extend(e){return new ze({...this._def,shape:()=>({...this._def.shape(),...e})})}merge(e){return new ze({unknownKeys:e._def.unknownKeys,catchall:e._def.catchall,shape:()=>({...this._def.shape(),...e._def.shape()}),typeName:he.ZodObject})}setKey(e,n){return this.augment({[e]:n})}catchall(e){return new ze({...this._def,catchall:e})}pick(e){const n={};return Oe.objectKeys(e).forEach(r=>{e[r]&&this.shape[r]&&(n[r]=this.shape[r])}),new ze({...this._def,shape:()=>n})}omit(e){const n={};return Oe.objectKeys(this.shape).forEach(r=>{e[r]||(n[r]=this.shape[r])}),new ze({...this._def,shape:()=>n})}deepPartial(){return Cn(this)}partial(e){const n={};return Oe.objectKeys(this.shape).forEach(r=>{const i=this.shape[r];e&&!e[r]?n[r]=i:n[r]=i.optional()}),new ze({...this._def,shape:()=>n})}required(e){const n={};return Oe.objectKeys(this.shape).forEach(r=>{if(e&&!e[r])n[r]=this.shape[r];else{let s=this.shape[r];for(;s instanceof zt;)s=s._def.innerType;n[r]=s}}),new ze({...this._def,shape:()=>n})}keyof(){return Wo(Oe.objectKeys(this.shape))}}ze.create=(t,e)=>new ze({shape:()=>t,unknownKeys:"strip",catchall:jt.create(),typeName:he.ZodObject,...we(e)});ze.strictCreate=(t,e)=>new ze({shape:()=>t,unknownKeys:"strict",catchall:jt.create(),typeName:he.ZodObject,...we(e)});ze.lazycreate=(t,e)=>new ze({shape:t,unknownKeys:"strip",catchall:jt.create(),typeName:he.ZodObject,...we(e)});class ir extends Ce{_parse(e){const{ctx:n}=this._processInputParams(e),r=this._def.options;function i(s){for(const l of s)if(l.result.status==="valid")return l.result;for(const l of s)if(l.result.status==="dirty")return n.common.issues.push(...l.ctx.common.issues),l.result;const o=s.map(l=>new xt(l.ctx.common.issues));return re(n,{code:X.invalid_union,unionErrors:o}),_e}if(n.common.async)return Promise.all(r.map(async s=>{const o={...n,common:{...n.common,issues:[]},parent:null};return{result:await s._parseAsync({data:n.data,path:n.path,parent:o}),ctx:o}})).then(i);{let s;const o=[];for(const a of r){const f={...n,common:{...n.common,issues:[]},parent:null},u=a._parseSync({data:n.data,path:n.path,parent:f});if(u.status==="valid")return u;u.status==="dirty"&&!s&&(s={result:u,ctx:f}),f.common.issues.length&&o.push(f.common.issues)}if(s)return n.common.issues.push(...s.ctx.common.issues),s.result;const l=o.map(a=>new xt(a));return re(n,{code:X.invalid_union,unionErrors:l}),_e}}get options(){return this._def.options}}ir.create=(t,e)=>new ir({options:t,typeName:he.ZodUnion,...we(e)});const xr=t=>t instanceof lr?xr(t.schema):t instanceof St?xr(t.innerType()):t instanceof ar?[t.value]:t instanceof tn?t.options:t instanceof ur?Object.keys(t.enum):t instanceof cr?xr(t._def.innerType):t instanceof nr?[void 0]:t instanceof rr?[null]:null;class oi extends Ce{_parse(e){const{ctx:n}=this._processInputParams(e);if(n.parsedType!==ee.object)return re(n,{code:X.invalid_type,expected:ee.object,received:n.parsedType}),_e;const r=this.discriminator,i=n.data[r],s=this.optionsMap.get(i);return s?n.common.async?s._parseAsync({data:n.data,path:n.path,parent:n}):s._parseSync({data:n.data,path:n.path,parent:n}):(re(n,{code:X.invalid_union_discriminator,options:Array.from(this.optionsMap.keys()),path:[r]}),_e)}get discriminator(){return this._def.discriminator}get options(){return this._def.options}get optionsMap(){return this._def.optionsMap}static create(e,n,r){const i=new Map;for(const s of n){const o=xr(s.shape[e]);if(!o)throw new Error(`A discriminator value for key \`${e}\` could not be extracted from all schema options`);for(const l of o){if(i.has(l))throw new Error(`Discriminator property ${String(e)} has duplicate value ${String(l)}`);i.set(l,s)}}return new oi({typeName:he.ZodDiscriminatedUnion,discriminator:e,options:n,optionsMap:i,...we(r)})}}function $i(t,e){const n=Yt(t),r=Yt(e);if(t===e)return{valid:!0,data:t};if(n===ee.object&&r===ee.object){const i=Oe.objectKeys(e),s=Oe.objectKeys(t).filter(l=>i.indexOf(l)!==-1),o={...t,...e};for(const l of s){const a=$i(t[l],e[l]);if(!a.valid)return{valid:!1};o[l]=a.data}return{valid:!0,data:o}}else if(n===ee.array&&r===ee.array){if(t.length!==e.length)return{valid:!1};const i=[];for(let s=0;s<t.length;s++){const o=t[s],l=e[s],a=$i(o,l);if(!a.valid)return{valid:!1};i.push(a.data)}return{valid:!0,data:i}}else return n===ee.date&&r===ee.date&&+t==+e?{valid:!0,data:t}:{valid:!1}}class sr extends Ce{_parse(e){const{status:n,ctx:r}=this._processInputParams(e),i=(s,o)=>{if(Si(s)||Si(o))return _e;const l=$i(s.value,o.value);return l.valid?((Ii(s)||Ii(o))&&n.dirty(),{status:n.value,value:l.data}):(re(r,{code:X.invalid_intersection_types}),_e)};return r.common.async?Promise.all([this._def.left._parseAsync({data:r.data,path:r.path,parent:r}),this._def.right._parseAsync({data:r.data,path:r.path,parent:r})]).then(([s,o])=>i(s,o)):i(this._def.left._parseSync({data:r.data,path:r.path,parent:r}),this._def.right._parseSync({data:r.data,path:r.path,parent:r}))}}sr.create=(t,e,n)=>new sr({left:t,right:e,typeName:he.ZodIntersection,...we(n)});class Ft extends Ce{_parse(e){const{status:n,ctx:r}=this._processInputParams(e);if(r.parsedType!==ee.array)return re(r,{code:X.invalid_type,expected:ee.array,received:r.parsedType}),_e;if(r.data.length<this._def.items.length)return re(r,{code:X.too_small,minimum:this._def.items.length,inclusive:!0,exact:!1,type:"array"}),_e;!this._def.rest&&r.data.length>this._def.items.length&&(re(r,{code:X.too_big,maximum:this._def.items.length,inclusive:!0,exact:!1,type:"array"}),n.dirty());const s=[...r.data].map((o,l)=>{const a=this._def.items[l]||this._def.rest;return a?a._parse(new qt(r,o,r.path,l)):null}).filter(o=>!!o);return r.common.async?Promise.all(s).then(o=>ct.mergeArray(n,o)):ct.mergeArray(n,s)}get items(){return this._def.items}rest(e){return new Ft({...this._def,rest:e})}}Ft.create=(t,e)=>{if(!Array.isArray(t))throw new Error("You must pass an array of schemas to z.tuple([ ... ])");return new Ft({items:t,typeName:he.ZodTuple,rest:null,...we(e)})};class or extends Ce{get keySchema(){return this._def.keyType}get valueSchema(){return this._def.valueType}_parse(e){const{status:n,ctx:r}=this._processInputParams(e);if(r.parsedType!==ee.object)return re(r,{code:X.invalid_type,expected:ee.object,received:r.parsedType}),_e;const i=[],s=this._def.keyType,o=this._def.valueType;for(const l in r.data)i.push({key:s._parse(new qt(r,l,r.path,l)),value:o._parse(new qt(r,r.data[l],r.path,l))});return r.common.async?ct.mergeObjectAsync(n,i):ct.mergeObjectSync(n,i)}get element(){return this._def.valueType}static create(e,n,r){return n instanceof Ce?new or({keyType:e,valueType:n,typeName:he.ZodRecord,...we(r)}):new or({keyType:Tt.create(),valueType:e,typeName:he.ZodRecord,...we(n)})}}class zr extends Ce{_parse(e){const{status:n,ctx:r}=this._processInputParams(e);if(r.parsedType!==ee.map)return re(r,{code:X.invalid_type,expected:ee.map,received:r.parsedType}),_e;const i=this._def.keyType,s=this._def.valueType,o=[...r.data.entries()].map(([l,a],f)=>({key:i._parse(new qt(r,l,r.path,[f,"key"])),value:s._parse(new qt(r,a,r.path,[f,"value"]))}));if(r.common.async){const l=new Map;return Promise.resolve().then(async()=>{for(const a of o){const f=await a.key,u=await a.value;if(f.status==="aborted"||u.status==="aborted")return _e;(f.status==="dirty"||u.status==="dirty")&&n.dirty(),l.set(f.value,u.value)}return{status:n.value,value:l}})}else{const l=new Map;for(const a of o){const f=a.key,u=a.value;if(f.status==="aborted"||u.status==="aborted")return _e;(f.status==="dirty"||u.status==="dirty")&&n.dirty(),l.set(f.value,u.value)}return{status:n.value,value:l}}}}zr.create=(t,e,n)=>new zr({valueType:e,keyType:t,typeName:he.ZodMap,...we(n)});class yn extends Ce{_parse(e){const{status:n,ctx:r}=this._processInputParams(e);if(r.parsedType!==ee.set)return re(r,{code:X.invalid_type,expected:ee.set,received:r.parsedType}),_e;const i=this._def;i.minSize!==null&&r.data.size<i.minSize.value&&(re(r,{code:X.too_small,minimum:i.minSize.value,type:"set",inclusive:!0,exact:!1,message:i.minSize.message}),n.dirty()),i.maxSize!==null&&r.data.size>i.maxSize.value&&(re(r,{code:X.too_big,maximum:i.maxSize.value,type:"set",inclusive:!0,exact:!1,message:i.maxSize.message}),n.dirty());const s=this._def.valueType;function o(a){const f=new Set;for(const u of a){if(u.status==="aborted")return _e;u.status==="dirty"&&n.dirty(),f.add(u.value)}return{status:n.value,value:f}}const l=[...r.data.values()].map((a,f)=>s._parse(new qt(r,a,r.path,f)));return r.common.async?Promise.all(l).then(a=>o(a)):o(l)}min(e,n){return new yn({...this._def,minSize:{value:e,message:ue.toString(n)}})}max(e,n){return new yn({...this._def,maxSize:{value:e,message:ue.toString(n)}})}size(e,n){return this.min(e,n).max(e,n)}nonempty(e){return this.min(1,e)}}yn.create=(t,e)=>new yn({valueType:t,minSize:null,maxSize:null,typeName:he.ZodSet,...we(e)});class On extends Ce{constructor(){super(...arguments),this.validate=this.implement}_parse(e){const{ctx:n}=this._processInputParams(e);if(n.parsedType!==ee.function)return re(n,{code:X.invalid_type,expected:ee.function,received:n.parsedType}),_e;function r(l,a){return qr({data:l,path:n.path,errorMaps:[n.common.contextualErrorMap,n.schemaErrorMap,Nr(),er].filter(f=>!!f),issueData:{code:X.invalid_arguments,argumentsError:a}})}function i(l,a){return qr({data:l,path:n.path,errorMaps:[n.common.contextualErrorMap,n.schemaErrorMap,Nr(),er].filter(f=>!!f),issueData:{code:X.invalid_return_type,returnTypeError:a}})}const s={errorMap:n.common.contextualErrorMap},o=n.data;return this._def.returns instanceof Pn?pt(async(...l)=>{const a=new xt([]),f=await this._def.args.parseAsync(l,s).catch(c=>{throw a.addIssue(r(l,c)),a}),u=await o(...f);return await this._def.returns._def.type.parseAsync(u,s).catch(c=>{throw a.addIssue(i(u,c)),a})}):pt((...l)=>{const a=this._def.args.safeParse(l,s);if(!a.success)throw new xt([r(l,a.error)]);const f=o(...a.data),u=this._def.returns.safeParse(f,s);if(!u.success)throw new xt([i(f,u.error)]);return u.data})}parameters(){return this._def.args}returnType(){return this._def.returns}args(...e){return new On({...this._def,args:Ft.create(e).rest(pn.create())})}returns(e){return new On({...this._def,returns:e})}implement(e){return this.parse(e)}strictImplement(e){return this.parse(e)}static create(e,n,r){return new On({args:e||Ft.create([]).rest(pn.create()),returns:n||pn.create(),typeName:he.ZodFunction,...we(r)})}}class lr extends Ce{get schema(){return this._def.getter()}_parse(e){const{ctx:n}=this._processInputParams(e);return this._def.getter()._parse({data:n.data,path:n.path,parent:n})}}lr.create=(t,e)=>new lr({getter:t,typeName:he.ZodLazy,...we(e)});class ar extends Ce{_parse(e){if(e.data!==this._def.value){const n=this._getOrReturnCtx(e);return re(n,{received:n.data,code:X.invalid_literal,expected:this._def.value}),_e}return{status:"valid",value:e.data}}get value(){return this._def.value}}ar.create=(t,e)=>new ar({value:t,typeName:he.ZodLiteral,...we(e)});function Wo(t,e){return new tn({values:t,typeName:he.ZodEnum,...we(e)})}class tn extends Ce{_parse(e){if(typeof e.data!="string"){const n=this._getOrReturnCtx(e),r=this._def.values;return re(n,{expected:Oe.joinValues(r),received:n.parsedType,code:X.invalid_type}),_e}if(this._def.values.indexOf(e.data)===-1){const n=this._getOrReturnCtx(e),r=this._def.values;return re(n,{received:n.data,code:X.invalid_enum_value,options:r}),_e}return pt(e.data)}get options(){return this._def.values}get enum(){const e={};for(const n of this._def.values)e[n]=n;return e}get Values(){const e={};for(const n of this._def.values)e[n]=n;return e}get Enum(){const e={};for(const n of this._def.values)e[n]=n;return e}extract(e){return tn.create(e)}exclude(e){return tn.create(this.options.filter(n=>!e.includes(n)))}}tn.create=Wo;class ur extends Ce{_parse(e){const n=Oe.getValidEnumValues(this._def.values),r=this._getOrReturnCtx(e);if(r.parsedType!==ee.string&&r.parsedType!==ee.number){const i=Oe.objectValues(n);return re(r,{expected:Oe.joinValues(i),received:r.parsedType,code:X.invalid_type}),_e}if(n.indexOf(e.data)===-1){const i=Oe.objectValues(n);return re(r,{received:r.data,code:X.invalid_enum_value,options:i}),_e}return pt(e.data)}get enum(){return this._def.values}}ur.create=(t,e)=>new ur({values:t,typeName:he.ZodNativeEnum,...we(e)});class Pn extends Ce{unwrap(){return this._def.type}_parse(e){const{ctx:n}=this._processInputParams(e);if(n.parsedType!==ee.promise&&n.common.async===!1)return re(n,{code:X.invalid_type,expected:ee.promise,received:n.parsedType}),_e;const r=n.parsedType===ee.promise?n.data:Promise.resolve(n.data);return pt(r.then(i=>this._def.type.parseAsync(i,{path:n.path,errorMap:n.common.contextualErrorMap})))}}Pn.create=(t,e)=>new Pn({type:t,typeName:he.ZodPromise,...we(e)});class St extends Ce{innerType(){return this._def.schema}sourceType(){return this._def.schema._def.typeName===he.ZodEffects?this._def.schema.sourceType():this._def.schema}_parse(e){const{status:n,ctx:r}=this._processInputParams(e),i=this._def.effect||null;if(i.type==="preprocess"){const o=i.transform(r.data);return r.common.async?Promise.resolve(o).then(l=>this._def.schema._parseAsync({data:l,path:r.path,parent:r})):this._def.schema._parseSync({data:o,path:r.path,parent:r})}const s={addIssue:o=>{re(r,o),o.fatal?n.abort():n.dirty()},get path(){return r.path}};if(s.addIssue=s.addIssue.bind(s),i.type==="refinement"){const o=l=>{const a=i.refinement(l,s);if(r.common.async)return Promise.resolve(a);if(a instanceof Promise)throw new Error("Async refinement encountered during synchronous parse operation. Use .parseAsync instead.");return l};if(r.common.async===!1){const l=this._def.schema._parseSync({data:r.data,path:r.path,parent:r});return l.status==="aborted"?_e:(l.status==="dirty"&&n.dirty(),o(l.value),{status:n.value,value:l.value})}else return this._def.schema._parseAsync({data:r.data,path:r.path,parent:r}).then(l=>l.status==="aborted"?_e:(l.status==="dirty"&&n.dirty(),o(l.value).then(()=>({status:n.value,value:l.value}))))}if(i.type==="transform")if(r.common.async===!1){const o=this._def.schema._parseSync({data:r.data,path:r.path,parent:r});if(!Fr(o))return o;const l=i.transform(o.value,s);if(l instanceof Promise)throw new Error("Asynchronous transform encountered during synchronous parse operation. Use .parseAsync instead.");return{status:n.value,value:l}}else return this._def.schema._parseAsync({data:r.data,path:r.path,parent:r}).then(o=>Fr(o)?Promise.resolve(i.transform(o.value,s)).then(l=>({status:n.value,value:l})):o);Oe.assertNever(i)}}St.create=(t,e,n)=>new St({schema:t,typeName:he.ZodEffects,effect:e,...we(n)});St.createWithPreprocess=(t,e,n)=>new St({schema:e,effect:{type:"preprocess",transform:t},typeName:he.ZodEffects,...we(n)});class zt extends Ce{_parse(e){return this._getType(e)===ee.undefined?pt(void 0):this._def.innerType._parse(e)}unwrap(){return this._def.innerType}}zt.create=(t,e)=>new zt({innerType:t,typeName:he.ZodOptional,...we(e)});class bn extends Ce{_parse(e){return this._getType(e)===ee.null?pt(null):this._def.innerType._parse(e)}unwrap(){return this._def.innerType}}bn.create=(t,e)=>new bn({innerType:t,typeName:he.ZodNullable,...we(e)});class cr extends Ce{_parse(e){const{ctx:n}=this._processInputParams(e);let r=n.data;return n.parsedType===ee.undefined&&(r=this._def.defaultValue()),this._def.innerType._parse({data:r,path:n.path,parent:n})}removeDefault(){return this._def.innerType}}cr.create=(t,e)=>new cr({innerType:t,typeName:he.ZodDefault,defaultValue:typeof e.default=="function"?e.default:()=>e.default,...we(e)});class Zr extends Ce{_parse(e){const{ctx:n}=this._processInputParams(e),r={...n,common:{...n.common,issues:[]}},i=this._def.innerType._parse({data:r.data,path:r.path,parent:{...r}});return Lr(i)?i.then(s=>({status:"valid",value:s.status==="valid"?s.value:this._def.catchValue({get error(){return new xt(r.common.issues)},input:r.data})})):{status:"valid",value:i.status==="valid"?i.value:this._def.catchValue({get error(){return new xt(r.common.issues)},input:r.data})}}removeCatch(){return this._def.innerType}}Zr.create=(t,e)=>new Zr({innerType:t,typeName:he.ZodCatch,catchValue:typeof e.catch=="function"?e.catch:()=>e.catch,...we(e)});class jr extends Ce{_parse(e){if(this._getType(e)!==ee.nan){const r=this._getOrReturnCtx(e);return re(r,{code:X.invalid_type,expected:ee.nan,received:r.parsedType}),_e}return{status:"valid",value:e.data}}}jr.create=t=>new jr({typeName:he.ZodNaN,...we(t)});const ja=Symbol("zod_brand");class Go extends Ce{_parse(e){const{ctx:n}=this._processInputParams(e),r=n.data;return this._def.type._parse({data:r,path:n.path,parent:n})}unwrap(){return this._def.type}}class br extends Ce{_parse(e){const{status:n,ctx:r}=this._processInputParams(e);if(r.common.async)return(async()=>{const s=await this._def.in._parseAsync({data:r.data,path:r.path,parent:r});return s.status==="aborted"?_e:s.status==="dirty"?(n.dirty(),Vo(s.value)):this._def.out._parseAsync({data:s.value,path:r.path,parent:r})})();{const i=this._def.in._parseSync({data:r.data,path:r.path,parent:r});return i.status==="aborted"?_e:i.status==="dirty"?(n.dirty(),{status:"dirty",value:i.value}):this._def.out._parseSync({data:i.value,path:r.path,parent:r})}}static create(e,n){return new br({in:e,out:n,typeName:he.ZodPipeline})}}const Qo=(t,e={},n)=>t?Ln.create().superRefine((r,i)=>{var s,o;if(!t(r)){const l=typeof e=="function"?e(r):typeof e=="string"?{message:e}:e,a=(o=(s=l.fatal)!==null&&s!==void 0?s:n)!==null&&o!==void 0?o:!0,f=typeof l=="string"?{message:l}:l;i.addIssue({code:"custom",...f,fatal:a})}}):Ln.create(),Ha={object:ze.lazycreate};var he;(function(t){t.ZodString="ZodString",t.ZodNumber="ZodNumber",t.ZodNaN="ZodNaN",t.ZodBigInt="ZodBigInt",t.ZodBoolean="ZodBoolean",t.ZodDate="ZodDate",t.ZodSymbol="ZodSymbol",t.ZodUndefined="ZodUndefined",t.ZodNull="ZodNull",t.ZodAny="ZodAny",t.ZodUnknown="ZodUnknown",t.ZodNever="ZodNever",t.ZodVoid="ZodVoid",t.ZodArray="ZodArray",t.ZodObject="ZodObject",t.ZodUnion="ZodUnion",t.ZodDiscriminatedUnion="ZodDiscriminatedUnion",t.ZodIntersection="ZodIntersection",t.ZodTuple="ZodTuple",t.ZodRecord="ZodRecord",t.ZodMap="ZodMap",t.ZodSet="ZodSet",t.ZodFunction="ZodFunction",t.ZodLazy="ZodLazy",t.ZodLiteral="ZodLiteral",t.ZodEnum="ZodEnum",t.ZodEffects="ZodEffects",t.ZodNativeEnum="ZodNativeEnum",t.ZodOptional="ZodOptional",t.ZodNullable="ZodNullable",t.ZodDefault="ZodDefault",t.ZodCatch="ZodCatch",t.ZodPromise="ZodPromise",t.ZodBranded="ZodBranded",t.ZodPipeline="ZodPipeline"})(he||(he={}));const Ba=(t,e={message:`Input not instance of ${t.name}`})=>Qo(n=>n instanceof t,e),Yo=Tt.create,Xo=Jt.create,Va=jr.create,Wa=en.create,Ko=tr.create,Ga=vn.create,Qa=Pr.create,Ya=nr.create,Xa=rr.create,Ka=Ln.create,Ja=pn.create,eu=jt.create,tu=Ur.create,nu=Dt.create,ru=ze.create,iu=ze.strictCreate,su=ir.create,ou=oi.create,lu=sr.create,au=Ft.create,uu=or.create,cu=zr.create,fu=yn.create,du=On.create,hu=lr.create,pu=ar.create,mu=tn.create,gu=ur.create,_u=Pn.create,_s=St.create,vu=zt.create,yu=bn.create,bu=St.createWithPreprocess,wu=br.create,ku=()=>Yo().optional(),Eu=()=>Xo().optional(),Au=()=>Ko().optional(),Tu={string:t=>Tt.create({...t,coerce:!0}),number:t=>Jt.create({...t,coerce:!0}),boolean:t=>tr.create({...t,coerce:!0}),bigint:t=>en.create({...t,coerce:!0}),date:t=>vn.create({...t,coerce:!0})},Cu=_e;var De=Object.freeze({__proto__:null,defaultErrorMap:er,setErrorMap:Ia,getErrorMap:Nr,makeIssue:qr,EMPTY_PATH:$a,addIssueToContext:re,ParseStatus:ct,INVALID:_e,DIRTY:Vo,OK:pt,isAborted:Si,isDirty:Ii,isValid:Fr,isAsync:Lr,get util(){return Oe},get objectUtil(){return Di},ZodParsedType:ee,getParsedType:Yt,ZodType:Ce,ZodString:Tt,ZodNumber:Jt,ZodBigInt:en,ZodBoolean:tr,ZodDate:vn,ZodSymbol:Pr,ZodUndefined:nr,ZodNull:rr,ZodAny:Ln,ZodUnknown:pn,ZodNever:jt,ZodVoid:Ur,ZodArray:Dt,ZodObject:ze,ZodUnion:ir,ZodDiscriminatedUnion:oi,ZodIntersection:sr,ZodTuple:Ft,ZodRecord:or,ZodMap:zr,ZodSet:yn,ZodFunction:On,ZodLazy:lr,ZodLiteral:ar,ZodEnum:tn,ZodNativeEnum:ur,ZodPromise:Pn,ZodEffects:St,ZodTransformer:St,ZodOptional:zt,ZodNullable:bn,ZodDefault:cr,ZodCatch:Zr,ZodNaN:jr,BRAND:ja,ZodBranded:Go,ZodPipeline:br,custom:Qo,Schema:Ce,ZodSchema:Ce,late:Ha,get ZodFirstPartyTypeKind(){return he},coerce:Tu,any:Ka,array:nu,bigint:Wa,boolean:Ko,date:Ga,discriminatedUnion:ou,effect:_s,enum:mu,function:du,instanceof:Ba,intersection:lu,lazy:hu,literal:pu,map:cu,nan:Va,nativeEnum:gu,never:eu,null:Xa,nullable:yu,number:Xo,object:ru,oboolean:Au,onumber:Eu,optional:vu,ostring:ku,pipeline:wu,preprocess:bu,promise:_u,record:uu,set:fu,strictObject:iu,string:Yo,symbol:Qa,transformer:_s,tuple:au,undefined:Ya,union:su,unknown:Ja,void:tu,NEVER:Cu,ZodIssueCode:X,quotelessJson:Sa,ZodError:xt});const Oi={MalformedMessage:"Unknown client sent invalid message",InvalidToken:"Invalid token provided",UsernameTaken:"Username already in use",InappropriateName:"That name is not allowed/inappropriate choose another name",NotJoinable:"Quiz is not joinable",CapacityReached:"Quiz is full",UnknownPlayer:"Target player not found",Unexpected:"Unexpected error occurred",InvalidPermission:"You don't have permission to do that",UnexpectedMessage:"Client and server out of sync",InvalidAnswer:"Invalid answer type"};var xn=(t=>(t.None="None",t.Low="Low",t.Medium="Medium",t.High="High",t))(xn||{}),st=(t=>(t.Lobby="Lobby",t.Starting="Starting",t.AwaitingReady="AwaitingReady",t.PreQuestion="PreQuestion",t.AwaitingAnswers="AwaitingAnswers",t.Marked="Marked",t.Finished="Finished",t))(st||{}),Un=(t=>(t.Start="Start",t.Next="Next",t.Reset="Reset",t))(Un||{}),Jo=(t=>(t.RemovedByHost="RemovedByHost",t.HostDisconnect="HostDisconnect",t.LostConnection="LostConnection",t.Disconnected="Disconnected",t))(Jo||{});const xu={RemovedByHost:"Removed by host",HostDisconnect:"Quiz host left",LostConnection:"Connection lost",Disconnected:"Disconnected"};var rt=(t=>(t.Single="Single",t.Multiple="Multiple",t))(rt||{});const Du={Single:"There can be multiple correct answers but the user can only choose 1",Multiple:"There can be 1 or more correct answers and the user can select many"},vs=De.object({id:De.number(),value:De.string(),correct:De.boolean()});var Dn=(t=>(t.Contain="Contain",t.Cover="Cover",t.Width="Width",t.Height="Height",t))(Dn||{});const Su=De.object({text:De.string(),image:De.object({uuid:De.string().uuid(),fit:De.enum(["Contain","Cover","Width","Height"])}).nullable(),answer_time:De.number(),bonus_score_time:De.number(),scoring:De.object({min_score:De.number(),max_score:De.number(),bonus_score:De.number()})}).and(De.discriminatedUnion("ty",[De.object({ty:De.literal("Single"),answers:De.array(vs)}),De.object({ty:De.literal("Multiple"),answers:De.array(vs),correct_answers:De.number()})]));var Mi=(t=>(t.Single="Single",t.Multiple="Multiple",t))(Mi||{}),Zt=(t=>(t.Correct="Correct",t.Incorrect="Incorrect",t.Partial="Partial",t))(Zt||{}),Ht=(t=>(t.Initialize="Initialize",t.Connect="Connect",t.Join="Join",t.Ready="Ready",t.HostAction="HostAction",t.Answer="Answer",t.Kick="Kick",t))(Ht||{}),Et=(t=>(t.Joined="Joined",t.Ok="Ok",t.PlayerData="PlayerData",t.GameState="GameState",t.Timer="Timer",t.Question="Question",t.Scores="Scores",t.Score="Score",t.Error="Error",t.Kicked="Kicked",t))(Et||{});function Wi(t){return new URL(t,window.location.origin)}function ys(t,e){return Wi(`/api/quiz/${t}/${e}`).toString()}function Iu(t,e,n){return new Promise((r,i)=>{const s=$u(t,e),o=new XMLHttpRequest;o.upload.onprogress=n,o.responseType="json",o.onload=()=>{if(Math.floor(o.status/100)!==2)return console.error("Failed invalid request",o),i(new Error("Invalid request try reloading the page or updating Quizler"));const f=o.response;r(f.uuid)},o.onerror=o.ontimeout=o.onabort=()=>i(new Error("Failed to connect"));const l=Wi("/api/quiz");o.open("POST",l),o.send(s)})}function $u(t,e){const n=new FormData;n.append("config",JSON.stringify(t));for(const r of e)t.questions.find(s=>{var o;return((o=s.image)==null?void 0:o.uuid)===r.uuid})!==void 0&&n.append(r.uuid,r.blob);return n}async function Ou(t,e){const n=e.image;if(n===null)return null;const r=6;let i=0,s=ys(t,n.uuid),o=null;for(;i<r;)try{await new Promise((l,a)=>{o=new Image,o.src=s,o.onload=l,o.onerror=a}),console.debug("Preloaded question image",s);break}catch(l){console.error("Failed to preload image trying again",s,l),i+=1,o=null,s=ys(t,n.uuid)+`?attempt=${i}`}return o}let el=0,Hr={},fr=[],tl=nl();const Ri={},Gi=Bt(!1);function Gt(t,e){Uo(()=>{Ri[t]=e,console.log("Added handler for",t),fr=fr.filter(n=>n.ty===t?(e(n),!1):!0)}),ua(()=>{delete Ri[t]})}function Mu(){for(const t of fr)console.warn("Message was not handled",t);fr=[]}function nl(){const t=Wi("/api/quiz/socket");t.protocol=t.protocol.includes("https")?"wss":"ws",console.debug("Connecting to socket server "+t);const e=new WebSocket(t);return e.onmessage=qu,e.onopen=()=>{e.readyState==WebSocket.OPEN?(console.debug("Connected to socket"),Gi.set(!0)):console.log(e.readyState)},e.onclose=n=>{console.error("WebSocket connetion closed",n),Ru()},e.onerror=n=>{console.error("WebSocket error",n)},e}function Ru(){el=0,Hr={},Mu(),yr(),Nu()}function Nu(){console.debug("Socket connection lost (Reconnecting in 1000ms)"),Gi.set(!1),setTimeout(()=>{tl=nl()},1e3)}function nn(t){return new Promise((e,n)=>{console.debug("Sending message to server",t),t.rid=el++,Hr[t.rid]={resolve:e,reject:n};const r=JSON.stringify(t);try{tl.send(r)}catch(i){console.error("Failed to send message",i),delete Hr[t.rid],n({ty:Et.Error,error:"Unexpected"})}})}function qu({data:t}){const e=JSON.parse(t);if(e.ty===void 0){console.error("Packet missing message type",t);return}const n=e.rid;if(n!==void 0){const i=Hr[n];if(i===void 0){console.error(`Missing return handle ${n} for message`,e);return}e.ty===Et.Error?i.reject(e.error):i.resolve(e)}const r=Ri[e.ty];if(r!==void 0){r(e);return}fr.push(e)}function Fu(t){let e;return{c(){e=D("div"),m(e,"class","spin svelte-1ku7j7d")},m(n,r){W(n,e,r)},p:Q,i:Q,o:Q,d(n){n&&V(e)}}}class rl extends be{constructor(e){super(),ye(this,e,null,Fu,ve,{})}}function Lu(t){let e,n,r,i,s,o;return n=new rl({}),{c(){e=D("main"),le(n.$$.fragment),r=U(),i=D("p"),s=te(t[0]),m(i,"class","text svelte-ce1y0i"),m(e,"class","page page--center page--overflow")},m(l,a){W(l,e,a),se(n,e,null),g(e,r),g(e,i),g(i,s),o=!0},p(l,[a]){(!o||a&1)&&Ie(s,l[0])},i(l){o||(B(n.$$.fragment,l),o=!0)},o(l){G(n.$$.fragment,l),o=!1},d(l){l&&V(e),oe(n)}}}function Pu(t,e,n){let{text:r}=e;return t.$$set=i=>{"text"in i&&n(0,r=i.text)},[r]}class il extends be{constructor(e){super(),ye(this,e,Pu,Lu,ve,{text:0})}}let Tr;const Uu=new Uint8Array(16);function zu(){if(!Tr&&(Tr=typeof crypto<"u"&&crypto.getRandomValues&&crypto.getRandomValues.bind(crypto),!Tr))throw new Error("crypto.getRandomValues() not supported. See https://github.com/uuidjs/uuid#getrandomvalues-not-supported");return Tr(Uu)}const et=[];for(let t=0;t<256;++t)et.push((t+256).toString(16).slice(1));function Zu(t,e=0){return(et[t[e+0]]+et[t[e+1]]+et[t[e+2]]+et[t[e+3]]+"-"+et[t[e+4]]+et[t[e+5]]+"-"+et[t[e+6]]+et[t[e+7]]+"-"+et[t[e+8]]+et[t[e+9]]+"-"+et[t[e+10]]+et[t[e+11]]+et[t[e+12]]+et[t[e+13]]+et[t[e+14]]+et[t[e+15]]).toLowerCase()}const ju=typeof crypto<"u"&&crypto.randomUUID&&crypto.randomUUID.bind(crypto),bs={randomUUID:ju};function Qi(t,e,n){if(bs.randomUUID&&!e&&!t)return bs.randomUUID();t=t||{};const r=t.random||(t.rng||zu)();if(r[6]=r[6]&15|64,r[8]=r[8]&63|128,e){n=n||0;for(let i=0;i<16;++i)e[n+i]=r[i];return e}return Zu(r)}const Dr=5,sl=1,ol=30,Hu=70,Bu=300,Vu="Example Quiz",Wu="Small description about your quiz",Gu=10,Qu=100,Yu=150,ws=8,ks=50,Xu=400,Ku=1e3*15,Ju=1e3,ec=1e3*60*30,tc=150,nc=1e3*3,Es=0,rc=1e4,ic=1e4,sc=1e3,oc=1e3*60*30,lc=1,ll=50,ac=1024*1024*15;function uc(){return{name:Vu,text:Wu,max_players:ll,filtering:xn.High,questions:[al()]}}function al(){return{id:Qi(),ty:rt.Single,text:"This is an example question, you should replace this with your first question",image:null,answer_time:Ku,bonus_score_time:nc,scoring:{min_score:Gu,max_score:Qu,bonus_score:tc},answers:[{id:0,value:"Example A",correct:!0},{id:1,value:"Example B",correct:!1},{id:2,value:"Example C",correct:!1},{id:3,value:"Example D",correct:!1}]}}function cc(t){let e,n,r,i,s,o,l;return n=new rl({}),{c(){e=D("main"),le(n.$$.fragment),r=U(),i=D("p"),s=te(t[0]),m(i,"class","text svelte-1cn2ffa"),m(e,"class","main svelte-1cn2ffa")},m(a,f){W(a,e,f),se(n,e,null),g(e,r),g(e,i),g(i,s),l=!0},p(a,[f]){(!l||f&1)&&Ie(s,a[0])},i(a){l||(B(n.$$.fragment,a),Le(()=>{l&&(o||(o=Ae(e,Nt,{},!0)),o.run(1))}),l=!0)},o(a){G(n.$$.fragment,a),o||(o=Ae(e,Nt,{},!1)),o.run(0),l=!1},d(a){a&&V(e),oe(n),a&&o&&o.end()}}}function fc(t,e,n){let{text:r=""}=e;return t.$$set=i=>{"text"in i&&n(0,r=i.text)},[r]}class Yi extends be{constructor(e){super(),ye(this,e,fc,cc,ve,{text:0})}}function dc(t){let e,n,r;return{c(){e=ce("svg"),n=ce("path"),r=ce("path"),m(n,"d","M0 0h24v24H0z"),m(n,"fill","none"),m(r,"d","M15.41 7.41L14 6l-6 6 6 6 1.41-1.41L10.83 12z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 0 24 24"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(i,s){W(i,e,s),g(e,n),g(e,r)},p:Q,i:Q,o:Q,d(i){i&&V(e)}}}class ul extends be{constructor(e){super(),ye(this,e,null,dc,ve,{})}}function hc(t){let e,n,r;return{c(){e=ce("svg"),n=ce("path"),r=ce("path"),m(n,"d","M0 0h24v24H0z"),m(n,"fill","none"),m(r,"d","M8 5v14l11-7z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 0 24 24"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(i,s){W(i,e,s),g(e,n),g(e,r)},p:Q,i:Q,o:Q,d(i){i&&V(e)}}}class li extends be{constructor(e){super(),ye(this,e,null,hc,ve,{})}}function As(t){let e,n;return e=new Yi({}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function pc(t){let e,n,r,i,s,o,l,a,f,u,d,c,h,p,y,k=t[4]&&Ts(t);return{c(){e=D("main"),n=D("p"),r=te(t[1]),i=U(),s=D("h1"),s.textContent="Enter Name",o=U(),l=D("p"),l.textContent="Please enter your desired name",a=U(),f=D("div"),u=D("input"),d=U(),k&&k.c(),m(u,"class","input input--small svelte-122cy5o"),m(u,"type","text"),m(u,"minlength",sl),m(u,"maxlength",ol),m(f,"class","form svelte-122cy5o"),m(e,"class","page page--center page--overflow svelte-122cy5o")},m(_,C){W(_,e,C),g(e,n),g(n,r),g(e,i),g(e,s),g(e,o),g(e,l),g(e,a),g(e,f),g(f,u),Ze(u,t[2]),g(f,d),k&&k.m(f,null),h=!0,p||(y=[J(u,"input",t[13]),J(u,"input",t[8])],p=!0)},p(_,C){(!h||C&2)&&Ie(r,_[1]),C&4&&u.value!==_[2]&&Ze(u,_[2]),_[4]?k?(k.p(_,C),C&16&&B(k,1)):(k=Ts(_),k.c(),B(k,1),k.m(f,null)):k&&(qe(),G(k,1,1,()=>{k=null}),Fe())},i(_){h||(B(k),Le(()=>{h&&(c||(c=Ae(e,Me,{},!0)),c.run(1))}),h=!0)},o(_){G(k),c||(c=Ae(e,Me,{},!1)),c.run(0),h=!1},d(_){_&&V(e),k&&k.d(),_&&c&&c.end(),p=!1,je(y)}}}function mc(t){let e,n,r,i,s,o,l,a,f,u,d,c,h=t[3]&&Cs(t);return{c(){e=D("main"),n=D("h1"),n.textContent="Enter Code",r=U(),i=D("p"),i.textContent="Please enter your quiz code below",s=U(),o=D("div"),l=D("input"),a=U(),h&&h.c(),m(l,"class","input svelte-122cy5o"),m(l,"type","text"),m(l,"minlength",Dr),m(l,"maxlength",Dr),m(l,"placeholder","X".repeat(Dr)),m(o,"class","form svelte-122cy5o"),m(e,"class","page page--center page--overflow svelte-122cy5o")},m(p,y){W(p,e,y),g(e,n),g(e,r),g(e,i),g(e,s),g(e,o),g(o,l),Ze(l,t[1]),g(o,a),h&&h.m(o,null),u=!0,d||(c=[J(l,"input",t[12]),J(l,"input",t[7])],d=!0)},p(p,y){y&2&&l.value!==p[1]&&Ze(l,p[1]),p[3]?h?(h.p(p,y),y&8&&B(h,1)):(h=Cs(p),h.c(),B(h,1),h.m(o,null)):h&&(qe(),G(h,1,1,()=>{h=null}),Fe())},i(p){u||(B(h),Le(()=>{u&&(f||(f=Ae(e,Me,{},!0)),f.run(1))}),u=!0)},o(p){G(h),f||(f=Ae(e,Me,{},!1)),f.run(0),u=!1},d(p){p&&V(e),h&&h.d(),p&&f&&f.end(),d=!1,je(c)}}}function Ts(t){let e,n,r,i,s,o;return n=new li({}),{c(){e=D("button"),le(n.$$.fragment),m(e,"class","play play--small svelte-122cy5o")},m(l,a){W(l,e,a),se(n,e,null),i=!0,s||(o=J(e,"click",t[10]),s=!0)},p:Q,i(l){i||(B(n.$$.fragment,l),Le(()=>{i&&(r||(r=Ae(e,Me,{axis:"x",duration:200},!0)),r.run(1))}),i=!0)},o(l){G(n.$$.fragment,l),r||(r=Ae(e,Me,{axis:"x",duration:200},!1)),r.run(0),i=!1},d(l){l&&V(e),oe(n),l&&r&&r.end(),s=!1,o()}}}function Cs(t){let e,n,r,i,s,o;return n=new li({}),{c(){e=D("button"),le(n.$$.fragment),m(e,"class","play svelte-122cy5o")},m(l,a){W(l,e,a),se(n,e,null),i=!0,s||(o=J(e,"click",t[9]),s=!0)},p:Q,i(l){i||(B(n.$$.fragment,l),Le(()=>{i&&(r||(r=Ae(e,Me,{axis:"x",duration:200},!0)),r.run(1))}),i=!0)},o(l){G(n.$$.fragment,l),r||(r=Ae(e,Me,{axis:"x",duration:200},!1)),r.run(0),i=!1},d(l){l&&V(e),oe(n),l&&r&&r.end(),s=!1,o()}}}function gc(t){let e,n,r,i,s,o,l,a,f,u,d=t[5]&&As();r=new ul({});const c=[mc,pc],h=[];function p(y,k){return y[0]===y[6].Connect?0:1}return s=p(t),o=h[s]=c[s](t),{c(){d&&d.c(),e=U(),n=D("button"),le(r.$$.fragment),i=U(),o.c(),l=Lt(),m(n,"class","back back--floating")},m(y,k){d&&d.m(y,k),W(y,e,k),W(y,n,k),se(r,n,null),W(y,i,k),h[s].m(y,k),W(y,l,k),a=!0,f||(u=J(n,"click",t[11]),f=!0)},p(y,[k]){y[5]?d?k&32&&B(d,1):(d=As(),d.c(),B(d,1),d.m(e.parentNode,e)):d&&(qe(),G(d,1,1,()=>{d=null}),Fe());let _=s;s=p(y),s===_?h[s].p(y,k):(qe(),G(h[_],1,1,()=>{h[_]=null}),Fe(),o=h[s],o?o.p(y,k):(o=h[s]=c[s](y),o.c()),B(o,1),o.m(l.parentNode,l))},i(y){a||(B(d),B(r.$$.fragment,y),B(o),a=!0)},o(y){G(d),G(r.$$.fragment,y),G(o),a=!1},d(y){d&&d.d(y),y&&V(e),y&&V(n),oe(r),y&&V(i),h[s].d(y),y&&V(l),f=!1,u()}}}function _c(t,e,n){var r=(_=>(_[_.Connect=0]="Connect",_[_.Join=1]="Join",_))(r||{});let i=0,s="",o="",l=!1,a=!1,f=!1;function u(){n(1,s=s.toUpperCase().replace(/[^A-Z0-9]/,"")),n(3,l=s.length===Dr)}function d(){n(4,a=o.length>=sl&&o.length<=ol)}function c(){n(5,f=!0),nn({ty:Ht.Connect,token:s}).then(()=>{n(0,i=1)}).catch(_=>{console.error("Failed to connect",_),Qt("Failed to connect",Oi[_])}).finally(()=>n(5,f=!1))}function h(){n(5,f=!0),nn({ty:Ht.Join,name:o}).then(({id:_,token:C,config:T})=>{Zo({id:_,token:C,config:T,host:!1,name:o})}).catch(_=>{console.error("Failed to join",_),Qt("Failed to join",Oi[_])}).finally(()=>n(5,f=!1))}function p(){i===0?yr():n(0,i=0)}function y(){s=this.value,n(1,s)}function k(){o=this.value,n(2,o)}return[i,s,o,l,a,f,r,u,d,c,h,p,y,k]}class vc extends be{constructor(e){super(),ye(this,e,_c,gc,ve,{})}}async function cl(t){try{await nn({ty:Ht.Kick,id:t})}catch(e){console.error("Error while attempting to kick",e)}}async function dr(t){try{await nn({ty:Ht.HostAction,action:t})}catch(e){console.error("Error while attempting host action",t,e)}}async function yc(){try{await nn({ty:Ht.Ready})}catch(t){console.error("Error while attempting to ready",t)}}async function Xi(t){const e=t.host?"Are you sure you want to leave? Leaving will remove all other players from the game":"Are you sure you want to leave?";await Vi("Confirm Leave",e)&&(yr(),cl(t.id))}function Br(t){return((t.total-t.elapsed)/1e3).toFixed(0)}function xs(t){const e=["bytes","kb","mb"];if(t==0)return"0 Byte";const n=Math.floor(Math.log(t)/Math.log(1024));return(t/Math.pow(1024,n)).toFixed(1)+e[n]}function mn(t,e){const n=Math.round(Math.random()*(e-t)+t);return n>e?e:n<t?t:n}function Ds(t){const e=["th","st","nd","rd"],n=t%100;return t+(e[(n-20)%10]||e[n]||e[0])}function bc(t){const e=mn(1,t.length/2);let n=0;for(;n<e;){const r=mn(0,t.length-1),i=mn(0,t.length-1);r!==i&&(fl(t,r,i),n++)}return t}function fl(t,e,n){const r=t[e],i=t[n];return(r!==void 0||i!==void 0)&&(t[e]=i,t[n]=r),t}const wc={Answered:["Hmm I wonder if you got it right....","It definitely had to be that one! 👌","Yeah it was probably that one... 😬","0_0 good luck i guess..","Are you sure it was that one..?","😅 Probbably right","It gotta be that one!","If you say so 🫠","1 + 1 = 5?","Was definitely that one! 😎","🤫 don't tell them the answer","Made it in time 😮‍💨","Nice answer 🥳","Mastermind 🧠"],[Zt.Correct]:["You did it!","That one was right!","Good job!","Yup that was it!","Nicely done 👍","Brainy 🧠","Smarty pants 🤓"],[Zt.Incorrect]:["Ooops..","Yeah not that one... 🤨","Better luck next time 🤧","Noooo your other left","Uh oh.... 😨","Yeah nope.... ","😤 Maybe next time"],[Zt.Partial]:["Almost there...","Not quite there","Just about...","Yeah not that one... 🤨","Better luck next time 🤧","Noooo your other left","Uh oh.... 😨","Yeah nope.... ","😤 Maybe next time"]};function dl(t){const e=wc[t];return e[mn(0,e.length-1)]}function kc(t){let e,n,r,i,s,o;return{c(){e=D("main"),n=D("h1"),n.textContent="Waiting...",r=U(),i=D("p"),i.textContent=`${t[0]}`,m(n,"class","title svelte-1p2ou5x"),m(i,"class","text svelte-1p2ou5x"),m(e,"class","main svelte-1p2ou5x")},m(l,a){W(l,e,a),g(e,n),g(e,r),g(e,i),o=!0},p:Q,i(l){o||(Le(()=>{o&&(s||(s=Ae(e,Me,{},!0)),s.run(1))}),o=!0)},o(l){s||(s=Ae(e,Me,{},!1)),s.run(0),o=!1},d(l){l&&V(e),l&&s&&s.end()}}}function Ec(t){return[dl("Answered")]}class Ac extends be{constructor(e){super(),ye(this,e,Ec,kc,ve,{})}}function wr(t,{from:e,to:n},r={}){const i=getComputedStyle(t),s=i.transform==="none"?"":i.transform,[o,l]=i.transformOrigin.split(" ").map(parseFloat),a=e.left+e.width*o/n.width-(n.left+o),f=e.top+e.height*l/n.height-(n.top+l),{delay:u=0,duration:d=h=>Math.sqrt(h)*120,easing:c=jo}=r;return{delay:u,duration:wn(d)?d(Math.sqrt(a*a+f*f)):d,easing:c,css:(h,p)=>{const y=p*a,k=p*f,_=h+p*e.width/n.width,C=h+p*e.height/n.height;return`transform: ${s} translate(${y}px, ${k}px) scale(${_}, ${C});`}}}function Ss(t){return Object.prototype.toString.call(t)==="[object Date]"}function Ni(t,e){if(t===e||t!==t)return()=>t;const n=typeof t;if(n!==typeof e||Array.isArray(t)!==Array.isArray(e))throw new Error("Cannot interpolate values of different type");if(Array.isArray(t)){const r=e.map((i,s)=>Ni(t[s],i));return i=>r.map(s=>s(i))}if(n==="object"){if(!t||!e)throw new Error("Object cannot be null");if(Ss(t)&&Ss(e)){t=t.getTime(),e=e.getTime();const s=e-t;return o=>new Date(t+o*s)}const r=Object.keys(e),i={};return r.forEach(s=>{i[s]=Ni(t[s],e[s])}),s=>{const o={};return r.forEach(l=>{o[l]=i[l](s)}),o}}if(n==="number"){const r=e-t;return i=>t+i*r}throw new Error(`Cannot interpolate ${n} values`)}function Ki(t,e={}){const n=Bt(t);let r,i=t;function s(o,l){if(t==null)return n.set(t=o),Promise.resolve();i=o;let a=r,f=!1,{delay:u=0,duration:d=400,easing:c=ri,interpolate:h=Ni}=Ei(Ei({},e),l);if(d===0)return a&&(a.abort(),a=null),n.set(t=i),Promise.resolve();const p=Zi()+u;let y;return r=Hi(k=>{if(k<p)return!0;f||(y=h(t,o),typeof d=="function"&&(d=d(t,o)),f=!0),a&&(a.abort(),a=null);const _=k-p;return _>d?(n.set(t=o),!1):(n.set(t=y(c(_/d))),!0)}),r.promise}return{set:s,update:(o,l)=>s(o(i,t),l),subscribe:n.subscribe}}function Tc(t){let e=t[0].toFixed(0)+"",n;return{c(){n=te(e)},m(r,i){W(r,n,i)},p(r,[i]){i&1&&e!==(e=r[0].toFixed(0)+"")&&Ie(n,e)},i:Q,o:Q,d(r){r&&V(n)}}}function Cc(t,e,n){let r,{value:i}=e;const s=Ki(0,{delay:500});return lt(t,s,o=>n(0,r=o)),t.$$set=o=>{"value"in o&&n(2,i=o.value)},t.$$.update=()=>{t.$$.dirty&4&&s.set(i)},[r,s,i]}class hl extends be{constructor(e){super(),ye(this,e,Cc,Tc,ve,{value:2})}}function xc(t){let e,n,r,i,s,o,l,a,f;return{c(){e=ce("svg"),n=ce("g"),r=ce("polygon"),i=ce("polygon"),s=ce("polygon"),o=ce("g"),l=ce("circle"),a=ce("circle"),f=ce("path"),Ai(r,"fill","#FFD102"),m(r,"points","371.342,210.903 469.209,150.675 379.787,95.648 281.922,155.874 	"),Ai(i,"fill","#FFD102"),m(i,"points","127.994,210.903 225.863,150.675 136.44,95.648 38.576,155.874 	"),m(s,"fill","#FFE368"),m(s,"points",`383.997,197.691 255.992,118.921 127.994,197.691 0,118.921 22.505,382.593 489.495,382.593 
-	512,118.921 `),m(l,"fill","#FFAA0E"),m(l,"cx","99.947"),m(l,"cy","320.636"),m(l,"r","8.44"),m(a,"fill","#FFAA0E"),m(a,"cx","66.184"),m(a,"cy","286.878"),m(a,"r","8.44"),m(f,"fill","#FFAA0E"),m(f,"d",`M489.495,371.341H22.505C10.076,371.341,0,381.416,0,393.846c0,12.43,10.076,22.505,22.505,22.505
-		h466.989c12.43,0,22.505-10.076,22.505-22.505C512,381.416,501.924,371.341,489.495,371.341z`),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"xmlns:xlink","http://www.w3.org/1999/xlink"),m(e,"viewBox","0 0 512 512"),m(e,"xml:space","preserve")},m(u,d){W(u,e,d),g(e,n),g(n,r),g(n,i),g(e,s),g(e,o),g(o,l),g(o,a),g(o,f)},p:Q,i:Q,o:Q,d(u){u&&V(e)}}}class Dc extends be{constructor(e){super(),ye(this,e,null,xc,ve,{})}}function Is(t,e,n){const r=t.slice();return r[4]=e[n],r[6]=n,r}function $s(t){let e,n,r;return{c(){e=D("button"),e.textContent="Restart",m(e,"class","btn")},m(i,s){W(i,e,s),n||(r=J(e,"click",t[2]),n=!0)},p:Q,d(i){i&&V(e),n=!1,r()}}}function Os(t){let e,n,r;return n=new Dc({}),{c(){e=D("div"),le(n.$$.fragment),m(e,"class","crown svelte-9bcas5")},m(i,s){W(i,e,s),se(n,e,null),r=!0},i(i){r||(B(n.$$.fragment,i),r=!0)},o(i){G(n.$$.fragment,i),r=!1},d(i){i&&V(e),oe(n)}}}function Ms(t,e){let n,r,i,s=Ds(e[6]+1)+"",o,l,a,f=e[4].name+"",u,d,c,h,p,y,k=Q,_,C=e[6]==0&&Os();return h=new hl({props:{value:e[4].score}}),{key:t,first:null,c(){n=D("tr"),r=D("td"),C&&C.c(),i=U(),o=te(s),l=U(),a=D("td"),u=te(f),d=U(),c=D("td"),le(h.$$.fragment),p=U(),m(r,"class","player__place svelte-9bcas5"),m(a,"class","player__name svelte-9bcas5"),m(c,"class","player_score"),m(n,"class","player svelte-9bcas5"),this.first=n},m(T,S){W(T,n,S),g(n,r),C&&C.m(r,null),g(r,i),g(r,o),g(n,l),g(n,a),g(a,u),g(n,d),g(n,c),se(h,c,null),g(n,p),_=!0},p(T,S){e=T,e[6]==0?C?S&1&&B(C,1):(C=Os(),C.c(),B(C,1),C.m(r,i)):C&&(qe(),G(C,1,1,()=>{C=null}),Fe()),(!_||S&1)&&s!==(s=Ds(e[6]+1)+"")&&Ie(o,s),(!_||S&1)&&f!==(f=e[4].name+"")&&Ie(u,f);const $={};S&1&&($.value=e[4].score),h.$set($)},r(){y=n.getBoundingClientRect()},f(){gr(n),k()},a(){k(),k=mr(n,y,wr,{})},i(T){_||(B(C),B(h.$$.fragment,T),_=!0)},o(T){G(C),G(h.$$.fragment,T),_=!1},d(T){T&&V(n),C&&C.d(),oe(h)}}}function Sc(t){let e,n,r,i=t[1].token+"",s,o,l,a=t[1].config.name+"",f,u,d,c=t[1].config.text+"",h,p,y,k,_,C,T,S,$,w,x=[],v=new Map,b,I,O,R,N=t[1].host&&$s(t),M=t[0].players;const L=A=>A[4].id;for(let A=0;A<M.length;A+=1){let E=Is(t,M,A),q=L(E);v.set(q,x[A]=Ms(q,E))}return{c(){e=D("main"),n=D("div"),r=D("h1"),s=te(i),o=U(),l=D("h2"),f=te(a),u=U(),d=D("p"),h=te(c),p=U(),y=D("div"),k=D("button"),k.textContent="Leave",_=U(),N&&N.c(),C=U(),T=D("table"),S=D("thead"),S.innerHTML=`<tr><th>Place</th> 
-          <th>Name</th> 
-          <th>Score</th></tr>`,$=U(),w=D("tbody");for(let A=0;A<x.length;A+=1)x[A].c();m(r,"class","token svelte-9bcas5"),m(l,"class","name"),m(d,"class","desc svelte-9bcas5"),m(k,"class","btn"),m(y,"class","btn-row btn-row--fill actions svelte-9bcas5"),m(n,"class","quiz svelte-9bcas5"),m(e,"class","page page--middle page--overflow svelte-9bcas5")},m(A,E){W(A,e,E),g(e,n),g(n,r),g(r,s),g(n,o),g(n,l),g(l,f),g(n,u),g(n,d),g(d,h),g(n,p),g(n,y),g(y,k),g(y,_),N&&N.m(y,null),g(n,C),g(n,T),g(T,S),g(T,$),g(T,w);for(let q=0;q<x.length;q+=1)x[q]&&x[q].m(w,null);I=!0,O||(R=J(k,"click",t[3]),O=!0)},p(A,[E]){if((!I||E&2)&&i!==(i=A[1].token+"")&&Ie(s,i),(!I||E&2)&&a!==(a=A[1].config.name+"")&&Ie(f,a),(!I||E&2)&&c!==(c=A[1].config.text+"")&&Ie(h,c),A[1].host?N?N.p(A,E):(N=$s(A),N.c(),N.m(y,null)):N&&(N.d(1),N=null),E&1){M=A[0].players,qe();for(let q=0;q<x.length;q+=1)x[q].r();x=_r(x,E,L,1,A,M,v,w,ii,Ms,null,Is);for(let q=0;q<x.length;q+=1)x[q].a();Fe()}},i(A){if(!I){for(let E=0;E<M.length;E+=1)B(x[E]);Le(()=>{I&&(b||(b=Ae(e,Me,{},!0)),b.run(1))}),I=!0}},o(A){for(let E=0;E<x.length;E+=1)G(x[E]);b||(b=Ae(e,Me,{},!1)),b.run(0),I=!1},d(A){A&&V(e),N&&N.d();for(let E=0;E<x.length;E+=1)x[E].d();A&&b&&b.end(),O=!1,R()}}}function Ic(t,e,n){let{summary:r}=e,{gameData:i}=e;const s=()=>dr(Un.Reset),o=()=>Xi(i);return t.$$set=l=>{"summary"in l&&n(0,r=l.summary),"gameData"in l&&n(1,i=l.gameData)},[r,i,s,o]}class $c extends be{constructor(e){super(),ye(this,e,Ic,Sc,ve,{summary:0,gameData:1})}}function Rs(t,e,n){const r=t.slice();return r[12]=e[n],r[14]=n,r}function Ns(t,e,n){const r=t.slice();return r[12]=e[n],r[14]=n,r}function qs(t){let e,n,r,i;return{c(){e=D("div"),m(e,"class","image-wrapper svelte-b8ytla")},m(s,o){W(s,e,o),r||(i=Ro(n=Nc.call(null,e,t[3])),r=!0)},p(s,o){n&&wn(n.update)&&o&8&&n.update.call(null,s[3])},d(s){s&&V(e),r=!1,i()}}}function Oc(t){let e,n,r,i=t[2].answers,s=[];for(let l=0;l<i.length;l+=1)s[l]=Fs(Rs(t,i,l));let o=!t[0].host&&Ls(t);return{c(){e=D("div");for(let l=0;l<s.length;l+=1)s[l].c();n=U(),o&&o.c(),r=Lt(),m(e,"class","answers svelte-b8ytla")},m(l,a){W(l,e,a);for(let f=0;f<s.length;f+=1)s[f]&&s[f].m(e,null);W(l,n,a),o&&o.m(l,a),W(l,r,a)},p(l,a){if(a&149){i=l[2].answers;let f;for(f=0;f<i.length;f+=1){const u=Rs(l,i,f);s[f]?s[f].p(u,a):(s[f]=Fs(u),s[f].c(),s[f].m(e,null))}for(;f<s.length;f+=1)s[f].d(1);s.length=i.length}l[0].host?o&&(o.d(1),o=null):o?o.p(l,a):(o=Ls(l),o.c(),o.m(r.parentNode,r))},d(l){l&&V(e),Kn(s,l),l&&V(n),o&&o.d(l),l&&V(r)}}}function Mc(t){let e,n=t[2].answers,r=[];for(let i=0;i<n.length;i+=1)r[i]=Ps(Ns(t,n,i));return{c(){e=D("div");for(let i=0;i<r.length;i+=1)r[i].c();m(e,"class","answers svelte-b8ytla")},m(i,s){W(i,e,s);for(let o=0;o<r.length;o+=1)r[o]&&r[o].m(e,null)},p(i,s){if(s&37){n=i[2].answers;let o;for(o=0;o<n.length;o+=1){const l=Ns(i,n,o);r[o]?r[o].p(l,s):(r[o]=Ps(l),r[o].c(),r[o].m(e,null))}for(;o<r.length;o+=1)r[o].d(1);r.length=n.length}},d(i){i&&V(e),Kn(r,i)}}}function Fs(t){let e,n=t[12].value+"",r,i,s,o,l,a;function f(){return t[11](t[14])}return{c(){e=D("button"),r=te(n),i=U(),m(e,"data-host",s=t[0].host),m(e,"class","answer btn btn--surface svelte-b8ytla"),e.disabled=o=t[0].host||!t[4].includes(t[14])&&t[4].length===t[2].correct_answers,Or(e,"answer--checked",t[4].includes(t[14]))},m(u,d){W(u,e,d),g(e,r),g(e,i),l||(a=J(e,"click",f),l=!0)},p(u,d){t=u,d&4&&n!==(n=t[12].value+"")&&Ie(r,n),d&1&&s!==(s=t[0].host)&&m(e,"data-host",s),d&21&&o!==(o=t[0].host||!t[4].includes(t[14])&&t[4].length===t[2].correct_answers)&&(e.disabled=o),d&16&&Or(e,"answer--checked",t[4].includes(t[14]))},d(u){u&&V(e),l=!1,a()}}}function Ls(t){let e,n,r,i,s;return{c(){e=D("button"),n=te("Submit"),m(e,"class","btn btn btn--surface submit svelte-b8ytla"),e.disabled=r=t[4].length<1},m(o,l){W(o,e,l),g(e,n),i||(s=J(e,"click",t[6]),i=!0)},p(o,l){l&16&&r!==(r=o[4].length<1)&&(e.disabled=r)},d(o){o&&V(e),i=!1,s()}}}function Ps(t){let e,n=t[12].value+"",r,i,s,o,l,a;function f(){return t[10](t[14])}return{c(){e=D("button"),r=te(n),i=U(),m(e,"data-host",s=t[0].host),m(e,"class","answer btn btn--surface svelte-b8ytla"),e.disabled=o=t[0].host},m(u,d){W(u,e,d),g(e,r),g(e,i),l||(a=J(e,"click",f),l=!0)},p(u,d){t=u,d&4&&n!==(n=t[12].value+"")&&Ie(r,n),d&1&&s!==(s=t[0].host)&&m(e,"data-host",s),d&1&&o!==(o=t[0].host)&&(e.disabled=o)},d(u){u&&V(e),l=!1,a()}}}function Us(t){let e,n,r;return{c(){e=D("button"),e.textContent="Skip",m(e,"class","btn btn--surface")},m(i,s){W(i,e,s),n||(r=J(e,"click",t[8]),n=!0)},p:Q,d(i){i&&V(e),n=!1,r()}}}function Rc(t){let e,n,r,i,s=t[2].text+"",o,l,a,f,u,d=t[0].token+"",c,h,p,y,k=Br(t[1])+"",_,C=t[3]!==null&&qs(t);function T(x,v){if(x[2].ty===rt.Single)return Mc;if(x[2].ty===rt.Multiple)return Oc}let S=T(t),$=S&&S(t),w=t[0].host&&Us(t);return{c(){e=D("main"),C&&C.c(),n=U(),r=D("div"),i=D("p"),o=te(s),l=U(),$&&$.c(),a=U(),f=D("div"),u=D("p"),c=te(d),h=U(),w&&w.c(),p=U(),y=D("p"),_=te(k),m(i,"class","text svelte-b8ytla"),m(r,"class","content svelte-b8ytla"),m(u,"class","token svelte-b8ytla"),m(y,"class","time svelte-b8ytla"),m(f,"class","bottom svelte-b8ytla"),m(e,"class","main svelte-b8ytla")},m(x,v){W(x,e,v),C&&C.m(e,null),g(e,n),g(e,r),g(r,i),g(i,o),g(r,l),$&&$.m(r,null),g(e,a),g(e,f),g(f,u),g(u,c),g(f,h),w&&w.m(f,null),g(f,p),g(f,y),g(y,_)},p(x,[v]){x[3]!==null?C?C.p(x,v):(C=qs(x),C.c(),C.m(e,n)):C&&(C.d(1),C=null),v&4&&s!==(s=x[2].text+"")&&Ie(o,s),S===(S=T(x))&&$?$.p(x,v):($&&$.d(1),$=S&&S(x),$&&($.c(),$.m(r,null))),v&1&&d!==(d=x[0].token+"")&&Ie(c,d),x[0].host?w?w.p(x,v):(w=Us(x),w.c(),w.m(f,p)):w&&(w.d(1),w=null),v&2&&k!==(k=Br(x[1])+"")&&Ie(_,k)},i:Q,o:Q,d(x){x&&V(e),C&&C.d(),$&&$.d(),w&&w.d()}}}function Nc(t,e){t.appendChild(e)}function qc(t,e,n){let{gameData:r}=e,{timer:i}=e,{question:s}=e,{answered:o}=e,{preloadedImage:l}=e,a=[];async function f(y){n(9,o=!0);try{await nn({ty:Ht.Answer,answer:{ty:Mi.Single,answer:y}})}catch(k){console.error("Error while attempting to answer",k)}}async function u(){n(9,o=!0);try{await nn({ty:Ht.Answer,answer:{ty:Mi.Multiple,answers:a}})}catch(y){console.error("Error while attempting to answer",y)}}function d(y){a.includes(y)?n(4,a=a.filter(k=>k!=y)):(a.push(y),n(4,a))}const c=()=>dr(Un.Next),h=y=>f(y),p=y=>d(y);return t.$$set=y=>{"gameData"in y&&n(0,r=y.gameData),"timer"in y&&n(1,i=y.timer),"question"in y&&n(2,s=y.question),"answered"in y&&n(9,o=y.answered),"preloadedImage"in y&&n(3,l=y.preloadedImage)},[r,i,s,l,a,f,u,d,c,o,h,p]}class Fc extends be{constructor(e){super(),ye(this,e,qc,Rc,ve,{gameData:0,timer:1,question:2,answered:9,preloadedImage:3})}}function zs(t,e,n){const r=t.slice();return r[7]=e[n],r}function Lc(t){let e,n,r;return{c(){e=D("button"),e.textContent="Start",m(e,"class","btn")},m(i,s){W(i,e,s),n||(r=J(e,"click",t[4]),n=!0)},p:Q,d(i){i&&V(e),n=!1,r()}}}function Pc(t){let e,n,r;return{c(){e=D("button"),e.textContent="Next",m(e,"class","btn")},m(i,s){W(i,e,s),n||(r=J(e,"click",t[4]),n=!0)},p:Q,d(i){i&&V(e),n=!1,r()}}}function Zs(t,e){let n,r,i=e[7].name+"",s,o,l,a,f,u,d,c,h,p=Q,y,k,_;a=new hl({props:{value:e[2][e[7].id]??0}});function C(){return e[6](e[7])}return{key:t,first:null,c(){n=D("tr"),r=D("td"),s=te(i),o=U(),l=D("td"),le(a.$$.fragment),f=U(),u=D("td"),d=D("button"),d.textContent="Kick",c=U(),m(r,"class","player__name svelte-t8tz1l"),m(l,"class","player_score"),m(d,"class","btn"),m(u,"class","player__action svelte-t8tz1l"),m(n,"class","player svelte-t8tz1l"),this.first=n},m(T,S){W(T,n,S),g(n,r),g(r,s),g(n,o),g(n,l),se(a,l,null),g(n,f),g(n,u),g(u,d),g(n,c),y=!0,k||(_=J(d,"click",C),k=!0)},p(T,S){e=T,(!y||S&2)&&i!==(i=e[7].name+"")&&Ie(s,i);const $={};S&6&&($.value=e[2][e[7].id]??0),a.$set($)},r(){h=n.getBoundingClientRect()},f(){gr(n),p()},a(){p(),p=mr(n,h,wr,{})},i(T){y||(B(a.$$.fragment,T),y=!0)},o(T){G(a.$$.fragment,T),y=!1},d(T){T&&V(n),oe(a),k=!1,_()}}}function Uc(t){let e,n,r,i=t[0].token+"",s,o,l,a=t[0].config.name+"",f,u,d,c=t[0].config.text+"",h,p,y,k,_,C,T,S,$,w,x=[],v=new Map,b,I,O,R;function N(q,F){if(q[3]===st.Marked)return Pc;if(q[1].length>0&&q[3]===st.Lobby)return Lc}let M=N(t),L=M&&M(t),A=t[1];const E=q=>q[7].id;for(let q=0;q<A.length;q+=1){let F=zs(t,A,q),P=E(F);v.set(P,x[q]=Zs(P,F))}return{c(){e=D("main"),n=D("div"),r=D("h1"),s=te(i),o=U(),l=D("h2"),f=te(a),u=U(),d=D("p"),h=te(c),p=U(),y=D("div"),k=D("button"),k.textContent="Leave",_=U(),L&&L.c(),C=U(),T=D("table"),S=D("thead"),S.innerHTML=`<tr><th>Name</th> 
-          <th>Score</th> 
-          <th>Actions</th></tr>`,$=U(),w=D("tbody");for(let q=0;q<x.length;q+=1)x[q].c();m(r,"class","token svelte-t8tz1l"),m(l,"class","name"),m(d,"class","desc svelte-t8tz1l"),m(k,"class","btn"),m(y,"class","btn-row btn-row--fill actions svelte-t8tz1l"),m(n,"class","quiz svelte-t8tz1l"),m(e,"class","page page--middle page--overflow svelte-t8tz1l")},m(q,F){W(q,e,F),g(e,n),g(n,r),g(r,s),g(n,o),g(n,l),g(l,f),g(n,u),g(n,d),g(d,h),g(n,p),g(n,y),g(y,k),g(y,_),L&&L.m(y,null),g(n,C),g(n,T),g(T,S),g(T,$),g(T,w);for(let P=0;P<x.length;P+=1)x[P]&&x[P].m(w,null);I=!0,O||(R=J(k,"click",t[5]),O=!0)},p(q,[F]){if((!I||F&1)&&i!==(i=q[0].token+"")&&Ie(s,i),(!I||F&1)&&a!==(a=q[0].config.name+"")&&Ie(f,a),(!I||F&1)&&c!==(c=q[0].config.text+"")&&Ie(h,c),M===(M=N(q))&&L?L.p(q,F):(L&&L.d(1),L=M&&M(q),L&&(L.c(),L.m(y,null))),F&6){A=q[1],qe();for(let P=0;P<x.length;P+=1)x[P].r();x=_r(x,F,E,1,q,A,v,w,ii,Zs,null,zs);for(let P=0;P<x.length;P+=1)x[P].a();Fe()}},i(q){if(!I){for(let F=0;F<A.length;F+=1)B(x[F]);Le(()=>{I&&(b||(b=Ae(e,Me,{},!0)),b.run(1))}),I=!0}},o(q){for(let F=0;F<x.length;F+=1)G(x[F]);b||(b=Ae(e,Me,{},!1)),b.run(0),I=!1},d(q){q&&V(e),L&&L.d();for(let F=0;F<x.length;F+=1)x[F].d();q&&b&&b.end(),O=!1,R()}}}function zc(t,e,n){let{gameData:r}=e,{players:i}=e,{scores:s}=e,{gameState:o}=e;const l=()=>dr(Un.Next),a=()=>Xi(r),f=u=>cl(u.id);return t.$$set=u=>{"gameData"in u&&n(0,r=u.gameData),"players"in u&&n(1,i=u.players),"scores"in u&&n(2,s=u.scores),"gameState"in u&&n(3,o=u.gameState)},[r,i,s,o,l,a,f]}class Zc extends be{constructor(e){super(),ye(this,e,zc,Uc,ve,{gameData:0,players:1,scores:2,gameState:3})}}function jc(t){let e,n=t[0].count+"",r,i,s=t[0].total+"",o,l,a,f,u=t[1].toFixed(0)+"",d;return{c(){e=D("p"),r=te(n),i=te(" / "),o=te(s),l=U(),a=D("p"),f=te("+"),d=te(u),m(e,"class","ratio"),m(a,"class","score svelte-1dywswa")},m(c,h){W(c,e,h),g(e,r),g(e,i),g(e,o),W(c,l,h),W(c,a,h),g(a,f),g(a,d)},p(c,h){h&1&&n!==(n=c[0].count+"")&&Ie(r,n),h&1&&s!==(s=c[0].total+"")&&Ie(o,s),h&2&&u!==(u=c[1].toFixed(0)+"")&&Ie(d,u)},d(c){c&&V(e),c&&V(l),c&&V(a)}}}function Hc(t){let e,n,r=t[1].toFixed(0)+"",i;return{c(){e=D("p"),n=te("+"),i=te(r),m(e,"class","score svelte-1dywswa")},m(s,o){W(s,e,o),g(e,n),g(e,i)},p(s,o){o&2&&r!==(r=s[1].toFixed(0)+"")&&Ie(i,r)},d(s){s&&V(e)}}}function Bc(t){let e,n,r=t[0].ty+"",i,s,o,l,a,f,u;function d(p,y){if(p[0].ty===Zt.Correct)return Hc;if(p[0].ty===Zt.Partial)return jc}let c=d(t),h=c&&c(t);return{c(){e=D("main"),n=D("h1"),i=te(r),s=U(),o=D("p"),o.textContent=`${t[2]}`,l=U(),h&&h.c(),m(n,"class","title svelte-1dywswa"),m(o,"class","text svelte-1dywswa"),m(e,"class","main svelte-1dywswa"),m(e,"data-type",a=t[0].ty)},m(p,y){W(p,e,y),g(e,n),g(n,i),g(e,s),g(e,o),g(e,l),h&&h.m(e,null),u=!0},p(p,[y]){(!u||y&1)&&r!==(r=p[0].ty+"")&&Ie(i,r),c===(c=d(p))&&h?h.p(p,y):(h&&h.d(1),h=c&&c(p),h&&(h.c(),h.m(e,null))),(!u||y&1&&a!==(a=p[0].ty))&&m(e,"data-type",a)},i(p){u||(Le(()=>{u&&(f||(f=Ae(e,Me,{},!0)),f.run(1))}),u=!0)},o(p){f||(f=Ae(e,Me,{},!1)),f.run(0),u=!1},d(p){p&&V(e),h&&h.d(),p&&f&&f.end()}}}function Vc(t,e,n){let r,{score:i}=e;const s=dl(i.ty),o=Ki(0,{delay:500});return lt(t,o,l=>n(1,r=l)),t.$$set=l=>{"score"in l&&n(0,i=l.score)},t.$$.update=()=>{t.$$.dirty&1&&(i.ty===Zt.Correct||i.ty===Zt.Partial)&&o.set(i.value)},[i,r,s,o]}class Wc extends be{constructor(e){super(),ye(this,e,Vc,Bc,ve,{score:0})}}function Gc(t){let e,n,r,i,s,o=t[0].token+"",l,a,f,u,d,c,h;return{c(){e=D("main"),n=D("div"),n.innerHTML=`<h1 class="title svelte-1hpr4aa">Waiting for host...</h1> 
-    <p class="text svelte-1hpr4aa">You&#39;re in, waiting for the host to start the quiz.</p>`,r=U(),i=D("div"),s=D("p"),l=te(o),a=U(),f=D("button"),f.textContent="Leave",m(n,"class","quiz svelte-1hpr4aa"),m(s,"class","token svelte-1hpr4aa"),m(f,"class","btn btn--surface"),m(i,"class","bottom svelte-1hpr4aa"),m(e,"class","page page--overflow")},m(p,y){W(p,e,y),g(e,n),g(e,r),g(e,i),g(i,s),g(s,l),g(i,a),g(i,f),d=!0,c||(h=J(f,"click",t[1]),c=!0)},p(p,[y]){(!d||y&1)&&o!==(o=p[0].token+"")&&Ie(l,o)},i(p){d||(Le(()=>{d&&(u||(u=Ae(e,Me,{},!0)),u.run(1))}),d=!0)},o(p){u||(u=Ae(e,Me,{},!1)),u.run(0),d=!1},d(p){p&&V(e),p&&u&&u.end(),c=!1,h()}}}function Qc(t,e,n){let{gameData:r}=e;const i=()=>Xi(r);return t.$$set=s=>{"gameData"in s&&n(0,r=s.gameData)},[r,i]}class Yc extends be{constructor(e){super(),ye(this,e,Qc,Gc,ve,{gameData:0})}}function js(t){let e,n=Br(t[1])+"",r;return{c(){e=D("span"),r=te(n),m(e,"class","time svelte-lus7sg")},m(i,s){W(i,e,s),g(e,r)},p(i,s){s&2&&n!==(n=Br(i[1])+"")&&Ie(r,n)},d(i){i&&V(e)}}}function Xc(t){let e;return{c(){e=D("p"),e.textContent="Get ready to answer",m(e,"class","text svelte-lus7sg")},m(n,r){W(n,e,r)},d(n){n&&V(e)}}}function Kc(t){let e;return{c(){e=D("p"),e.textContent="Get ready, the quiz is about to start.",m(e,"class","text svelte-lus7sg")},m(n,r){W(n,e,r)},d(n){n&&V(e)}}}function Hs(t){let e,n,r,i,s=t[0]===st.Starting&&Bs(t);return{c(){s&&s.c(),e=U(),n=D("button"),n.textContent="Skip",m(n,"class","btn btn--surface")},m(o,l){s&&s.m(o,l),W(o,e,l),W(o,n,l),r||(i=J(n,"click",t[3]),r=!0)},p(o,l){o[0]===st.Starting?s?s.p(o,l):(s=Bs(o),s.c(),s.m(e.parentNode,e)):s&&(s.d(1),s=null)},d(o){s&&s.d(o),o&&V(e),o&&V(n),r=!1,i()}}}function Bs(t){let e,n,r;return{c(){e=D("button"),e.textContent="Cancel",m(e,"class","btn btn--surface")},m(i,s){W(i,e,s),n||(r=J(e,"click",t[4]),n=!0)},p:Q,d(i){i&&V(e),n=!1,r()}}}function Jc(t){let e,n,r,i,s,o,l=t[2].token+"",a,f,u,d,c=t[1].elapsed!==t[1].total&&js(t);function h(_,C){return _[0]===st.Starting?Kc:Xc}let p=h(t),y=p(t),k=t[2].host&&Hs(t);return{c(){e=D("main"),n=D("div"),c&&c.c(),r=U(),y.c(),i=U(),s=D("div"),o=D("p"),a=te(l),f=U(),k&&k.c(),m(n,"class","quiz svelte-lus7sg"),m(o,"class","token svelte-lus7sg"),m(s,"class","bottom svelte-lus7sg"),m(e,"class","page page--overflow")},m(_,C){W(_,e,C),g(e,n),c&&c.m(n,null),g(n,r),y.m(n,null),g(e,i),g(e,s),g(s,o),g(o,a),g(s,f),k&&k.m(s,null),d=!0},p(_,[C]){_[1].elapsed!==_[1].total?c?c.p(_,C):(c=js(_),c.c(),c.m(n,r)):c&&(c.d(1),c=null),p!==(p=h(_))&&(y.d(1),y=p(_),y&&(y.c(),y.m(n,null))),(!d||C&4)&&l!==(l=_[2].token+"")&&Ie(a,l),_[2].host?k?k.p(_,C):(k=Hs(_),k.c(),k.m(s,null)):k&&(k.d(1),k=null)},i(_){d||(Le(()=>{d&&(u||(u=Ae(e,Me,{},!0)),u.run(1))}),d=!0)},o(_){u||(u=Ae(e,Me,{},!1)),u.run(0),d=!1},d(_){_&&V(e),c&&c.d(),y.d(),k&&k.d(),_&&u&&u.end()}}}function ef(t,e,n){let{gameState:r}=e,{timer:i}=e,{gameData:s}=e;const o=()=>dr(Un.Next),l=()=>dr(Un.Reset);return t.$$set=a=>{"gameState"in a&&n(0,r=a.gameState),"timer"in a&&n(1,i=a.timer),"gameData"in a&&n(2,s=a.gameData)},[r,i,s,o,l]}class tf extends be{constructor(e){super(),ye(this,e,ef,Jc,ve,{gameState:0,timer:1,gameData:2})}}function nf(t){let e,n;return e=new il({props:{text:"..."}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p:Q,i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function rf(t){let e,n;return e=new Wc({props:{score:t[6]}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p(r,i){const s={};i&64&&(s.score=r[6]),e.$set(s)},i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function sf(t){let e,n;return e=new Yc({props:{gameData:t[0]}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p(r,i){const s={};i&1&&(s.gameData=r[0]),e.$set(s)},i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function of(t){let e,n;return e=new Zc({props:{gameData:t[0],gameState:t[3],players:t[2],scores:t[7]}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p(r,i){const s={};i&1&&(s.gameData=r[0]),i&8&&(s.gameState=r[3]),i&4&&(s.players=r[2]),i&128&&(s.scores=r[7]),e.$set(s)},i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function lf(t){let e,n,r,i;const s=[ff,cf],o=[];function l(a,f){return a[8]?a[2].length!==1?1:-1:0}return~(e=l(t))&&(n=o[e]=s[e](t)),{c(){n&&n.c(),r=Lt()},m(a,f){~e&&o[e].m(a,f),W(a,r,f),i=!0},p(a,f){let u=e;e=l(a),e===u?~e&&o[e].p(a,f):(n&&(qe(),G(o[u],1,1,()=>{o[u]=null}),Fe()),~e?(n=o[e],n?n.p(a,f):(n=o[e]=s[e](a),n.c()),B(n,1),n.m(r.parentNode,r)):n=null)},i(a){i||(B(n),i=!0)},o(a){G(n),i=!1},d(a){~e&&o[e].d(a),a&&V(r)}}}function af(t){let e,n;return e=new tf({props:{gameState:t[3],gameData:t[0],timer:t[9]}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p(r,i){const s={};i&8&&(s.gameState=r[3]),i&1&&(s.gameData=r[0]),i&512&&(s.timer=r[9]),e.$set(s)},i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function uf(t){let e,n;return e=new $c({props:{gameData:t[0],summary:t[4]}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p(r,i){const s={};i&1&&(s.gameData=r[0]),i&16&&(s.summary=r[4]),e.$set(s)},i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function cf(t){let e,n;return e=new Ac({}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p:Q,i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function ff(t){let e,n,r;function i(o){t[10](o)}let s={gameData:t[0],question:t[5],timer:t[9],preloadedImage:t[1]};return t[8]!==void 0&&(s.answered=t[8]),e=new Fc({props:s}),Ke.push(()=>ut(e,"answered",i)),{c(){le(e.$$.fragment)},m(o,l){se(e,o,l),r=!0},p(o,l){const a={};l&1&&(a.gameData=o[0]),l&32&&(a.question=o[5]),l&512&&(a.timer=o[9]),l&2&&(a.preloadedImage=o[1]),!n&&l&256&&(n=!0,a.answered=o[8],at(()=>n=!1)),e.$set(a)},i(o){r||(B(e.$$.fragment,o),r=!0)},o(o){G(e.$$.fragment,o),r=!1},d(o){oe(e,o)}}}function df(t){let e,n,r,i;const s=[uf,af,lf,of,sf,rf,nf],o=[];function l(a,f){return a[3]===st.Finished&&a[4]!=null?0:a[3]===st.Starting||a[3]===st.PreQuestion||a[3]===st.AwaitingReady?1:a[3]===st.AwaitingAnswers&&a[5]!=null?2:a[0].host?3:a[3]===st.Lobby?4:a[3]===st.Marked?5:6}return e=l(t),n=o[e]=s[e](t),{c(){n.c(),r=Lt()},m(a,f){o[e].m(a,f),W(a,r,f),i=!0},p(a,[f]){let u=e;e=l(a),e===u?o[e].p(a,f):(qe(),G(o[u],1,1,()=>{o[u]=null}),Fe(),n=o[e],n?n.p(a,f):(n=o[e]=s[e](a),n.c()),B(n,1),n.m(r.parentNode,r))},i(a){i||(B(n),i=!0)},o(a){G(n),i=!1},d(a){o[e].d(a),a&&V(r)}}}function hf(t,e,n){let{gameData:r}=e,i=null,s=[],o=st.Lobby,l=null;r.host||s.push({id:r.id,name:r.name??""});let a=null,f={ty:Zt.Incorrect},u={},d,c={total:0,elapsed:0},h=0;function p(){if(c.elapsed===c.total)return;const k=performance.now(),_=k-h;n(9,c.elapsed+=_,c),h=k,c.elapsed>c.total?n(9,c.elapsed=c.total,c):requestAnimationFrame(p)}Gt(Et.PlayerData,k=>{console.debug("Other player message",k),s.push(k),n(2,s)}),Gt(Et.GameState,k=>{if(console.debug("Game state message",k),n(3,o=k.state),n(8,d=!1),k.state===st.Lobby)n(7,u={});else if(k.state===st.Finished){const _=s.map(C=>({score:u[C.id]??0,...C}));_.sort((C,T)=>T.score-C.score),n(4,l={players:_})}}),Gt(Et.Timer,k=>{console.debug("Time sync message",k),h=performance.now(),n(9,c.total=k.value,c),n(9,c.elapsed=0,c),p()}),Gt(Et.Question,async k=>{console.debug("Question message",k),n(5,a=k.question),n(1,i=await Ou(r.token,a)),i!==null&&a.image!==null&&(i.classList.add("question-image"),i.setAttribute("data-fit",a.image.fit),n(1,i.alt=a.text,i),i.decode!==void 0&&await i.decode()),await yc(),console.debug("Server acknowledged ready state")}),Gt(Et.Scores,k=>{console.debug("Score message",k),n(7,u=k.scores);let _=C=>u[C]??0;n(2,s=s.sort((C,T)=>_(T.id)-_(C.id)))}),Gt(Et.Score,k=>{console.debug("Score message",k),n(6,f=k.score)}),Gt(Et.Error,k=>{console.error("Server error",k.error)}),Gt(Et.Kicked,k=>{if(console.debug("Kick message",k),n(2,s=s.filter(_=>_.id!==k.id)),k.id===r.id&&(yr(),k.reason!==Jo.Disconnected)){const _=xu[k.reason];Qt("Removed from game",_)}}),Uo(()=>{});function y(k){d=k,n(8,d)}return t.$$set=k=>{"gameData"in k&&n(0,r=k.gameData)},[r,i,s,o,l,a,f,u,d,c,y]}class pf extends be{constructor(e){super(),ye(this,e,hf,df,ve,{gameData:0})}}const mf="finalize",gf="consider";function zn(t,e,n){t.dispatchEvent(new CustomEvent(mf,{detail:{items:e,info:n}}))}function rn(t,e,n){t.dispatchEvent(new CustomEvent(gf,{detail:{items:e,info:n}}))}const ai="draggedEntered",kr="draggedLeft",ui="draggedOverIndex",Ji="draggedLeftDocument",Vr={LEFT_FOR_ANOTHER:"leftForAnother",OUTSIDE_OF_ANY:"outsideOfAny"};function _f(t,e,n){t.dispatchEvent(new CustomEvent(ai,{detail:{indexObj:e,draggedEl:n}}))}function vf(t,e,n){t.dispatchEvent(new CustomEvent(kr,{detail:{draggedEl:e,type:Vr.LEFT_FOR_ANOTHER,theOtherDz:n}}))}function yf(t,e){t.dispatchEvent(new CustomEvent(kr,{detail:{draggedEl:e,type:Vr.OUTSIDE_OF_ANY}}))}function bf(t,e,n){t.dispatchEvent(new CustomEvent(ui,{detail:{indexObj:e,draggedEl:n}}))}function wf(t){window.dispatchEvent(new CustomEvent(Ji,{detail:{draggedEl:t}}))}const ht={DRAG_STARTED:"dragStarted",DRAGGED_ENTERED:ai,DRAGGED_ENTERED_ANOTHER:"dragEnteredAnother",DRAGGED_OVER_INDEX:ui,DRAGGED_LEFT:kr,DRAGGED_LEFT_ALL:"draggedLeftAll",DROPPED_INTO_ZONE:"droppedIntoZone",DROPPED_INTO_ANOTHER:"droppedIntoAnother",DROPPED_OUTSIDE_OF_ANY:"droppedOutsideOfAny",DRAG_STOPPED:"dragStopped"},gt={POINTER:"pointer",KEYBOARD:"keyboard"},Zn="isDndShadowItem",es="data-is-dnd-shadow-item",ts="id:dnd-shadow-placeholder-0000",kf="dnd-action-dragged-el";let Xe="id",qi=0;function pl(){qi++}function ml(){if(qi===0)throw new Error("Bug! trying to decrement when there are no dropzones");qi--}const ns=typeof window>"u";function gl(t){let e;const n=t.getBoundingClientRect(),r=getComputedStyle(t),i=r.transform;if(i){let s,o,l,a;if(i.startsWith("matrix3d("))e=i.slice(9,-1).split(/, /),s=+e[0],o=+e[5],l=+e[12],a=+e[13];else if(i.startsWith("matrix("))e=i.slice(7,-1).split(/, /),s=+e[0],o=+e[3],l=+e[4],a=+e[5];else return n;const f=r.transformOrigin,u=n.x-l-(1-s)*parseFloat(f),d=n.y-a-(1-o)*parseFloat(f.slice(f.indexOf(" ")+1)),c=s?n.width/s:t.offsetWidth,h=o?n.height/o:t.offsetHeight;return{x:u,y:d,width:c,height:h,top:d,right:u+c,bottom:d+h,left:u}}else return n}function _l(t){const e=gl(t);return{top:e.top+window.scrollY,bottom:e.bottom+window.scrollY,left:e.left+window.scrollX,right:e.right+window.scrollX}}function rs(t){const e=t.getBoundingClientRect();return{top:e.top+window.scrollY,bottom:e.bottom+window.scrollY,left:e.left+window.scrollX,right:e.right+window.scrollX}}function vl(t){return{x:(t.left+t.right)/2,y:(t.top+t.bottom)/2}}function Ef(t,e){return Math.sqrt(Math.pow(t.x-e.x,2)+Math.pow(t.y-e.y,2))}function is(t,e){return t.y<=e.bottom&&t.y>=e.top&&t.x>=e.left&&t.x<=e.right}function hr(t){return vl(rs(t))}function Vs(t,e){const n=hr(t),r=_l(e);return is(n,r)}function Af(t,e){const n=hr(t),r=hr(e);return Ef(n,r)}function Tf(t){const e=rs(t);return e.right<0||e.left>document.documentElement.scrollWidth||e.bottom<0||e.top>document.documentElement.scrollHeight}function Cf(t,e){const n=rs(e);return is(t,n)?{top:t.y-n.top,bottom:n.bottom-t.y,left:t.x-n.left,right:Math.min(n.right,document.documentElement.clientWidth)-t.x}:null}let gn;function ss(){gn=new Map}ss();function xf(t){gn.delete(t)}function Df(t){const e=Array.from(t.children).findIndex(n=>n.getAttribute(es));if(e>=0)return gn.has(t)||gn.set(t,new Map),gn.get(t).set(e,_l(t.children[e])),e}function Sf(t,e){if(!Vs(t,e))return null;const n=e.children;if(n.length===0)return{index:0,isProximityBased:!0};const r=Df(e);for(let o=0;o<n.length;o++)if(Vs(t,n[o])){const l=gn.has(e)&&gn.get(e).get(o);return l&&!is(hr(t),l)?{index:r,isProximityBased:!1}:{index:o,isProximityBased:!1}}let i=Number.MAX_VALUE,s;for(let o=0;o<n.length;o++){const l=Af(t,n[o]);l<i&&(i=l,s=o)}return{index:s,isProximityBased:!0}}const Qn=25;function yl(){let t;function e(){t={directionObj:void 0,stepPx:0}}e();function n(s){const{directionObj:o,stepPx:l}=t;o&&(s.scrollBy(o.x*l,o.y*l),window.requestAnimationFrame(()=>n(s)))}function r(s){return Qn-s}function i(s,o){if(!o)return!1;const l=Cf(s,o);if(l===null)return e(),!1;const a=!!t.directionObj;let[f,u]=[!1,!1];return o.scrollHeight>o.clientHeight&&(l.bottom<Qn?(f=!0,t.directionObj={x:0,y:1},t.stepPx=r(l.bottom)):l.top<Qn&&(f=!0,t.directionObj={x:0,y:-1},t.stepPx=r(l.top)),!a&&f)||o.scrollWidth>o.clientWidth&&(l.right<Qn?(u=!0,t.directionObj={x:1,y:0},t.stepPx=r(l.right)):l.left<Qn&&(u=!0,t.directionObj={x:-1,y:0},t.stepPx=r(l.left)),!a&&u)?(n(o),!0):(e(),!1)}return{scrollIfNeeded:i,resetScrolling:e}}function bi(t){return JSON.stringify(t,null,2)}function Ws(t){if(!t)throw new Error("cannot get depth of a falsy node");return bl(t,0)}function bl(t,e=0){return t.parentElement?bl(t.parentElement,e+1):e-1}function If(t,e){if(Object.keys(t).length!==Object.keys(e).length)return!1;for(const n in t)if(!{}.hasOwnProperty.call(e,n)||e[n]!==t[n])return!1;return!0}function $f(t,e){if(t.length!==e.length)return!1;for(let n=0;n<t.length;n++)if(t[n]!==e[n])return!1;return!0}const Of=200,Gs=10,{scrollIfNeeded:Mf,resetScrolling:Rf}=yl();let Fi;function Nf(t,e,n=Of){let r,i,s=!1,o;const l=Array.from(e).sort((f,u)=>Ws(u)-Ws(f));function a(){const f=hr(t),u=Mf(f,r);if(!u&&o&&Math.abs(o.x-f.x)<Gs&&Math.abs(o.y-f.y)<Gs){Fi=window.setTimeout(a,n);return}if(Tf(t)){wf(t);return}o=f;let d=!1;for(const c of l){u&&xf(r);const h=Sf(t,c);if(h===null)continue;const{index:p}=h;d=!0,c!==r?(r&&vf(r,t,c),_f(c,h,t),r=c):p!==i&&(bf(c,h,t),i=p);break}!d&&s&&r?(yf(r,t),r=void 0,i=void 0,s=!1):s=!0,Fi=window.setTimeout(a,n)}a()}function qf(){clearTimeout(Fi),Rf(),ss()}const Ff=300;let Wr;function Gr(t){const e=t.touches?t.touches[0]:t;Wr={x:e.clientX,y:e.clientY}}const{scrollIfNeeded:Lf,resetScrolling:Pf}=yl();let wl;function kl(){Wr&&Lf(Wr,document.documentElement)&&ss(),wl=window.setTimeout(kl,Ff)}function Uf(){window.addEventListener("mousemove",Gr),window.addEventListener("touchmove",Gr),kl()}function zf(){window.removeEventListener("mousemove",Gr),window.removeEventListener("touchmove",Gr),Wr=void 0,window.clearTimeout(wl),Pf()}function Zf(t){const e=t.cloneNode(!0),n=[],r=t.tagName==="SELECT",i=r?[t]:[...t.querySelectorAll("select")];for(const o of i)n.push(o.value);if(i.length<=0)return e;const s=r?[e]:[...e.querySelectorAll("select")];for(let o=0;o<s.length;o++){const l=s[o],a=n[o],f=l.querySelector(`option[value="${a}"`);f&&f.setAttribute("selected",!0)}return e}const jf=.2;function cn(t){return`${t} ${jf}s ease`}function Hf(t,e){const n=t.getBoundingClientRect(),r=Zf(t);El(t,r),r.id=kf,r.style.position="fixed";let i=n.top,s=n.left;if(r.style.top=`${i}px`,r.style.left=`${s}px`,e){const o=vl(n);i-=o.y-e.y,s-=o.x-e.x,window.setTimeout(()=>{r.style.top=`${i}px`,r.style.left=`${s}px`},0)}return r.style.margin="0",r.style.boxSizing="border-box",r.style.height=`${n.height}px`,r.style.width=`${n.width}px`,r.style.transition=`${cn("top")}, ${cn("left")}, ${cn("background-color")}, ${cn("opacity")}, ${cn("color")} `,window.setTimeout(()=>r.style.transition+=`, ${cn("width")}, ${cn("height")}`,0),r.style.zIndex="9999",r.style.cursor="grabbing",r}function Bf(t){t.style.cursor="grab"}function Vf(t,e,n,r){El(e,t);const i=e.getBoundingClientRect(),s=t.getBoundingClientRect(),o=i.width-s.width,l=i.height-s.height;if(o||l){const a={left:(n-s.left)/s.width,top:(r-s.top)/s.height};t.style.height=`${i.height}px`,t.style.width=`${i.width}px`,t.style.left=`${parseFloat(t.style.left)-a.left*o}px`,t.style.top=`${parseFloat(t.style.top)-a.top*l}px`}}function El(t,e){const n=window.getComputedStyle(t);Array.from(n).filter(r=>r.startsWith("background")||r.startsWith("padding")||r.startsWith("font")||r.startsWith("text")||r.startsWith("align")||r.startsWith("justify")||r.startsWith("display")||r.startsWith("flex")||r.startsWith("border")||r==="opacity"||r==="color"||r==="list-style-type").forEach(r=>e.style.setProperty(r,n.getPropertyValue(r),n.getPropertyPriority(r)))}function Wf(t,e){t.draggable=!1,t.ondragstart=()=>!1,e?(t.style.userSelect="",t.style.WebkitUserSelect="",t.style.cursor=""):(t.style.userSelect="none",t.style.WebkitUserSelect="none",t.style.cursor="grab")}function Al(t){t.style.display="none",t.style.position="fixed",t.style.zIndex="-5"}function Gf(t){t.style.visibility="hidden",t.setAttribute(es,"true")}function Qf(t){t.style.visibility="",t.removeAttribute(es)}function Sr(t,e=()=>{},n=()=>[]){t.forEach(r=>{const i=e(r);Object.keys(i).forEach(s=>{r.style[s]=i[s]}),n(r).forEach(s=>r.classList.add(s))})}function Qr(t,e=()=>{},n=()=>[]){t.forEach(r=>{const i=e(r);Object.keys(i).forEach(s=>{r.style[s]=""}),n(r).forEach(s=>r.classList.contains(s)&&r.classList.remove(s))})}function Yf(t){const e=t.style.minHeight;t.style.minHeight=window.getComputedStyle(t).getPropertyValue("height");const n=t.style.minWidth;return t.style.minWidth=window.getComputedStyle(t).getPropertyValue("width"),function(){t.style.minHeight=e,t.style.minWidth=n}}const Xf="--any--",Kf=100,Qs=3,Ys={outline:"rgba(255, 255, 102, 0.7) solid 2px"};let Pt,nt,vt,ci,We,fi,jn,_t,Xt,bt,dn=!1,os=!1,ls,Er=!1,Ir=[];const Ct=new Map,Ve=new Map,wi=new WeakMap;function Jf(t,e){Ct.has(e)||Ct.set(e,new Set),Ct.get(e).has(t)||(Ct.get(e).add(t),pl())}function Xs(t,e){Ct.get(e).delete(t),ml(),Ct.get(e).size===0&&Ct.delete(e)}function ed(){Uf();const t=Ct.get(ci);for(const n of t)n.addEventListener(ai,Tl),n.addEventListener(kr,Cl),n.addEventListener(ui,xl);window.addEventListener(Ji,Hn);const e=Math.max(Kf,...Array.from(t.keys()).map(n=>Ve.get(n).dropAnimationDurationMs));Nf(nt,t,e*1.07)}function td(){zf();const t=Ct.get(ci);for(const e of t)e.removeEventListener(ai,Tl),e.removeEventListener(kr,Cl),e.removeEventListener(ui,xl);window.removeEventListener(Ji,Hn),qf()}function nd(t){return t.findIndex(e=>e[Xe]===ts)}function di(t){return t.findIndex(e=>!!e[Zn]&&e[Xe]!==ts)}function Tl(t){let{items:e,dropFromOthersDisabled:n}=Ve.get(t.currentTarget);if(n&&t.currentTarget!==We)return;if(Er=!1,e=e.filter(o=>o[Xe]!==jn[Xe]),We!==t.currentTarget){const l=Ve.get(We).items.filter(a=>!a[Zn]);rn(We,l,{trigger:ht.DRAGGED_ENTERED_ANOTHER,id:vt[Xe],source:gt.POINTER})}else{const o=nd(e);o!==-1&&e.splice(o,1)}const{index:r,isProximityBased:i}=t.detail.indexObj,s=i&&r===t.currentTarget.children.length-1?r+1:r;_t=t.currentTarget,e.splice(s,0,jn),rn(t.currentTarget,e,{trigger:ht.DRAGGED_ENTERED,id:vt[Xe],source:gt.POINTER})}function Cl(t){if(!dn)return;const{items:e,dropFromOthersDisabled:n}=Ve.get(t.currentTarget);if(n&&t.currentTarget!==We&&t.currentTarget!==_t)return;const r=di(e),i=e.splice(r,1)[0];_t=void 0;const{type:s,theOtherDz:o}=t.detail;if(s===Vr.OUTSIDE_OF_ANY||s===Vr.LEFT_FOR_ANOTHER&&o!==We&&Ve.get(o).dropFromOthersDisabled){Er=!0,_t=We;const l=Ve.get(We).items;l.splice(fi,0,i),rn(We,l,{trigger:ht.DRAGGED_LEFT_ALL,id:vt[Xe],source:gt.POINTER})}rn(t.currentTarget,e,{trigger:ht.DRAGGED_LEFT,id:vt[Xe],source:gt.POINTER})}function xl(t){const{items:e,dropFromOthersDisabled:n}=Ve.get(t.currentTarget);if(n&&t.currentTarget!==We)return;Er=!1;const{index:r}=t.detail.indexObj,i=di(e);e.splice(i,1),e.splice(r,0,jn),rn(t.currentTarget,e,{trigger:ht.DRAGGED_OVER_INDEX,id:vt[Xe],source:gt.POINTER})}function Yr(t){t.preventDefault();const e=t.touches?t.touches[0]:t;bt={x:e.clientX,y:e.clientY},nt.style.transform=`translate3d(${bt.x-Xt.x}px, ${bt.y-Xt.y}px, 0)`}function Hn(){os=!0,window.removeEventListener("mousemove",Yr),window.removeEventListener("touchmove",Yr),window.removeEventListener("mouseup",Hn),window.removeEventListener("touchend",Hn),td(),Bf(nt),_t||(_t=We);let{items:t,type:e}=Ve.get(_t);Qr(Ct.get(e),i=>Ve.get(i).dropTargetStyle,i=>Ve.get(i).dropTargetClasses);let n=di(t);n===-1&&(n=fi),t=t.map(i=>i[Zn]?vt:i);function r(){ls(),zn(_t,t,{trigger:Er?ht.DROPPED_OUTSIDE_OF_ANY:ht.DROPPED_INTO_ZONE,id:vt[Xe],source:gt.POINTER}),_t!==We&&zn(We,Ve.get(We).items,{trigger:ht.DROPPED_INTO_ANOTHER,id:vt[Xe],source:gt.POINTER}),Qf(_t.children[n]),sd()}rd(n,r)}function rd(t,e){const n=gl(_t.children[t]),r={x:n.left-parseFloat(nt.style.left),y:n.top-parseFloat(nt.style.top)},{dropAnimationDurationMs:i}=Ve.get(_t),s=`transform ${i}ms ease`;nt.style.transition=nt.style.transition?nt.style.transition+","+s:s,nt.style.transform=`translate3d(${r.x}px, ${r.y}px, 0)`,window.setTimeout(e,i)}function id(t,e){Ir.push({dz:t,destroy:e}),window.requestAnimationFrame(()=>{Al(t),document.body.appendChild(t)})}function sd(){nt.remove(),Pt.remove(),Ir.length&&(Ir.forEach(({dz:t,destroy:e})=>{e(),t.remove()}),Ir=[]),nt=void 0,Pt=void 0,vt=void 0,ci=void 0,We=void 0,fi=void 0,jn=void 0,_t=void 0,Xt=void 0,bt=void 0,dn=!1,os=!1,ls=void 0,Er=!1}function od(t,e){let n=!1;const r={items:void 0,type:void 0,flipDurationMs:0,dragDisabled:!1,morphDisabled:!1,dropFromOthersDisabled:!1,dropTargetStyle:Ys,dropTargetClasses:[],transformDraggedElement:()=>{},centreDraggedOnCursor:!1};let i=new Map;function s(){window.addEventListener("mousemove",a,{passive:!1}),window.addEventListener("touchmove",a,{passive:!1,capture:!1}),window.addEventListener("mouseup",l,{passive:!1}),window.addEventListener("touchend",l,{passive:!1})}function o(){window.removeEventListener("mousemove",a),window.removeEventListener("touchmove",a),window.removeEventListener("mouseup",l),window.removeEventListener("touchend",l)}function l(){o(),Pt=void 0,Xt=void 0,bt=void 0}function a(c){c.preventDefault();const h=c.touches?c.touches[0]:c;bt={x:h.clientX,y:h.clientY},(Math.abs(bt.x-Xt.x)>=Qs||Math.abs(bt.y-Xt.y)>=Qs)&&(o(),u())}function f(c){if(c.target!==c.currentTarget&&(c.target.value!==void 0||c.target.isContentEditable)||c.button||dn)return;c.stopPropagation();const h=c.touches?c.touches[0]:c;Xt={x:h.clientX,y:h.clientY},bt={...Xt},Pt=c.currentTarget,s()}function u(){dn=!0;const c=i.get(Pt);fi=c,We=Pt.parentElement;const h=We.getRootNode(),p=h.body||h,{items:y,type:k,centreDraggedOnCursor:_}=r;vt={...y[c]},ci=k,jn={...vt,[Zn]:!0};const C={...jn,[Xe]:ts};nt=Hf(Pt,_&&bt);function T(){nt.parentElement?window.requestAnimationFrame(T):(p.appendChild(nt),nt.focus(),ed(),Al(Pt),p.appendChild(Pt))}window.requestAnimationFrame(T),Sr(Array.from(Ct.get(r.type)).filter(S=>S===We||!Ve.get(S).dropFromOthersDisabled),S=>Ve.get(S).dropTargetStyle,S=>Ve.get(S).dropTargetClasses),y.splice(c,1,C),ls=Yf(We),rn(We,y,{trigger:ht.DRAG_STARTED,id:vt[Xe],source:gt.POINTER}),window.addEventListener("mousemove",Yr,{passive:!1}),window.addEventListener("touchmove",Yr,{passive:!1,capture:!1}),window.addEventListener("mouseup",Hn,{passive:!1}),window.addEventListener("touchend",Hn,{passive:!1})}function d({items:c=void 0,flipDurationMs:h=0,type:p=Xf,dragDisabled:y=!1,morphDisabled:k=!1,dropFromOthersDisabled:_=!1,dropTargetStyle:C=Ys,dropTargetClasses:T=[],transformDraggedElement:S=()=>{},centreDraggedOnCursor:$=!1}){r.dropAnimationDurationMs=h,r.type&&p!==r.type&&Xs(t,r.type),r.type=p,Jf(t,p),r.items=[...c],r.dragDisabled=y,r.morphDisabled=k,r.transformDraggedElement=S,r.centreDraggedOnCursor=$,n&&dn&&!os&&(!If(C,r.dropTargetStyle)||!$f(T,r.dropTargetClasses))&&(Qr([t],()=>r.dropTargetStyle,()=>T),Sr([t],()=>C,()=>T)),r.dropTargetStyle=C,r.dropTargetClasses=[...T];function w(v,b){return Ve.get(v)?Ve.get(v)[b]:r[b]}n&&dn&&r.dropFromOthersDisabled!==_&&(_?Qr([t],v=>w(v,"dropTargetStyle"),v=>w(v,"dropTargetClasses")):Sr([t],v=>w(v,"dropTargetStyle"),v=>w(v,"dropTargetClasses"))),r.dropFromOthersDisabled=_,Ve.set(t,r);const x=di(r.items);for(let v=0;v<t.children.length;v++){const b=t.children[v];if(Wf(b,y),v===x){r.transformDraggedElement(nt,vt,v),k||Vf(nt,b,bt.x,bt.y),Gf(b);continue}b.removeEventListener("mousedown",wi.get(b)),b.removeEventListener("touchstart",wi.get(b)),y||(b.addEventListener("mousedown",f),b.addEventListener("touchstart",f),wi.set(b,f)),i.set(b,v),n||(n=!0)}}return d(e),{update:c=>{d(c)},destroy:()=>{function c(){Xs(t,Ve.get(t).type),Ve.delete(t)}dn?id(t,c):c()}}}const Li={DND_ZONE_ACTIVE:"dnd-zone-active",DND_ZONE_DRAG_DISABLED:"dnd-zone-drag-disabled"},Dl={[Li.DND_ZONE_ACTIVE]:"Tab to one the items and press space-bar or enter to start dragging it",[Li.DND_ZONE_DRAG_DISABLED]:"This is a disabled drag and drop list"},ld="dnd-action-aria-alert";let Be;function Pi(){Be||(Be=document.createElement("div"),function(){Be.id=ld,Be.style.position="fixed",Be.style.bottom="0",Be.style.left="0",Be.style.zIndex="-5",Be.style.opacity="0",Be.style.height="0",Be.style.width="0",Be.setAttribute("role","alert")}(),document.body.prepend(Be),Object.entries(Dl).forEach(([t,e])=>document.body.prepend(cd(t,e))))}function ad(){return ns?null:(document.readyState==="complete"?Pi():window.addEventListener("DOMContentLoaded",Pi),{...Li})}function ud(){ns||!Be||(Object.keys(Dl).forEach(t=>{var e;return(e=document.getElementById(t))==null?void 0:e.remove()}),Be.remove(),Be=void 0)}function cd(t,e){const n=document.createElement("div");return n.id=t,n.innerHTML=`<p>${e}</p>`,n.style.display="none",n.style.position="fixed",n.style.zIndex="-5",n}function Mn(t){if(ns)return;Be||Pi(),Be.innerHTML="";const e=document.createTextNode(t);Be.appendChild(e),Be.style.display="none",Be.style.display="inline"}const fd="--any--",Ks={outline:"rgba(255, 255, 102, 0.7) solid 2px"};let wt=!1,Ui,ot,_n="",fn,Mt,Kt="";const Xr=new WeakSet,Js=new WeakMap,eo=new WeakMap,zi=new Map,tt=new Map,At=new Map;let Kr;function dd(t,e){At.size===0&&(Kr=ad(),window.addEventListener("keydown",Sl),window.addEventListener("click",Il)),At.has(e)||At.set(e,new Set),At.get(e).has(t)||(At.get(e).add(t),pl())}function to(t,e){ot===t&&pr(),At.get(e).delete(t),ml(),At.get(e).size===0&&At.delete(e),At.size===0&&(window.removeEventListener("keydown",Sl),window.removeEventListener("click",Il),Kr=void 0,ud())}function Sl(t){if(wt)switch(t.key){case"Escape":{pr();break}}}function Il(){wt&&(Xr.has(document.activeElement)||pr())}function hd(t){if(!wt)return;const e=t.currentTarget;if(e===ot)return;_n=e.getAttribute("aria-label")||"";const{items:n}=tt.get(ot),r=n.find(f=>f[Xe]===Mt),i=n.indexOf(r),s=n.splice(i,1)[0],{items:o,autoAriaDisabled:l}=tt.get(e);e.getBoundingClientRect().top<ot.getBoundingClientRect().top||e.getBoundingClientRect().left<ot.getBoundingClientRect().left?(o.push(s),l||Mn(`Moved item ${Kt} to the end of the list ${_n}`)):(o.unshift(s),l||Mn(`Moved item ${Kt} to the beginning of the list ${_n}`)),zn(ot,n,{trigger:ht.DROPPED_INTO_ANOTHER,id:Mt,source:gt.KEYBOARD}),zn(e,o,{trigger:ht.DROPPED_INTO_ZONE,id:Mt,source:gt.KEYBOARD}),ot=e}function $l(){zi.forEach(({update:t},e)=>t(tt.get(e)))}function pr(t=!0){tt.get(ot).autoAriaDisabled||Mn(`Stopped dragging item ${Kt}`),Xr.has(document.activeElement)&&document.activeElement.blur(),t&&rn(ot,tt.get(ot).items,{trigger:ht.DRAG_STOPPED,id:Mt,source:gt.KEYBOARD}),Qr(At.get(Ui),e=>tt.get(e).dropTargetStyle,e=>tt.get(e).dropTargetClasses),fn=null,Mt=null,Kt="",Ui=null,ot=null,_n="",wt=!1,$l()}function pd(t,e){const n={items:void 0,type:void 0,dragDisabled:!1,zoneTabIndex:0,dropFromOthersDisabled:!1,dropTargetStyle:Ks,dropTargetClasses:[],autoAriaDisabled:!1};function r(u,d,c){u.length<=1||u.splice(c,1,u.splice(d,1,u[c])[0])}function i(u){switch(u.key){case"Enter":case" ":{if((u.target.disabled!==void 0||u.target.href||u.target.isContentEditable)&&!Xr.has(u.target))return;u.preventDefault(),u.stopPropagation(),wt?pr():s(u);break}case"ArrowDown":case"ArrowRight":{if(!wt)return;u.preventDefault(),u.stopPropagation();const{items:d}=tt.get(t),c=Array.from(t.children),h=c.indexOf(u.currentTarget);h<c.length-1&&(n.autoAriaDisabled||Mn(`Moved item ${Kt} to position ${h+2} in the list ${_n}`),r(d,h,h+1),zn(t,d,{trigger:ht.DROPPED_INTO_ZONE,id:Mt,source:gt.KEYBOARD}));break}case"ArrowUp":case"ArrowLeft":{if(!wt)return;u.preventDefault(),u.stopPropagation();const{items:d}=tt.get(t),h=Array.from(t.children).indexOf(u.currentTarget);h>0&&(n.autoAriaDisabled||Mn(`Moved item ${Kt} to position ${h} in the list ${_n}`),r(d,h,h-1),zn(t,d,{trigger:ht.DROPPED_INTO_ZONE,id:Mt,source:gt.KEYBOARD}));break}}}function s(u){l(u.currentTarget),ot=t,Ui=n.type,wt=!0;const d=Array.from(At.get(n.type)).filter(c=>c===ot||!tt.get(c).dropFromOthersDisabled);if(Sr(d,c=>tt.get(c).dropTargetStyle,c=>tt.get(c).dropTargetClasses),!n.autoAriaDisabled){let c=`Started dragging item ${Kt}. Use the arrow keys to move it within its list ${_n}`;d.length>1&&(c+=", or tab to another list in order to move the item into it"),Mn(c)}rn(t,tt.get(t).items,{trigger:ht.DRAG_STARTED,id:Mt,source:gt.KEYBOARD}),$l()}function o(u){wt&&u.currentTarget!==fn&&(u.stopPropagation(),pr(!1),s(u))}function l(u){const{items:d}=tt.get(t),c=Array.from(t.children),h=c.indexOf(u);fn=u,fn.tabIndex=0,Mt=d[h][Xe],Kt=c[h].getAttribute("aria-label")||""}function a({items:u=[],type:d=fd,dragDisabled:c=!1,zoneTabIndex:h=0,dropFromOthersDisabled:p=!1,dropTargetStyle:y=Ks,dropTargetClasses:k=[],autoAriaDisabled:_=!1}){n.items=[...u],n.dragDisabled=c,n.dropFromOthersDisabled=p,n.zoneTabIndex=h,n.dropTargetStyle=y,n.dropTargetClasses=k,n.autoAriaDisabled=_,n.type&&d!==n.type&&to(t,n.type),n.type=d,dd(t,d),_||(t.setAttribute("aria-disabled",c),t.setAttribute("role","list"),t.setAttribute("aria-describedby",c?Kr.DND_ZONE_DRAG_DISABLED:Kr.DND_ZONE_ACTIVE)),tt.set(t,n),wt?t.tabIndex=t===ot||fn.contains(t)||n.dropFromOthersDisabled||ot&&n.type!==tt.get(ot).type?-1:0:t.tabIndex=n.zoneTabIndex,t.addEventListener("focus",hd);for(let C=0;C<t.children.length;C++){const T=t.children[C];Xr.add(T),T.tabIndex=wt?-1:0,_||T.setAttribute("role","listitem"),T.removeEventListener("keydown",Js.get(T)),T.removeEventListener("click",eo.get(T)),c||(T.addEventListener("keydown",i),Js.set(T,i),T.addEventListener("click",o),eo.set(T,o)),wt&&n.items[C][Xe]===Mt&&(fn=T,fn.tabIndex=0,T.focus())}}a(e);const f={update:u=>{a(u)},destroy:()=>{to(t,n.type),tt.delete(t),zi.delete(t)}};return zi.set(t,f),f}function md(t,e){no(e);const n=od(t,e),r=pd(t,e);return{update:i=>{no(i),n.update(i),r.update(i)},destroy:()=>{n.destroy(),r.destroy()}}}function no(t){const{items:e,flipDurationMs:n,type:r,dragDisabled:i,morphDisabled:s,dropFromOthersDisabled:o,zoneTabIndex:l,dropTargetStyle:a,dropTargetClasses:f,transformDraggedElement:u,autoAriaDisabled:d,centreDraggedOnCursor:c,...h}=t;if(Object.keys(h).length>0&&console.warn("dndzone will ignore unknown options",h),!e)throw new Error("no 'items' key provided to dndzone");const p=e.find(y=>!{}.hasOwnProperty.call(y,Xe));if(p)throw new Error(`missing '${Xe}' property for item ${bi(p)}`);if(f&&!Array.isArray(f))throw new Error(`dropTargetClasses should be an array but instead it is a ${typeof f}, ${bi(f)}`);if(l&&!gd(l))throw new Error(`zoneTabIndex should be a number but instead it is a ${typeof l}, ${bi(l)}`)}function gd(t){return!isNaN(t)&&function(e){return(e|0)===e}(parseFloat(t))}const dt=Bt(uc()),sn=Bt(null);sn.subscribe(()=>{dt.update(t=>t)});function _d(t){dt.set(t),t.questions.length>0?sn.set(t.questions[0]):sn.set(null)}function vd(){dt.update(t=>{const e=al();return t.questions.push(e),t})}function yd(t,e){dt.update(n=>(fl(n.questions,t,e),n))}function bd(t){dt.update(e=>{const n=e.questions.findIndex(r=>r.id===t.id);return sn.update(r=>r===null||r.id===t.id?null:r),e.questions.splice(n,1),e})}function wd(){dt.update(t=>{const e=mn(1,t.questions.length/2);let n=0;for(;n<e;){const r=mn(0,t.questions.length-1),i=mn(0,t.questions.length-1);r!==i&&(yd(r,i),n++)}return t})}function ro(t){if((t.ty===rt.Multiple||t.ty===rt.Single)&&(t.answers=t.answers??[],t.ty===rt.Multiple)){let e=0;for(let n=0;n<t.answers.length;n++)t.answers[n].correct&&e++;t.correct_answers=e}return t}function kd(t,e){return e.forEach(function(n){n&&typeof n!="string"&&!Array.isArray(n)&&Object.keys(n).forEach(function(r){if(r!=="default"&&!(r in t)){var i=Object.getOwnPropertyDescriptor(n,r);Object.defineProperty(t,r,i.get?i:{enumerable:!0,get:function(){return n[r]}})}})}),Object.freeze(t)}function Ol(t,e){return new Promise(function(n,r){let i;return Ed(t).then(function(s){try{return i=s,n(new Blob([e.slice(0,2),i,e.slice(2)],{type:"image/jpeg"}))}catch(o){return r(o)}},r)})}const Ed=t=>new Promise((e,n)=>{const r=new FileReader;r.addEventListener("load",({target:{result:i}})=>{const s=new DataView(i);let o=0;if(s.getUint16(o)!==65496)return n("not a valid JPEG");for(o+=2;;){const l=s.getUint16(o);if(l===65498)break;const a=s.getUint16(o+2);if(l===65505&&s.getUint32(o+4)===1165519206){const f=o+10;let u;switch(s.getUint16(f)){case 18761:u=!0;break;case 19789:u=!1;break;default:return n("TIFF header contains invalid endian")}if(s.getUint16(f+2,u)!==42)return n("TIFF header contains invalid version");const d=s.getUint32(f+4,u),c=f+d+2+12*s.getUint16(f+d,u);for(let h=f+d+2;h<c;h+=12)if(s.getUint16(h,u)==274){if(s.getUint16(h+2,u)!==3)return n("Orientation data type is invalid");if(s.getUint32(h+4,u)!==1)return n("Orientation data count is invalid");s.setUint16(h+8,1,u);break}return e(i.slice(o,o+2+a))}o+=2+a}return e(new Blob)}),r.readAsArrayBuffer(t)});var Jr={},Ad={get exports(){return Jr},set exports(t){Jr=t}};(function(t){var e,n,r={};Ad.exports=r,r.parse=function(i,s){for(var o=r.bin.readUshort,l=r.bin.readUint,a=0,f={},u=new Uint8Array(i),d=u.length-4;l(u,d)!=101010256;)d--;a=d,a+=4;var c=o(u,a+=4);o(u,a+=2);var h=l(u,a+=2),p=l(u,a+=4);a+=4,a=p;for(var y=0;y<c;y++){l(u,a),a+=4,a+=4,a+=4,l(u,a+=4),h=l(u,a+=4);var k=l(u,a+=4),_=o(u,a+=4),C=o(u,a+2),T=o(u,a+4);a+=6;var S=l(u,a+=8);a+=4,a+=_+C+T,r._readLocal(u,S,f,h,k,s)}return f},r._readLocal=function(i,s,o,l,a,f){var u=r.bin.readUshort,d=r.bin.readUint;d(i,s),u(i,s+=4),u(i,s+=2);var c=u(i,s+=2);d(i,s+=2),d(i,s+=4),s+=4;var h=u(i,s+=8),p=u(i,s+=2);s+=2;var y=r.bin.readUTF8(i,s,h);if(s+=h,s+=p,f)o[y]={size:a,csize:l};else{var k=new Uint8Array(i.buffer,s);if(c==0)o[y]=new Uint8Array(k.buffer.slice(s,s+l));else{if(c!=8)throw"unknown compression method: "+c;var _=new Uint8Array(a);r.inflateRaw(k,_),o[y]=_}}},r.inflateRaw=function(i,s){return r.F.inflate(i,s)},r.inflate=function(i,s){return i[0],i[1],r.inflateRaw(new Uint8Array(i.buffer,i.byteOffset+2,i.length-6),s)},r.deflate=function(i,s){s==null&&(s={level:6});var o=0,l=new Uint8Array(50+Math.floor(1.1*i.length));l[o]=120,l[o+1]=156,o+=2,o=r.F.deflateRaw(i,l,o,s.level);var a=r.adler(i,0,i.length);return l[o+0]=a>>>24&255,l[o+1]=a>>>16&255,l[o+2]=a>>>8&255,l[o+3]=a>>>0&255,new Uint8Array(l.buffer,0,o+4)},r.deflateRaw=function(i,s){s==null&&(s={level:6});var o=new Uint8Array(50+Math.floor(1.1*i.length)),l=r.F.deflateRaw(i,o,l,s.level);return new Uint8Array(o.buffer,0,l)},r.encode=function(i,s){s==null&&(s=!1);var o=0,l=r.bin.writeUint,a=r.bin.writeUshort,f={};for(var u in i){var d=!r._noNeed(u)&&!s,c=i[u],h=r.crc.crc(c,0,c.length);f[u]={cpr:d,usize:c.length,crc:h,file:d?r.deflateRaw(c):c}}for(var u in f)o+=f[u].file.length+30+46+2*r.bin.sizeUTF8(u);o+=22;var p=new Uint8Array(o),y=0,k=[];for(var u in f){var _=f[u];k.push(y),y=r._writeHeader(p,y,u,_,0)}var C=0,T=y;for(var u in f)_=f[u],k.push(y),y=r._writeHeader(p,y,u,_,1,k[C++]);var S=y-T;return l(p,y,101010256),y+=4,a(p,y+=4,C),a(p,y+=2,C),l(p,y+=2,S),l(p,y+=4,T),y+=4,y+=2,p.buffer},r._noNeed=function(i){var s=i.split(".").pop().toLowerCase();return"png,jpg,jpeg,zip".indexOf(s)!=-1},r._writeHeader=function(i,s,o,l,a,f){var u=r.bin.writeUint,d=r.bin.writeUshort,c=l.file;return u(i,s,a==0?67324752:33639248),s+=4,a==1&&(s+=2),d(i,s,20),d(i,s+=2,0),d(i,s+=2,l.cpr?8:0),u(i,s+=2,0),u(i,s+=4,l.crc),u(i,s+=4,c.length),u(i,s+=4,l.usize),d(i,s+=4,r.bin.sizeUTF8(o)),d(i,s+=2,0),s+=2,a==1&&(s+=2,s+=2,u(i,s+=6,f),s+=4),s+=r.bin.writeUTF8(i,s,o),a==0&&(i.set(c,s),s+=c.length),s},r.crc={table:function(){for(var i=new Uint32Array(256),s=0;s<256;s++){for(var o=s,l=0;l<8;l++)1&o?o=3988292384^o>>>1:o>>>=1;i[s]=o}return i}(),update:function(i,s,o,l){for(var a=0;a<l;a++)i=r.crc.table[255&(i^s[o+a])]^i>>>8;return i},crc:function(i,s,o){return 4294967295^r.crc.update(4294967295,i,s,o)}},r.adler=function(i,s,o){for(var l=1,a=0,f=s,u=s+o;f<u;){for(var d=Math.min(f+5552,u);f<d;)a+=l+=i[f++];l%=65521,a%=65521}return a<<16|l},r.bin={readUshort:function(i,s){return i[s]|i[s+1]<<8},writeUshort:function(i,s,o){i[s]=255&o,i[s+1]=o>>8&255},readUint:function(i,s){return 16777216*i[s+3]+(i[s+2]<<16|i[s+1]<<8|i[s])},writeUint:function(i,s,o){i[s]=255&o,i[s+1]=o>>8&255,i[s+2]=o>>16&255,i[s+3]=o>>24&255},readASCII:function(i,s,o){for(var l="",a=0;a<o;a++)l+=String.fromCharCode(i[s+a]);return l},writeASCII:function(i,s,o){for(var l=0;l<o.length;l++)i[s+l]=o.charCodeAt(l)},pad:function(i){return i.length<2?"0"+i:i},readUTF8:function(i,s,o){for(var l,a="",f=0;f<o;f++)a+="%"+r.bin.pad(i[s+f].toString(16));try{l=decodeURIComponent(a)}catch{return r.bin.readASCII(i,s,o)}return l},writeUTF8:function(i,s,o){for(var l=o.length,a=0,f=0;f<l;f++){var u=o.charCodeAt(f);if(!(4294967168&u))i[s+a]=u,a++;else if(!(4294965248&u))i[s+a]=192|u>>6,i[s+a+1]=128|u>>0&63,a+=2;else if(!(4294901760&u))i[s+a]=224|u>>12,i[s+a+1]=128|u>>6&63,i[s+a+2]=128|u>>0&63,a+=3;else{if(4292870144&u)throw"e";i[s+a]=240|u>>18,i[s+a+1]=128|u>>12&63,i[s+a+2]=128|u>>6&63,i[s+a+3]=128|u>>0&63,a+=4}}return a},sizeUTF8:function(i){for(var s=i.length,o=0,l=0;l<s;l++){var a=i.charCodeAt(l);if(!(4294967168&a))o++;else if(!(4294965248&a))o+=2;else if(!(4294901760&a))o+=3;else{if(4292870144&a)throw"e";o+=4}}return o}},r.F={},r.F.deflateRaw=function(i,s,o,l){var a=[[0,0,0,0,0],[4,4,8,4,0],[4,5,16,8,0],[4,6,16,16,0],[4,10,16,32,0],[8,16,32,32,0],[8,16,128,128,0],[8,32,128,256,0],[32,128,258,1024,1],[32,258,258,4096,1]][l],f=r.F.U,u=r.F._goodIndex;r.F._hash;var d=r.F._putsE,c=0,h=o<<3,p=0,y=i.length;if(l==0){for(;c<y;)d(s,h,c+(O=Math.min(65535,y-c))==y?1:0),h=r.F._copyExact(i,c,O,s,h+8),c+=O;return h>>>3}var k=f.lits,_=f.strt,C=f.prev,T=0,S=0,$=0,w=0,x=0,v=0;for(y>2&&(_[v=r.F._hash(i,0)]=0),c=0;c<y;c++){if(x=v,c+1<y-2){v=r.F._hash(i,c+1);var b=c+1&32767;C[b]=_[v],_[v]=b}if(p<=c){(T>14e3||S>26697)&&y-c>100&&(p<c&&(k[T]=c-p,T+=2,p=c),h=r.F._writeBlock(c==y-1||p==y?1:0,k,T,w,i,$,c-$,s,h),T=S=w=0,$=c);var I=0;c<y-2&&(I=r.F._bestMatch(i,c,C,x,Math.min(a[2],y-c),a[3]));var O=I>>>16,R=65535&I;if(I!=0){R=65535&I;var N=u(O=I>>>16,f.of0);f.lhst[257+N]++;var M=u(R,f.df0);f.dhst[M]++,w+=f.exb[N]+f.dxb[M],k[T]=O<<23|c-p,k[T+1]=R<<16|N<<8|M,T+=2,p=c+O}else f.lhst[i[c]]++;S++}}for($==c&&i.length!=0||(p<c&&(k[T]=c-p,T+=2,p=c),h=r.F._writeBlock(1,k,T,w,i,$,c-$,s,h),T=0,S=0,T=S=w=0,$=c);7&h;)h++;return h>>>3},r.F._bestMatch=function(i,s,o,l,a,f){var u=32767&s,d=o[u],c=u-d+32768&32767;if(d==u||l!=r.F._hash(i,s-c))return 0;for(var h=0,p=0,y=Math.min(32767,s);c<=y&&--f!=0&&d!=u;){if(h==0||i[s+h]==i[s+h-c]){var k=r.F._howLong(i,s,c);if(k>h){if(p=c,(h=k)>=a)break;c+2<k&&(k=c+2);for(var _=0,C=0;C<k-2;C++){var T=s-c+C+32768&32767,S=T-o[T]+32768&32767;S>_&&(_=S,d=T)}}}c+=(u=d)-(d=o[u])+32768&32767}return h<<16|p},r.F._howLong=function(i,s,o){if(i[s]!=i[s-o]||i[s+1]!=i[s+1-o]||i[s+2]!=i[s+2-o])return 0;var l=s,a=Math.min(i.length,s+258);for(s+=3;s<a&&i[s]==i[s-o];)s++;return s-l},r.F._hash=function(i,s){return(i[s]<<8|i[s+1])+(i[s+2]<<4)&65535},r.saved=0,r.F._writeBlock=function(i,s,o,l,a,f,u,d,c){var h,p,y,k,_,C,T,S,$,w=r.F.U,x=r.F._putsF,v=r.F._putsE;w.lhst[256]++,p=(h=r.F.getTrees())[0],y=h[1],k=h[2],_=h[3],C=h[4],T=h[5],S=h[6],$=h[7];var b=32+(c+3&7?8-(c+3&7):0)+(u<<3),I=l+r.F.contSize(w.fltree,w.lhst)+r.F.contSize(w.fdtree,w.dhst),O=l+r.F.contSize(w.ltree,w.lhst)+r.F.contSize(w.dtree,w.dhst);O+=14+3*T+r.F.contSize(w.itree,w.ihst)+(2*w.ihst[16]+3*w.ihst[17]+7*w.ihst[18]);for(var R=0;R<286;R++)w.lhst[R]=0;for(R=0;R<30;R++)w.dhst[R]=0;for(R=0;R<19;R++)w.ihst[R]=0;var N=b<I&&b<O?0:I<O?1:2;if(x(d,c,i),x(d,c+1,N),c+=3,N==0){for(;7&c;)c++;c=r.F._copyExact(a,f,u,d,c)}else{var M,L;if(N==1&&(M=w.fltree,L=w.fdtree),N==2){r.F.makeCodes(w.ltree,p),r.F.revCodes(w.ltree,p),r.F.makeCodes(w.dtree,y),r.F.revCodes(w.dtree,y),r.F.makeCodes(w.itree,k),r.F.revCodes(w.itree,k),M=w.ltree,L=w.dtree,v(d,c,_-257),v(d,c+=5,C-1),v(d,c+=5,T-4),c+=4;for(var A=0;A<T;A++)v(d,c+3*A,w.itree[1+(w.ordr[A]<<1)]);c+=3*T,c=r.F._codeTiny(S,w.itree,d,c),c=r.F._codeTiny($,w.itree,d,c)}for(var E=f,q=0;q<o;q+=2){for(var F=s[q],P=F>>>23,Y=E+(8388607&F);E<Y;)c=r.F._writeLit(a[E++],M,d,c);if(P!=0){var j=s[q+1],H=j>>16,z=j>>8&255,Z=255&j;v(d,c=r.F._writeLit(257+z,M,d,c),P-w.of0[z]),c+=w.exb[z],x(d,c=r.F._writeLit(Z,L,d,c),H-w.df0[Z]),c+=w.dxb[Z],E+=P}}c=r.F._writeLit(256,M,d,c)}return c},r.F._copyExact=function(i,s,o,l,a){var f=a>>>3;return l[f]=o,l[f+1]=o>>>8,l[f+2]=255-l[f],l[f+3]=255-l[f+1],f+=4,l.set(new Uint8Array(i.buffer,s,o),f),a+(o+4<<3)},r.F.getTrees=function(){for(var i=r.F.U,s=r.F._hufTree(i.lhst,i.ltree,15),o=r.F._hufTree(i.dhst,i.dtree,15),l=[],a=r.F._lenCodes(i.ltree,l),f=[],u=r.F._lenCodes(i.dtree,f),d=0;d<l.length;d+=2)i.ihst[l[d]]++;for(d=0;d<f.length;d+=2)i.ihst[f[d]]++;for(var c=r.F._hufTree(i.ihst,i.itree,7),h=19;h>4&&i.itree[1+(i.ordr[h-1]<<1)]==0;)h--;return[s,o,c,a,u,h,l,f]},r.F.getSecond=function(i){for(var s=[],o=0;o<i.length;o+=2)s.push(i[o+1]);return s},r.F.nonZero=function(i){for(var s="",o=0;o<i.length;o+=2)i[o+1]!=0&&(s+=(o>>1)+",");return s},r.F.contSize=function(i,s){for(var o=0,l=0;l<s.length;l++)o+=s[l]*i[1+(l<<1)];return o},r.F._codeTiny=function(i,s,o,l){for(var a=0;a<i.length;a+=2){var f=i[a],u=i[a+1];l=r.F._writeLit(f,s,o,l);var d=f==16?2:f==17?3:7;f>15&&(r.F._putsE(o,l,u,d),l+=d)}return l},r.F._lenCodes=function(i,s){for(var o=i.length;o!=2&&i[o-1]==0;)o-=2;for(var l=0;l<o;l+=2){var a=i[l+1],f=l+3<o?i[l+3]:-1,u=l+5<o?i[l+5]:-1,d=l==0?-1:i[l-1];if(a==0&&f==a&&u==a){for(var c=l+5;c+2<o&&i[c+2]==a;)c+=2;(h=Math.min(c+1-l>>>1,138))<11?s.push(17,h-3):s.push(18,h-11),l+=2*h-2}else if(a==d&&f==a&&u==a){for(c=l+5;c+2<o&&i[c+2]==a;)c+=2;var h=Math.min(c+1-l>>>1,6);s.push(16,h-3),l+=2*h-2}else s.push(a,0)}return o>>>1},r.F._hufTree=function(i,s,o){var l=[],a=i.length,f=s.length,u=0;for(u=0;u<f;u+=2)s[u]=0,s[u+1]=0;for(u=0;u<a;u++)i[u]!=0&&l.push({lit:u,f:i[u]});var d=l.length,c=l.slice(0);if(d==0)return 0;if(d==1){var h=l[0].lit;return c=h==0?1:0,s[1+(h<<1)]=1,s[1+(c<<1)]=1,1}l.sort(function(S,$){return S.f-$.f});var p=l[0],y=l[1],k=0,_=1,C=2;for(l[0]={lit:-1,f:p.f+y.f,l:p,r:y,d:0};_!=d-1;)p=k!=_&&(C==d||l[k].f<l[C].f)?l[k++]:l[C++],y=k!=_&&(C==d||l[k].f<l[C].f)?l[k++]:l[C++],l[_++]={lit:-1,f:p.f+y.f,l:p,r:y};var T=r.F.setDepth(l[_-1],0);for(T>o&&(r.F.restrictDepth(c,o,T),T=o),u=0;u<d;u++)s[1+(c[u].lit<<1)]=c[u].d;return T},r.F.setDepth=function(i,s){return i.lit!=-1?(i.d=s,s):Math.max(r.F.setDepth(i.l,s+1),r.F.setDepth(i.r,s+1))},r.F.restrictDepth=function(i,s,o){var l=0,a=1<<o-s,f=0;for(i.sort(function(d,c){return c.d==d.d?d.f-c.f:c.d-d.d}),l=0;l<i.length&&i[l].d>s;l++){var u=i[l].d;i[l].d=s,f+=a-(1<<o-u)}for(f>>>=o-s;f>0;)(u=i[l].d)<s?(i[l].d++,f-=1<<s-u-1):l++;for(;l>=0;l--)i[l].d==s&&f<0&&(i[l].d--,f++);f!=0&&console.log("debt left")},r.F._goodIndex=function(i,s){var o=0;return s[16|o]<=i&&(o|=16),s[8|o]<=i&&(o|=8),s[4|o]<=i&&(o|=4),s[2|o]<=i&&(o|=2),s[1|o]<=i&&(o|=1),o},r.F._writeLit=function(i,s,o,l){return r.F._putsF(o,l,s[i<<1]),l+s[1+(i<<1)]},r.F.inflate=function(i,s){var o=Uint8Array;if(i[0]==3&&i[1]==0)return s||new o(0);var l=r.F,a=l._bitsF,f=l._bitsE,u=l._decodeTiny,d=l.makeCodes,c=l.codes2map,h=l._get17,p=l.U,y=s==null;y&&(s=new o(i.length>>>2<<3));for(var k,_,C=0,T=0,S=0,$=0,w=0,x=0,v=0,b=0,I=0;C==0;)if(C=a(i,I,1),T=a(i,I+1,2),I+=3,T!=0){if(y&&(s=r.F._check(s,b+(1<<17))),T==1&&(k=p.flmap,_=p.fdmap,x=511,v=31),T==2){S=f(i,I,5)+257,$=f(i,I+5,5)+1,w=f(i,I+10,4)+4,I+=14;for(var O=0;O<38;O+=2)p.itree[O]=0,p.itree[O+1]=0;var R=1;for(O=0;O<w;O++){var N=f(i,I+3*O,3);p.itree[1+(p.ordr[O]<<1)]=N,N>R&&(R=N)}I+=3*w,d(p.itree,R),c(p.itree,R,p.imap),k=p.lmap,_=p.dmap,I=u(p.imap,(1<<R)-1,S+$,i,I,p.ttree);var M=l._copyOut(p.ttree,0,S,p.ltree);x=(1<<M)-1;var L=l._copyOut(p.ttree,S,$,p.dtree);v=(1<<L)-1,d(p.ltree,M),c(p.ltree,M,k),d(p.dtree,L),c(p.dtree,L,_)}for(;;){var A=k[h(i,I)&x];I+=15&A;var E=A>>>4;if(!(E>>>8))s[b++]=E;else{if(E==256)break;var q=b+E-254;if(E>264){var F=p.ldef[E-257];q=b+(F>>>3)+f(i,I,7&F),I+=7&F}var P=_[h(i,I)&v];I+=15&P;var Y=P>>>4,j=p.ddef[Y],H=(j>>>4)+a(i,I,15&j);for(I+=15&j,y&&(s=r.F._check(s,b+(1<<17)));b<q;)s[b]=s[b++-H],s[b]=s[b++-H],s[b]=s[b++-H],s[b]=s[b++-H];b=q}}}else{7&I&&(I+=8-(7&I));var z=4+(I>>>3),Z=i[z-4]|i[z-3]<<8;y&&(s=r.F._check(s,b+Z)),s.set(new o(i.buffer,i.byteOffset+z,Z),b),I=z+Z<<3,b+=Z}return s.length==b?s:s.slice(0,b)},r.F._check=function(i,s){var o=i.length;if(s<=o)return i;var l=new Uint8Array(Math.max(o<<1,s));return l.set(i,0),l},r.F._decodeTiny=function(i,s,o,l,a,f){for(var u=r.F._bitsE,d=r.F._get17,c=0;c<o;){var h=i[d(l,a)&s];a+=15&h;var p=h>>>4;if(p<=15)f[c]=p,c++;else{var y=0,k=0;p==16?(k=3+u(l,a,2),a+=2,y=f[c-1]):p==17?(k=3+u(l,a,3),a+=3):p==18&&(k=11+u(l,a,7),a+=7);for(var _=c+k;c<_;)f[c]=y,c++}}return a},r.F._copyOut=function(i,s,o,l){for(var a=0,f=0,u=l.length>>>1;f<o;){var d=i[f+s];l[f<<1]=0,l[1+(f<<1)]=d,d>a&&(a=d),f++}for(;f<u;)l[f<<1]=0,l[1+(f<<1)]=0,f++;return a},r.F.makeCodes=function(i,s){for(var o,l,a,f,u=r.F.U,d=i.length,c=u.bl_count,h=0;h<=s;h++)c[h]=0;for(h=1;h<d;h+=2)c[i[h]]++;var p=u.next_code;for(o=0,c[0]=0,l=1;l<=s;l++)o=o+c[l-1]<<1,p[l]=o;for(a=0;a<d;a+=2)(f=i[a+1])!=0&&(i[a]=p[f],p[f]++)},r.F.codes2map=function(i,s,o){for(var l=i.length,a=r.F.U.rev15,f=0;f<l;f+=2)if(i[f+1]!=0)for(var u=f>>1,d=i[f+1],c=u<<4|d,h=s-d,p=i[f]<<h,y=p+(1<<h);p!=y;)o[a[p]>>>15-s]=c,p++},r.F.revCodes=function(i,s){for(var o=r.F.U.rev15,l=15-s,a=0;a<i.length;a+=2){var f=i[a]<<s-i[a+1];i[a]=o[f]>>>l}},r.F._putsE=function(i,s,o){o<<=7&s;var l=s>>>3;i[l]|=o,i[l+1]|=o>>>8},r.F._putsF=function(i,s,o){o<<=7&s;var l=s>>>3;i[l]|=o,i[l+1]|=o>>>8,i[l+2]|=o>>>16},r.F._bitsE=function(i,s,o){return(i[s>>>3]|i[1+(s>>>3)]<<8)>>>(7&s)&(1<<o)-1},r.F._bitsF=function(i,s,o){return(i[s>>>3]|i[1+(s>>>3)]<<8|i[2+(s>>>3)]<<16)>>>(7&s)&(1<<o)-1},r.F._get17=function(i,s){return(i[s>>>3]|i[1+(s>>>3)]<<8|i[2+(s>>>3)]<<16)>>>(7&s)},r.F._get25=function(i,s){return(i[s>>>3]|i[1+(s>>>3)]<<8|i[2+(s>>>3)]<<16|i[3+(s>>>3)]<<24)>>>(7&s)},r.F.U=(e=Uint16Array,n=Uint32Array,{next_code:new e(16),bl_count:new e(16),ordr:[16,17,18,0,8,7,9,6,10,5,11,4,12,3,13,2,14,1,15],of0:[3,4,5,6,7,8,9,10,11,13,15,17,19,23,27,31,35,43,51,59,67,83,99,115,131,163,195,227,258,999,999,999],exb:[0,0,0,0,0,0,0,0,1,1,1,1,2,2,2,2,3,3,3,3,4,4,4,4,5,5,5,5,0,0,0,0],ldef:new e(32),df0:[1,2,3,4,5,7,9,13,17,25,33,49,65,97,129,193,257,385,513,769,1025,1537,2049,3073,4097,6145,8193,12289,16385,24577,65535,65535],dxb:[0,0,0,0,1,1,2,2,3,3,4,4,5,5,6,6,7,7,8,8,9,9,10,10,11,11,12,12,13,13,0,0],ddef:new n(32),flmap:new e(512),fltree:[],fdmap:new e(32),fdtree:[],lmap:new e(32768),ltree:[],ttree:[],dmap:new e(32768),dtree:[],imap:new e(512),itree:[],rev15:new e(32768),lhst:new n(286),dhst:new n(30),ihst:new n(19),lits:new n(15e3),strt:new e(65536),prev:new e(32768)}),function(){for(var i=r.F.U,s=0;s<32768;s++){var o=s;o=(4278255360&(o=(4042322160&(o=(3435973836&(o=(2863311530&o)>>>1|(1431655765&o)<<1))>>>2|(858993459&o)<<2))>>>4|(252645135&o)<<4))>>>8|(16711935&o)<<8,i.rev15[s]=(o>>>16|o<<16)>>>17}function l(a,f,u){for(;f--!=0;)a.push(0,u)}for(s=0;s<32;s++)i.ldef[s]=i.of0[s]<<3|i.exb[s],i.ddef[s]=i.df0[s]<<4|i.dxb[s];l(i.fltree,144,8),l(i.fltree,112,9),l(i.fltree,24,7),l(i.fltree,8,8),r.F.makeCodes(i.fltree,9),r.F.codes2map(i.fltree,9,i.flmap),r.F.revCodes(i.fltree,9),l(i.fdtree,32,5),r.F.makeCodes(i.fdtree,5),r.F.codes2map(i.fdtree,5,i.fdmap),r.F.revCodes(i.fdtree,5),l(i.itree,19,0),l(i.ltree,286,0),l(i.dtree,30,0),l(i.ttree,320,0)}()})();var Td=kd({__proto__:null,default:Jr},[Jr]);const kt=function(){var t={nextZero(u,d){for(;u[d]!=0;)d++;return d},readUshort:(u,d)=>u[d]<<8|u[d+1],writeUshort(u,d,c){u[d]=c>>8&255,u[d+1]=255&c},readUint:(u,d)=>16777216*u[d]+(u[d+1]<<16|u[d+2]<<8|u[d+3]),writeUint(u,d,c){u[d]=c>>24&255,u[d+1]=c>>16&255,u[d+2]=c>>8&255,u[d+3]=255&c},readASCII(u,d,c){let h="";for(let p=0;p<c;p++)h+=String.fromCharCode(u[d+p]);return h},writeASCII(u,d,c){for(let h=0;h<c.length;h++)u[d+h]=c.charCodeAt(h)},readBytes(u,d,c){const h=[];for(let p=0;p<c;p++)h.push(u[d+p]);return h},pad:u=>u.length<2?`0${u}`:u,readUTF8(u,d,c){let h,p="";for(let y=0;y<c;y++)p+=`%${t.pad(u[d+y].toString(16))}`;try{h=decodeURIComponent(p)}catch{return t.readASCII(u,d,c)}return h}};function e(u,d,c,h){const p=d*c,y=s(h),k=Math.ceil(d*y/8),_=new Uint8Array(4*p),C=new Uint32Array(_.buffer),{ctype:T}=h,{depth:S}=h,$=t.readUshort;if(T==6){const F=p<<2;if(S==8)for(var w=0;w<F;w+=4)_[w]=u[w],_[w+1]=u[w+1],_[w+2]=u[w+2],_[w+3]=u[w+3];if(S==16)for(w=0;w<F;w++)_[w]=u[w<<1]}else if(T==2){const F=h.tabs.tRNS;if(F==null){if(S==8)for(w=0;w<p;w++){var x=3*w;C[w]=255<<24|u[x+2]<<16|u[x+1]<<8|u[x]}if(S==16)for(w=0;w<p;w++)x=6*w,C[w]=255<<24|u[x+4]<<16|u[x+2]<<8|u[x]}else{var v=F[0];const P=F[1],Y=F[2];if(S==8)for(w=0;w<p;w++){var b=w<<2;x=3*w,C[w]=255<<24|u[x+2]<<16|u[x+1]<<8|u[x],u[x]==v&&u[x+1]==P&&u[x+2]==Y&&(_[b+3]=0)}if(S==16)for(w=0;w<p;w++)b=w<<2,x=6*w,C[w]=255<<24|u[x+4]<<16|u[x+2]<<8|u[x],$(u,x)==v&&$(u,x+2)==P&&$(u,x+4)==Y&&(_[b+3]=0)}}else if(T==3){const F=h.tabs.PLTE,P=h.tabs.tRNS,Y=P?P.length:0;if(S==1)for(var I=0;I<c;I++){var O=I*k,R=I*d;for(w=0;w<d;w++){b=R+w<<2;var N=3*(M=u[O+(w>>3)]>>7-((7&w)<<0)&1);_[b]=F[N],_[b+1]=F[N+1],_[b+2]=F[N+2],_[b+3]=M<Y?P[M]:255}}if(S==2)for(I=0;I<c;I++)for(O=I*k,R=I*d,w=0;w<d;w++)b=R+w<<2,N=3*(M=u[O+(w>>2)]>>6-((3&w)<<1)&3),_[b]=F[N],_[b+1]=F[N+1],_[b+2]=F[N+2],_[b+3]=M<Y?P[M]:255;if(S==4)for(I=0;I<c;I++)for(O=I*k,R=I*d,w=0;w<d;w++)b=R+w<<2,N=3*(M=u[O+(w>>1)]>>4-((1&w)<<2)&15),_[b]=F[N],_[b+1]=F[N+1],_[b+2]=F[N+2],_[b+3]=M<Y?P[M]:255;if(S==8)for(w=0;w<p;w++){var M;b=w<<2,N=3*(M=u[w]),_[b]=F[N],_[b+1]=F[N+1],_[b+2]=F[N+2],_[b+3]=M<Y?P[M]:255}}else if(T==4){if(S==8)for(w=0;w<p;w++){b=w<<2;var L=u[A=w<<1];_[b]=L,_[b+1]=L,_[b+2]=L,_[b+3]=u[A+1]}if(S==16)for(w=0;w<p;w++){var A;b=w<<2,L=u[A=w<<2],_[b]=L,_[b+1]=L,_[b+2]=L,_[b+3]=u[A+2]}}else if(T==0)for(v=h.tabs.tRNS?h.tabs.tRNS:-1,I=0;I<c;I++){const F=I*k,P=I*d;if(S==1)for(var E=0;E<d;E++){var q=(L=255*(u[F+(E>>>3)]>>>7-(7&E)&1))==255*v?0:255;C[P+E]=q<<24|L<<16|L<<8|L}else if(S==2)for(E=0;E<d;E++)q=(L=85*(u[F+(E>>>2)]>>>6-((3&E)<<1)&3))==85*v?0:255,C[P+E]=q<<24|L<<16|L<<8|L;else if(S==4)for(E=0;E<d;E++)q=(L=17*(u[F+(E>>>1)]>>>4-((1&E)<<2)&15))==17*v?0:255,C[P+E]=q<<24|L<<16|L<<8|L;else if(S==8)for(E=0;E<d;E++)q=(L=u[F+E])==v?0:255,C[P+E]=q<<24|L<<16|L<<8|L;else if(S==16)for(E=0;E<d;E++)L=u[F+(E<<1)],q=$(u,F+(E<<1))==v?0:255,C[P+E]=q<<24|L<<16|L<<8|L}return _}function n(u,d,c,h){const p=s(u),y=Math.ceil(c*p/8),k=new Uint8Array((y+1+u.interlace)*h);return d=u.tabs.CgBI?i(d,k):r(d,k),u.interlace==0?d=o(d,u,0,c,h):u.interlace==1&&(d=function(C,T){const S=T.width,$=T.height,w=s(T),x=w>>3,v=Math.ceil(S*w/8),b=new Uint8Array($*v);let I=0;const O=[0,0,4,0,2,0,1],R=[0,4,0,2,0,1,0],N=[8,8,8,4,4,2,2],M=[8,8,4,4,2,2,1];let L=0;for(;L<7;){const E=N[L],q=M[L];let F=0,P=0,Y=O[L];for(;Y<$;)Y+=E,P++;let j=R[L];for(;j<S;)j+=q,F++;const H=Math.ceil(F*w/8);o(C,T,I,F,P);let z=0,Z=O[L];for(;Z<$;){let ne=R[L],de=I+z*H<<3;for(;ne<S;){var A;if(w==1&&(A=(A=C[de>>3])>>7-(7&de)&1,b[Z*v+(ne>>3)]|=A<<7-((7&ne)<<0)),w==2&&(A=(A=C[de>>3])>>6-(7&de)&3,b[Z*v+(ne>>2)]|=A<<6-((3&ne)<<1)),w==4&&(A=(A=C[de>>3])>>4-(7&de)&15,b[Z*v+(ne>>1)]|=A<<4-((1&ne)<<2)),w>=8){const me=Z*v+ne*x;for(let ke=0;ke<x;ke++)b[me+ke]=C[(de>>3)+ke]}de+=w,ne+=q}z++,Z+=E}F*P!=0&&(I+=P*(1+H)),L+=1}return b}(d,u)),d}function r(u,d){return i(new Uint8Array(u.buffer,2,u.length-6),d)}var i=function(){const u={H:{}};return u.H.N=function(d,c){const h=Uint8Array;let p,y,k=0,_=0,C=0,T=0,S=0,$=0,w=0,x=0,v=0;if(d[0]==3&&d[1]==0)return c||new h(0);const b=u.H,I=b.b,O=b.e,R=b.R,N=b.n,M=b.A,L=b.Z,A=b.m,E=c==null;for(E&&(c=new h(d.length>>>2<<5));k==0;)if(k=I(d,v,1),_=I(d,v+1,2),v+=3,_!=0){if(E&&(c=u.H.W(c,x+(1<<17))),_==1&&(p=A.J,y=A.h,$=511,w=31),_==2){C=O(d,v,5)+257,T=O(d,v+5,5)+1,S=O(d,v+10,4)+4,v+=14;let F=1;for(var q=0;q<38;q+=2)A.Q[q]=0,A.Q[q+1]=0;for(q=0;q<S;q++){const j=O(d,v+3*q,3);A.Q[1+(A.X[q]<<1)]=j,j>F&&(F=j)}v+=3*S,N(A.Q,F),M(A.Q,F,A.u),p=A.w,y=A.d,v=R(A.u,(1<<F)-1,C+T,d,v,A.v);const P=b.V(A.v,0,C,A.C);$=(1<<P)-1;const Y=b.V(A.v,C,T,A.D);w=(1<<Y)-1,N(A.C,P),M(A.C,P,p),N(A.D,Y),M(A.D,Y,y)}for(;;){const F=p[L(d,v)&$];v+=15&F;const P=F>>>4;if(!(P>>>8))c[x++]=P;else{if(P==256)break;{let Y=x+P-254;if(P>264){const ne=A.q[P-257];Y=x+(ne>>>3)+O(d,v,7&ne),v+=7&ne}const j=y[L(d,v)&w];v+=15&j;const H=j>>>4,z=A.c[H],Z=(z>>>4)+I(d,v,15&z);for(v+=15&z;x<Y;)c[x]=c[x++-Z],c[x]=c[x++-Z],c[x]=c[x++-Z],c[x]=c[x++-Z];x=Y}}}}else{7&v&&(v+=8-(7&v));const F=4+(v>>>3),P=d[F-4]|d[F-3]<<8;E&&(c=u.H.W(c,x+P)),c.set(new h(d.buffer,d.byteOffset+F,P),x),v=F+P<<3,x+=P}return c.length==x?c:c.slice(0,x)},u.H.W=function(d,c){const h=d.length;if(c<=h)return d;const p=new Uint8Array(h<<1);return p.set(d,0),p},u.H.R=function(d,c,h,p,y,k){const _=u.H.e,C=u.H.Z;let T=0;for(;T<h;){const S=d[C(p,y)&c];y+=15&S;const $=S>>>4;if($<=15)k[T]=$,T++;else{let w=0,x=0;$==16?(x=3+_(p,y,2),y+=2,w=k[T-1]):$==17?(x=3+_(p,y,3),y+=3):$==18&&(x=11+_(p,y,7),y+=7);const v=T+x;for(;T<v;)k[T]=w,T++}}return y},u.H.V=function(d,c,h,p){let y=0,k=0;const _=p.length>>>1;for(;k<h;){const C=d[k+c];p[k<<1]=0,p[1+(k<<1)]=C,C>y&&(y=C),k++}for(;k<_;)p[k<<1]=0,p[1+(k<<1)]=0,k++;return y},u.H.n=function(d,c){const h=u.H.m,p=d.length;let y,k,_,C;const T=h.j;for(var S=0;S<=c;S++)T[S]=0;for(S=1;S<p;S+=2)T[d[S]]++;const $=h.K;for(y=0,T[0]=0,k=1;k<=c;k++)y=y+T[k-1]<<1,$[k]=y;for(_=0;_<p;_+=2)C=d[_+1],C!=0&&(d[_]=$[C],$[C]++)},u.H.A=function(d,c,h){const p=d.length,y=u.H.m.r;for(let k=0;k<p;k+=2)if(d[k+1]!=0){const _=k>>1,C=d[k+1],T=_<<4|C,S=c-C;let $=d[k]<<S;const w=$+(1<<S);for(;$!=w;)h[y[$]>>>15-c]=T,$++}},u.H.l=function(d,c){const h=u.H.m.r,p=15-c;for(let y=0;y<d.length;y+=2){const k=d[y]<<c-d[y+1];d[y]=h[k]>>>p}},u.H.M=function(d,c,h){h<<=7&c;const p=c>>>3;d[p]|=h,d[p+1]|=h>>>8},u.H.I=function(d,c,h){h<<=7&c;const p=c>>>3;d[p]|=h,d[p+1]|=h>>>8,d[p+2]|=h>>>16},u.H.e=function(d,c,h){return(d[c>>>3]|d[1+(c>>>3)]<<8)>>>(7&c)&(1<<h)-1},u.H.b=function(d,c,h){return(d[c>>>3]|d[1+(c>>>3)]<<8|d[2+(c>>>3)]<<16)>>>(7&c)&(1<<h)-1},u.H.Z=function(d,c){return(d[c>>>3]|d[1+(c>>>3)]<<8|d[2+(c>>>3)]<<16)>>>(7&c)},u.H.i=function(d,c){return(d[c>>>3]|d[1+(c>>>3)]<<8|d[2+(c>>>3)]<<16|d[3+(c>>>3)]<<24)>>>(7&c)},u.H.m=function(){const d=Uint16Array,c=Uint32Array;return{K:new d(16),j:new d(16),X:[16,17,18,0,8,7,9,6,10,5,11,4,12,3,13,2,14,1,15],S:[3,4,5,6,7,8,9,10,11,13,15,17,19,23,27,31,35,43,51,59,67,83,99,115,131,163,195,227,258,999,999,999],T:[0,0,0,0,0,0,0,0,1,1,1,1,2,2,2,2,3,3,3,3,4,4,4,4,5,5,5,5,0,0,0,0],q:new d(32),p:[1,2,3,4,5,7,9,13,17,25,33,49,65,97,129,193,257,385,513,769,1025,1537,2049,3073,4097,6145,8193,12289,16385,24577,65535,65535],z:[0,0,0,0,1,1,2,2,3,3,4,4,5,5,6,6,7,7,8,8,9,9,10,10,11,11,12,12,13,13,0,0],c:new c(32),J:new d(512),_:[],h:new d(32),$:[],w:new d(32768),C:[],v:[],d:new d(32768),D:[],u:new d(512),Q:[],r:new d(32768),s:new c(286),Y:new c(30),a:new c(19),t:new c(15e3),k:new d(65536),g:new d(32768)}}(),function(){const d=u.H.m;for(var c=0;c<32768;c++){let p=c;p=(2863311530&p)>>>1|(1431655765&p)<<1,p=(3435973836&p)>>>2|(858993459&p)<<2,p=(4042322160&p)>>>4|(252645135&p)<<4,p=(4278255360&p)>>>8|(16711935&p)<<8,d.r[c]=(p>>>16|p<<16)>>>17}function h(p,y,k){for(;y--!=0;)p.push(0,k)}for(c=0;c<32;c++)d.q[c]=d.S[c]<<3|d.T[c],d.c[c]=d.p[c]<<4|d.z[c];h(d._,144,8),h(d._,112,9),h(d._,24,7),h(d._,8,8),u.H.n(d._,9),u.H.A(d._,9,d.J),u.H.l(d._,9),h(d.$,32,5),u.H.n(d.$,5),u.H.A(d.$,5,d.h),u.H.l(d.$,5),h(d.Q,19,0),h(d.C,286,0),h(d.D,30,0),h(d.v,320,0)}(),u.H.N}();function s(u){return[1,null,3,1,2,null,4][u.ctype]*u.depth}function o(u,d,c,h,p){let y=s(d);const k=Math.ceil(h*y/8);let _,C;y=Math.ceil(y/8);let T=u[c],S=0;if(T>1&&(u[c]=[0,0,1][T-2]),T==3)for(S=y;S<k;S++)u[S+1]=u[S+1]+(u[S+1-y]>>>1)&255;for(let $=0;$<p;$++)if(_=c+$*k,C=_+$+1,T=u[C-1],S=0,T==0)for(;S<k;S++)u[_+S]=u[C+S];else if(T==1){for(;S<y;S++)u[_+S]=u[C+S];for(;S<k;S++)u[_+S]=u[C+S]+u[_+S-y]}else if(T==2)for(;S<k;S++)u[_+S]=u[C+S]+u[_+S-k];else if(T==3){for(;S<y;S++)u[_+S]=u[C+S]+(u[_+S-k]>>>1);for(;S<k;S++)u[_+S]=u[C+S]+(u[_+S-k]+u[_+S-y]>>>1)}else{for(;S<y;S++)u[_+S]=u[C+S]+l(0,u[_+S-k],0);for(;S<k;S++)u[_+S]=u[C+S]+l(u[_+S-y],u[_+S-k],u[_+S-y-k])}return u}function l(u,d,c){const h=u+d-c,p=h-u,y=h-d,k=h-c;return p*p<=y*y&&p*p<=k*k?u:y*y<=k*k?d:c}function a(u,d,c){c.width=t.readUint(u,d),d+=4,c.height=t.readUint(u,d),d+=4,c.depth=u[d],d++,c.ctype=u[d],d++,c.compress=u[d],d++,c.filter=u[d],d++,c.interlace=u[d],d++}function f(u,d,c,h,p,y,k,_,C){const T=Math.min(d,p),S=Math.min(c,y);let $=0,w=0;for(let L=0;L<S;L++)for(let A=0;A<T;A++)if(k>=0&&_>=0?($=L*d+A<<2,w=(_+L)*p+k+A<<2):($=(-_+L)*d-k+A<<2,w=L*p+A<<2),C==0)h[w]=u[$],h[w+1]=u[$+1],h[w+2]=u[$+2],h[w+3]=u[$+3];else if(C==1){var x=u[$+3]*.00392156862745098,v=u[$]*x,b=u[$+1]*x,I=u[$+2]*x,O=h[w+3]*(1/255),R=h[w]*O,N=h[w+1]*O,M=h[w+2]*O;const E=1-x,q=x+O*E,F=q==0?0:1/q;h[w+3]=255*q,h[w+0]=(v+R*E)*F,h[w+1]=(b+N*E)*F,h[w+2]=(I+M*E)*F}else if(C==2)x=u[$+3],v=u[$],b=u[$+1],I=u[$+2],O=h[w+3],R=h[w],N=h[w+1],M=h[w+2],x==O&&v==R&&b==N&&I==M?(h[w]=0,h[w+1]=0,h[w+2]=0,h[w+3]=0):(h[w]=v,h[w+1]=b,h[w+2]=I,h[w+3]=x);else if(C==3){if(x=u[$+3],v=u[$],b=u[$+1],I=u[$+2],O=h[w+3],R=h[w],N=h[w+1],M=h[w+2],x==O&&v==R&&b==N&&I==M)continue;if(x<220&&O>20)return!1}return!0}return{decode:function(d){const c=new Uint8Array(d);let h=8;const p=t,y=p.readUshort,k=p.readUint,_={tabs:{},frames:[]},C=new Uint8Array(c.length);let T,S=0,$=0;const w=[137,80,78,71,13,10,26,10];for(var x=0;x<8;x++)if(c[x]!=w[x])throw"The input is not a PNG file!";for(;h<c.length;){const L=p.readUint(c,h);h+=4;const A=p.readASCII(c,h,4);if(h+=4,A=="IHDR")a(c,h,_);else if(A=="iCCP"){for(var v=h;c[v]!=0;)v++;p.readASCII(c,h,v-h),c[v+1];const E=c.slice(v+2,h+L);let q=null;try{q=r(E)}catch{q=i(E)}_.tabs[A]=q}else if(A=="CgBI")_.tabs[A]=c.slice(h,h+4);else if(A=="IDAT"){for(x=0;x<L;x++)C[S+x]=c[h+x];S+=L}else if(A=="acTL")_.tabs[A]={num_frames:k(c,h),num_plays:k(c,h+4)},T=new Uint8Array(c.length);else if(A=="fcTL"){$!=0&&((M=_.frames[_.frames.length-1]).data=n(_,T.slice(0,$),M.rect.width,M.rect.height),$=0);const E={x:k(c,h+12),y:k(c,h+16),width:k(c,h+4),height:k(c,h+8)};let q=y(c,h+22);q=y(c,h+20)/(q==0?100:q);const F={rect:E,delay:Math.round(1e3*q),dispose:c[h+24],blend:c[h+25]};_.frames.push(F)}else if(A=="fdAT"){for(x=0;x<L-4;x++)T[$+x]=c[h+x+4];$+=L-4}else if(A=="pHYs")_.tabs[A]=[p.readUint(c,h),p.readUint(c,h+4),c[h+8]];else if(A=="cHRM")for(_.tabs[A]=[],x=0;x<8;x++)_.tabs[A].push(p.readUint(c,h+4*x));else if(A=="tEXt"||A=="zTXt"){_.tabs[A]==null&&(_.tabs[A]={});var b=p.nextZero(c,h),I=p.readASCII(c,h,b-h),O=h+L-b-1;if(A=="tEXt")N=p.readASCII(c,b+1,O);else{var R=r(c.slice(b+2,b+2+O));N=p.readUTF8(R,0,R.length)}_.tabs[A][I]=N}else if(A=="iTXt"){_.tabs[A]==null&&(_.tabs[A]={}),b=0,v=h,b=p.nextZero(c,v),I=p.readASCII(c,v,b-v);const E=c[v=b+1];var N;c[v+1],v+=2,b=p.nextZero(c,v),p.readASCII(c,v,b-v),v=b+1,b=p.nextZero(c,v),p.readUTF8(c,v,b-v),O=L-((v=b+1)-h),E==0?N=p.readUTF8(c,v,O):(R=r(c.slice(v,v+O)),N=p.readUTF8(R,0,R.length)),_.tabs[A][I]=N}else if(A=="PLTE")_.tabs[A]=p.readBytes(c,h,L);else if(A=="hIST"){const E=_.tabs.PLTE.length/3;for(_.tabs[A]=[],x=0;x<E;x++)_.tabs[A].push(y(c,h+2*x))}else if(A=="tRNS")_.ctype==3?_.tabs[A]=p.readBytes(c,h,L):_.ctype==0?_.tabs[A]=y(c,h):_.ctype==2&&(_.tabs[A]=[y(c,h),y(c,h+2),y(c,h+4)]);else if(A=="gAMA")_.tabs[A]=p.readUint(c,h)/1e5;else if(A=="sRGB")_.tabs[A]=c[h];else if(A=="bKGD")_.ctype==0||_.ctype==4?_.tabs[A]=[y(c,h)]:_.ctype==2||_.ctype==6?_.tabs[A]=[y(c,h),y(c,h+2),y(c,h+4)]:_.ctype==3&&(_.tabs[A]=c[h]);else if(A=="IEND")break;h+=L,p.readUint(c,h),h+=4}var M;return $!=0&&((M=_.frames[_.frames.length-1]).data=n(_,T.slice(0,$),M.rect.width,M.rect.height)),_.data=n(_,C,_.width,_.height),delete _.compress,delete _.interlace,delete _.filter,_},toRGBA8:function(d){const c=d.width,h=d.height;if(d.tabs.acTL==null)return[e(d.data,c,h,d).buffer];const p=[];d.frames[0].data==null&&(d.frames[0].data=d.data);const y=c*h*4,k=new Uint8Array(y),_=new Uint8Array(y),C=new Uint8Array(y);for(let S=0;S<d.frames.length;S++){const $=d.frames[S],w=$.rect.x,x=$.rect.y,v=$.rect.width,b=$.rect.height,I=e($.data,v,b,d);if(S!=0)for(var T=0;T<y;T++)C[T]=k[T];if($.blend==0?f(I,v,b,k,c,h,w,x,0):$.blend==1&&f(I,v,b,k,c,h,w,x,1),p.push(k.buffer.slice(0)),$.dispose!=0){if($.dispose==1)f(_,v,b,k,c,h,w,x,0);else if($.dispose==2)for(T=0;T<y;T++)k[T]=C[T]}}return p},_paeth:l,_copyTile:f,_bin:t}}();(function(){const{_copyTile:t}=kt,{_bin:e}=kt,n=kt._paeth;var r={table:function(){const v=new Uint32Array(256);for(let b=0;b<256;b++){let I=b;for(let O=0;O<8;O++)1&I?I=3988292384^I>>>1:I>>>=1;v[b]=I}return v}(),update(v,b,I,O){for(let R=0;R<O;R++)v=r.table[255&(v^b[I+R])]^v>>>8;return v},crc:(v,b,I)=>4294967295^r.update(4294967295,v,b,I)};function i(v,b,I,O){b[I]+=v[0]*O>>4,b[I+1]+=v[1]*O>>4,b[I+2]+=v[2]*O>>4,b[I+3]+=v[3]*O>>4}function s(v){return Math.max(0,Math.min(255,v))}function o(v,b){const I=v[0]-b[0],O=v[1]-b[1],R=v[2]-b[2],N=v[3]-b[3];return I*I+O*O+R*R+N*N}function l(v,b,I,O,R,N,M){M==null&&(M=1);const L=O.length,A=[];for(var E=0;E<L;E++){const Z=O[E];A.push([Z>>>0&255,Z>>>8&255,Z>>>16&255,Z>>>24&255])}for(E=0;E<L;E++){let Z=4294967295;for(var q=0,F=0;F<L;F++){var P=o(A[E],A[F]);F!=E&&P<Z&&(Z=P,q=F)}}const Y=new Uint32Array(R.buffer),j=new Int16Array(b*I*4),H=[0,8,2,10,12,4,14,6,3,11,1,9,15,7,13,5];for(E=0;E<H.length;E++)H[E]=255*((H[E]+.5)/16-.5);for(let Z=0;Z<I;Z++)for(let ne=0;ne<b;ne++){var z;E=4*(Z*b+ne),M!=2?z=[s(v[E]+j[E]),s(v[E+1]+j[E+1]),s(v[E+2]+j[E+2]),s(v[E+3]+j[E+3])]:(P=H[4*(3&Z)+(3&ne)],z=[s(v[E]+P),s(v[E+1]+P),s(v[E+2]+P),s(v[E+3]+P)]),q=0;let de=16777215;for(F=0;F<L;F++){const Te=o(z,A[F]);Te<de&&(de=Te,q=F)}const me=A[q],ke=[z[0]-me[0],z[1]-me[1],z[2]-me[2],z[3]-me[3]];M==1&&(ne!=b-1&&i(ke,j,E+4,7),Z!=I-1&&(ne!=0&&i(ke,j,E+4*b-4,3),i(ke,j,E+4*b,5),ne!=b-1&&i(ke,j,E+4*b+4,1))),N[E>>2]=q,Y[E>>2]=O[q]}}function a(v,b,I,O,R){R==null&&(R={});const{crc:N}=r,M=e.writeUint,L=e.writeUshort,A=e.writeASCII;let E=8;const q=v.frames.length>1;let F,P=!1,Y=33+(q?20:0);if(R.sRGB!=null&&(Y+=13),R.pHYs!=null&&(Y+=21),R.iCCP!=null&&(F=pako.deflate(R.iCCP),Y+=21+F.length+4),v.ctype==3){for(var j=v.plte.length,H=0;H<j;H++)v.plte[H]>>>24!=255&&(P=!0);Y+=8+3*j+4+(P?8+1*j+4:0)}for(var z=0;z<v.frames.length;z++)q&&(Y+=38),Y+=(me=v.frames[z]).cimg.length+12,z!=0&&(Y+=4);Y+=12;const Z=new Uint8Array(Y),ne=[137,80,78,71,13,10,26,10];for(H=0;H<8;H++)Z[H]=ne[H];if(M(Z,E,13),E+=4,A(Z,E,"IHDR"),E+=4,M(Z,E,b),E+=4,M(Z,E,I),E+=4,Z[E]=v.depth,E++,Z[E]=v.ctype,E++,Z[E]=0,E++,Z[E]=0,E++,Z[E]=0,E++,M(Z,E,N(Z,E-17,17)),E+=4,R.sRGB!=null&&(M(Z,E,1),E+=4,A(Z,E,"sRGB"),E+=4,Z[E]=R.sRGB,E++,M(Z,E,N(Z,E-5,5)),E+=4),R.iCCP!=null){const ke=13+F.length;M(Z,E,ke),E+=4,A(Z,E,"iCCP"),E+=4,A(Z,E,"ICC profile"),E+=11,E+=2,Z.set(F,E),E+=F.length,M(Z,E,N(Z,E-(ke+4),ke+4)),E+=4}if(R.pHYs!=null&&(M(Z,E,9),E+=4,A(Z,E,"pHYs"),E+=4,M(Z,E,R.pHYs[0]),E+=4,M(Z,E,R.pHYs[1]),E+=4,Z[E]=R.pHYs[2],E++,M(Z,E,N(Z,E-13,13)),E+=4),q&&(M(Z,E,8),E+=4,A(Z,E,"acTL"),E+=4,M(Z,E,v.frames.length),E+=4,M(Z,E,R.loop!=null?R.loop:0),E+=4,M(Z,E,N(Z,E-12,12)),E+=4),v.ctype==3){for(M(Z,E,3*(j=v.plte.length)),E+=4,A(Z,E,"PLTE"),E+=4,H=0;H<j;H++){const ke=3*H,Te=v.plte[H],ae=255&Te,fe=Te>>>8&255,Ge=Te>>>16&255;Z[E+ke+0]=ae,Z[E+ke+1]=fe,Z[E+ke+2]=Ge}if(E+=3*j,M(Z,E,N(Z,E-3*j-4,3*j+4)),E+=4,P){for(M(Z,E,j),E+=4,A(Z,E,"tRNS"),E+=4,H=0;H<j;H++)Z[E+H]=v.plte[H]>>>24&255;E+=j,M(Z,E,N(Z,E-j-4,j+4)),E+=4}}let de=0;for(z=0;z<v.frames.length;z++){var me=v.frames[z];q&&(M(Z,E,26),E+=4,A(Z,E,"fcTL"),E+=4,M(Z,E,de++),E+=4,M(Z,E,me.rect.width),E+=4,M(Z,E,me.rect.height),E+=4,M(Z,E,me.rect.x),E+=4,M(Z,E,me.rect.y),E+=4,L(Z,E,O[z]),E+=2,L(Z,E,1e3),E+=2,Z[E]=me.dispose,E++,Z[E]=me.blend,E++,M(Z,E,N(Z,E-30,30)),E+=4);const ke=me.cimg;M(Z,E,(j=ke.length)+(z==0?0:4)),E+=4;const Te=E;A(Z,E,z==0?"IDAT":"fdAT"),E+=4,z!=0&&(M(Z,E,de++),E+=4),Z.set(ke,E),E+=j,M(Z,E,N(Z,Te,E-Te)),E+=4}return M(Z,E,0),E+=4,A(Z,E,"IEND"),E+=4,M(Z,E,N(Z,E-4,4)),E+=4,Z.buffer}function f(v,b,I){for(let O=0;O<v.frames.length;O++){const R=v.frames[O];R.rect.width;const N=R.rect.height,M=new Uint8Array(N*R.bpl+N);R.cimg=h(R.img,N,R.bpp,R.bpl,M,b,I)}}function u(v,b,I,O,R){const N=R[0],M=R[1],L=R[2],A=R[3],E=R[4],q=R[5];let F=6,P=8,Y=255;for(var j=0;j<v.length;j++){const Ee=new Uint8Array(v[j]);for(var H=Ee.length,z=0;z<H;z+=4)Y&=Ee[z+3]}const Z=Y!=255,ne=function(pe,Se,Ue,Re,xe,Pe){const K=[];for(var ie=0;ie<pe.length;ie++){const ge=new Uint8Array(pe[ie]),He=new Uint32Array(ge.buffer);var Ne;let ft=0,yt=0,$t=Se,ln=Ue,mi=Re?1:0;if(ie!=0){const Vl=Pe||Re||ie==1||K[ie-2].dispose!=0?1:2;let gi=0,fs=1e9;for(let Vn=0;Vn<Vl;Vn++){var Qe=new Uint8Array(pe[ie-1-Vn]);const Wl=new Uint32Array(pe[ie-1-Vn]);let Vt=Se,Wt=Ue,kn=-1,Wn=-1;for(let an=0;an<Ue;an++)for(let un=0;un<Se;un++)He[Ye=an*Se+un]!=Wl[Ye]&&(un<Vt&&(Vt=un),un>kn&&(kn=un),an<Wt&&(Wt=an),an>Wn&&(Wn=an));kn==-1&&(Vt=Wt=kn=Wn=0),xe&&((1&Vt)==1&&Vt--,(1&Wt)==1&&Wt--);const ds=(kn-Vt+1)*(Wn-Wt+1);ds<fs&&(fs=ds,gi=Vn,ft=Vt,yt=Wt,$t=kn-Vt+1,ln=Wn-Wt+1)}Qe=new Uint8Array(pe[ie-1-gi]),gi==1&&(K[ie-1].dispose=2),Ne=new Uint8Array($t*ln*4),t(Qe,Se,Ue,Ne,$t,ln,-ft,-yt,0),mi=t(ge,Se,Ue,Ne,$t,ln,-ft,-yt,3)?1:0,mi==1?c(ge,Se,Ue,Ne,{x:ft,y:yt,width:$t,height:ln}):t(ge,Se,Ue,Ne,$t,ln,-ft,-yt,0)}else Ne=ge.slice(0);K.push({rect:{x:ft,y:yt,width:$t,height:ln},img:Ne,blend:mi,dispose:0})}if(Re)for(ie=0;ie<K.length;ie++){if((It=K[ie]).blend==1)continue;const ge=It.rect,He=K[ie-1].rect,ft=Math.min(ge.x,He.x),yt=Math.min(ge.y,He.y),$t={x:ft,y:yt,width:Math.max(ge.x+ge.width,He.x+He.width)-ft,height:Math.max(ge.y+ge.height,He.y+He.height)-yt};K[ie-1].dispose=1,ie-1!=0&&d(pe,Se,Ue,K,ie-1,$t,xe),d(pe,Se,Ue,K,ie,$t,xe)}let on=0;if(pe.length!=1)for(var Ye=0;Ye<K.length;Ye++){var It;on+=(It=K[Ye]).rect.width*It.rect.height}return K}(v,b,I,N,M,L),de={},me=[],ke=[];if(O!=0){const Ee=[];for(z=0;z<ne.length;z++)Ee.push(ne[z].img.buffer);const pe=function(xe){let Pe=0;for(var K=0;K<xe.length;K++)Pe+=xe[K].byteLength;const ie=new Uint8Array(Pe);let Ne=0;for(K=0;K<xe.length;K++){const Qe=new Uint8Array(xe[K]),on=Qe.length;for(let Ye=0;Ye<on;Ye+=4){let It=Qe[Ye],ge=Qe[Ye+1],He=Qe[Ye+2];const ft=Qe[Ye+3];ft==0&&(It=ge=He=0),ie[Ne+Ye]=It,ie[Ne+Ye+1]=ge,ie[Ne+Ye+2]=He,ie[Ne+Ye+3]=ft}Ne+=on}return ie.buffer}(Ee),Se=y(pe,O);for(z=0;z<Se.plte.length;z++)me.push(Se.plte[z].est.rgba);let Ue=0;for(z=0;z<ne.length;z++){const Re=(ae=ne[z]).img.length;var Te=new Uint8Array(Se.inds.buffer,Ue>>2,Re>>2);ke.push(Te);const xe=new Uint8Array(Se.abuf,Ue,Re);q&&l(ae.img,ae.rect.width,ae.rect.height,me,xe,Te),ae.img.set(xe),Ue+=Re}}else for(j=0;j<ne.length;j++){var ae=ne[j];const Ee=new Uint32Array(ae.img.buffer);var fe=ae.rect.width;for(H=Ee.length,Te=new Uint8Array(H),ke.push(Te),z=0;z<H;z++){const pe=Ee[z];if(z!=0&&pe==Ee[z-1])Te[z]=Te[z-1];else if(z>fe&&pe==Ee[z-fe])Te[z]=Te[z-fe];else{let Se=de[pe];if(Se==null&&(de[pe]=Se=me.length,me.push(pe),me.length>=300))break;Te[z]=Se}}}const Ge=me.length;for(Ge<=256&&E==0&&(P=Ge<=2?1:Ge<=4?2:Ge<=16?4:8,P=Math.max(P,A)),j=0;j<ne.length;j++){(ae=ne[j]).rect.x,ae.rect.y,fe=ae.rect.width;const Ee=ae.rect.height;let pe=ae.img;new Uint32Array(pe.buffer);let Se=4*fe,Ue=4;if(Ge<=256&&E==0){Se=Math.ceil(P*fe/8);var Je=new Uint8Array(Se*Ee);const Re=ke[j];for(let xe=0;xe<Ee;xe++){z=xe*Se;const Pe=xe*fe;if(P==8)for(var $e=0;$e<fe;$e++)Je[z+$e]=Re[Pe+$e];else if(P==4)for($e=0;$e<fe;$e++)Je[z+($e>>1)]|=Re[Pe+$e]<<4-4*(1&$e);else if(P==2)for($e=0;$e<fe;$e++)Je[z+($e>>2)]|=Re[Pe+$e]<<6-2*(3&$e);else if(P==1)for($e=0;$e<fe;$e++)Je[z+($e>>3)]|=Re[Pe+$e]<<7-1*(7&$e)}pe=Je,F=3,Ue=1}else if(Z==0&&ne.length==1){Je=new Uint8Array(fe*Ee*3);const Re=fe*Ee;for(z=0;z<Re;z++){const xe=3*z,Pe=4*z;Je[xe]=pe[Pe],Je[xe+1]=pe[Pe+1],Je[xe+2]=pe[Pe+2]}pe=Je,F=2,Ue=3,Se=3*fe}ae.img=pe,ae.bpl=Se,ae.bpp=Ue}return{ctype:F,depth:P,plte:me,frames:ne}}function d(v,b,I,O,R,N,M){const L=Uint8Array,A=Uint32Array,E=new L(v[R-1]),q=new A(v[R-1]),F=R+1<v.length?new L(v[R+1]):null,P=new L(v[R]),Y=new A(P.buffer);let j=b,H=I,z=-1,Z=-1;for(let de=0;de<N.height;de++)for(let me=0;me<N.width;me++){const ke=N.x+me,Te=N.y+de,ae=Te*b+ke,fe=Y[ae];fe==0||O[R-1].dispose==0&&q[ae]==fe&&(F==null||F[4*ae+3]!=0)||(ke<j&&(j=ke),ke>z&&(z=ke),Te<H&&(H=Te),Te>Z&&(Z=Te))}z==-1&&(j=H=z=Z=0),M&&((1&j)==1&&j--,(1&H)==1&&H--),N={x:j,y:H,width:z-j+1,height:Z-H+1};const ne=O[R];ne.rect=N,ne.blend=1,ne.img=new Uint8Array(N.width*N.height*4),O[R-1].dispose==0?(t(E,b,I,ne.img,N.width,N.height,-N.x,-N.y,0),c(P,b,I,ne.img,N)):t(P,b,I,ne.img,N.width,N.height,-N.x,-N.y,0)}function c(v,b,I,O,R){t(v,b,I,O,R.width,R.height,-R.x,-R.y,2)}function h(v,b,I,O,R,N,M){const L=[];let A,E=[0,1,2,3,4];N!=-1?E=[N]:(b*O>5e5||I==1)&&(E=[0]),M&&(A={level:0});const q=Td;for(var F=0;F<E.length;F++){for(let j=0;j<b;j++)p(R,v,j,O,I,E[F]);L.push(q.deflate(R,A))}let P,Y=1e9;for(F=0;F<L.length;F++)L[F].length<Y&&(P=F,Y=L[F].length);return L[P]}function p(v,b,I,O,R,N){const M=I*O;let L=M+I;if(v[L]=N,L++,N==0)if(O<500)for(var A=0;A<O;A++)v[L+A]=b[M+A];else v.set(new Uint8Array(b.buffer,M,O),L);else if(N==1){for(A=0;A<R;A++)v[L+A]=b[M+A];for(A=R;A<O;A++)v[L+A]=b[M+A]-b[M+A-R]+256&255}else if(I==0){for(A=0;A<R;A++)v[L+A]=b[M+A];if(N==2)for(A=R;A<O;A++)v[L+A]=b[M+A];if(N==3)for(A=R;A<O;A++)v[L+A]=b[M+A]-(b[M+A-R]>>1)+256&255;if(N==4)for(A=R;A<O;A++)v[L+A]=b[M+A]-n(b[M+A-R],0,0)+256&255}else{if(N==2)for(A=0;A<O;A++)v[L+A]=b[M+A]+256-b[M+A-O]&255;if(N==3){for(A=0;A<R;A++)v[L+A]=b[M+A]+256-(b[M+A-O]>>1)&255;for(A=R;A<O;A++)v[L+A]=b[M+A]+256-(b[M+A-O]+b[M+A-R]>>1)&255}if(N==4){for(A=0;A<R;A++)v[L+A]=b[M+A]+256-n(0,b[M+A-O],0)&255;for(A=R;A<O;A++)v[L+A]=b[M+A]+256-n(b[M+A-R],b[M+A-O],b[M+A-R-O])&255}}}function y(v,b){const I=new Uint8Array(v),O=I.slice(0),R=new Uint32Array(O.buffer),N=k(O,b),M=N[0],L=N[1],A=I.length,E=new Uint8Array(A>>2);let q;if(I.length<2e7)for(var F=0;F<A;F+=4)q=_(M,P=I[F]*(1/255),Y=I[F+1]*(1/255),j=I[F+2]*(1/255),H=I[F+3]*(1/255)),E[F>>2]=q.ind,R[F>>2]=q.est.rgba;else for(F=0;F<A;F+=4){var P=I[F]*.00392156862745098,Y=I[F+1]*(1/255),j=I[F+2]*(1/255),H=I[F+3]*(1/255);for(q=M;q.left;)q=C(q.est,P,Y,j,H)<=0?q.left:q.right;E[F>>2]=q.ind,R[F>>2]=q.est.rgba}return{abuf:O.buffer,inds:E,plte:L}}function k(v,b,I){I==null&&(I=1e-4);const O=new Uint32Array(v.buffer),R={i0:0,i1:v.length,bst:null,est:null,tdst:0,left:null,right:null};R.bst=$(v,R.i0,R.i1),R.est=w(R.bst);const N=[R];for(;N.length<b;){let L=0,A=0;for(var M=0;M<N.length;M++)N[M].est.L>L&&(L=N[M].est.L,A=M);if(L<I)break;const E=N[A],q=T(v,O,E.i0,E.i1,E.est.e,E.est.eMq255);if(E.i0>=q||E.i1<=q){E.est.L=0;continue}const F={i0:E.i0,i1:q,bst:null,est:null,tdst:0,left:null,right:null};F.bst=$(v,F.i0,F.i1),F.est=w(F.bst);const P={i0:q,i1:E.i1,bst:null,est:null,tdst:0,left:null,right:null};for(P.bst={R:[],m:[],N:E.bst.N-F.bst.N},M=0;M<16;M++)P.bst.R[M]=E.bst.R[M]-F.bst.R[M];for(M=0;M<4;M++)P.bst.m[M]=E.bst.m[M]-F.bst.m[M];P.est=w(P.bst),E.left=F,E.right=P,N[A]=F,N.push(P)}for(N.sort((L,A)=>A.bst.N-L.bst.N),M=0;M<N.length;M++)N[M].ind=M;return[R,N]}function _(v,b,I,O,R){if(v.left==null)return v.tdst=function(F,P,Y,j,H){const z=P-F[0],Z=Y-F[1],ne=j-F[2],de=H-F[3];return z*z+Z*Z+ne*ne+de*de}(v.est.q,b,I,O,R),v;const N=C(v.est,b,I,O,R);let M=v.left,L=v.right;N>0&&(M=v.right,L=v.left);const A=_(M,b,I,O,R);if(A.tdst<=N*N)return A;const E=_(L,b,I,O,R);return E.tdst<A.tdst?E:A}function C(v,b,I,O,R){const{e:N}=v;return N[0]*b+N[1]*I+N[2]*O+N[3]*R-v.eMq}function T(v,b,I,O,R,N){for(O-=4;I<O;){for(;S(v,I,R)<=N;)I+=4;for(;S(v,O,R)>N;)O-=4;if(I>=O)break;const M=b[I>>2];b[I>>2]=b[O>>2],b[O>>2]=M,I+=4,O-=4}for(;S(v,I,R)>N;)I-=4;return I+4}function S(v,b,I){return v[b]*I[0]+v[b+1]*I[1]+v[b+2]*I[2]+v[b+3]*I[3]}function $(v,b,I){const O=[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],R=[0,0,0,0],N=I-b>>2;for(let M=b;M<I;M+=4){const L=v[M]*.00392156862745098,A=v[M+1]*(1/255),E=v[M+2]*(1/255),q=v[M+3]*(1/255);R[0]+=L,R[1]+=A,R[2]+=E,R[3]+=q,O[0]+=L*L,O[1]+=L*A,O[2]+=L*E,O[3]+=L*q,O[5]+=A*A,O[6]+=A*E,O[7]+=A*q,O[10]+=E*E,O[11]+=E*q,O[15]+=q*q}return O[4]=O[1],O[8]=O[2],O[9]=O[6],O[12]=O[3],O[13]=O[7],O[14]=O[11],{R:O,m:R,N}}function w(v){const{R:b}=v,{m:I}=v,{N:O}=v,R=I[0],N=I[1],M=I[2],L=I[3],A=O==0?0:1/O,E=[b[0]-R*R*A,b[1]-R*N*A,b[2]-R*M*A,b[3]-R*L*A,b[4]-N*R*A,b[5]-N*N*A,b[6]-N*M*A,b[7]-N*L*A,b[8]-M*R*A,b[9]-M*N*A,b[10]-M*M*A,b[11]-M*L*A,b[12]-L*R*A,b[13]-L*N*A,b[14]-L*M*A,b[15]-L*L*A],q=E,F=x;let P=[Math.random(),Math.random(),Math.random(),Math.random()],Y=0,j=0;if(O!=0)for(let z=0;z<16&&(P=F.multVec(q,P),j=Math.sqrt(F.dot(P,P)),P=F.sml(1/j,P),!(z!=0&&Math.abs(j-Y)<1e-9));z++)Y=j;const H=[R*A,N*A,M*A,L*A];return{Cov:E,q:H,e:P,L:Y,eMq255:F.dot(F.sml(255,H),P),eMq:F.dot(P,H),rgba:(Math.round(255*H[3])<<24|Math.round(255*H[2])<<16|Math.round(255*H[1])<<8|Math.round(255*H[0])<<0)>>>0}}var x={multVec:(v,b)=>[v[0]*b[0]+v[1]*b[1]+v[2]*b[2]+v[3]*b[3],v[4]*b[0]+v[5]*b[1]+v[6]*b[2]+v[7]*b[3],v[8]*b[0]+v[9]*b[1]+v[10]*b[2]+v[11]*b[3],v[12]*b[0]+v[13]*b[1]+v[14]*b[2]+v[15]*b[3]],dot:(v,b)=>v[0]*b[0]+v[1]*b[1]+v[2]*b[2]+v[3]*b[3],sml:(v,b)=>[v*b[0],v*b[1],v*b[2],v*b[3]]};kt.encode=function(b,I,O,R,N,M,L){R==null&&(R=0),L==null&&(L=!1);const A=u(b,I,O,R,[!1,!1,!1,0,L,!1]);return f(A,-1),a(A,I,O,N,M)},kt.encodeLL=function(b,I,O,R,N,M,L,A){const E={ctype:0+(R==1?0:2)+(N==0?0:4),depth:M,frames:[]},q=(R+N)*M,F=q*I;for(let P=0;P<b.length;P++)E.frames.push({rect:{x:0,y:0,width:I,height:O},img:new Uint8Array(b[P]),blend:0,dispose:1,bpp:Math.ceil(q/8),bpl:Math.ceil(F/8)});return f(E,0,!0),a(E,I,O,L,A)},kt.encode.compress=u,kt.encode.dither=l,kt.quantize=y,kt.quantize.getKDtree=k,kt.quantize.getNearest=_})();const Ml={toArrayBuffer(t,e){const n=t.width,r=t.height,i=n<<2,s=t.getContext("2d").getImageData(0,0,n,r),o=new Uint32Array(s.data.buffer),l=(32*n+31)/32<<2,a=l*r,f=122+a,u=new ArrayBuffer(f),d=new DataView(u),c=1<<20;let h,p,y,k,_=c,C=0,T=0,S=0;function $(v){d.setUint16(T,v,!0),T+=2}function w(v){d.setUint32(T,v,!0),T+=4}function x(v){T+=v}$(19778),w(f),x(4),w(122),w(108),w(n),w(-r>>>0),$(1),$(32),w(3),w(a),w(2835),w(2835),x(8),w(16711680),w(65280),w(255),w(4278190080),w(1466527264),function v(){for(;C<r&&_>0;){for(k=122+C*l,h=0;h<i;)_--,p=o[S++],y=p>>>24,d.setUint32(k+h,p<<8|y),h+=4;C++}S<o.length?(_=c,setTimeout(v,Ml._dly)):e(u)}()},toBlob(t,e){this.toArrayBuffer(t,n=>{e(new Blob([n],{type:"image/bmp"}))})},_dly:9};var mt={CHROME:"CHROME",FIREFOX:"FIREFOX",DESKTOP_SAFARI:"DESKTOP_SAFARI",IE:"IE",IOS:"IOS",ETC:"ETC"},Cd={[mt.CHROME]:16384,[mt.FIREFOX]:11180,[mt.DESKTOP_SAFARI]:16384,[mt.IE]:8192,[mt.IOS]:4096,[mt.ETC]:8192};const as=typeof window<"u",Rl=typeof WorkerGlobalScope<"u"&&self instanceof WorkerGlobalScope,ei=as&&window.cordova&&window.cordova.require&&window.cordova.require("cordova/modulemapper"),xd=(as||Rl)&&(ei&&ei.getOriginalSymbol(window,"File")||typeof File<"u"&&File),Nl=(as||Rl)&&(ei&&ei.getOriginalSymbol(window,"FileReader")||typeof FileReader<"u"&&FileReader);function us(t,e,n=Date.now()){return new Promise(r=>{const i=t.split(","),s=i[0].match(/:(.*?);/)[1],o=globalThis.atob(i[1]);let l=o.length;const a=new Uint8Array(l);for(;l--;)a[l]=o.charCodeAt(l);const f=new Blob([a],{type:s});f.name=e,f.lastModified=n,r(f)})}function ql(t){return new Promise((e,n)=>{const r=new Nl;r.onload=()=>e(r.result),r.onerror=i=>n(i),r.readAsDataURL(t)})}function Fl(t){return new Promise((e,n)=>{const r=new Image;r.onload=()=>e(r),r.onerror=i=>n(i),r.src=t})}function hn(){if(hn.cachedResult!==void 0)return hn.cachedResult;let t=mt.ETC;const{userAgent:e}=navigator;return/Chrom(e|ium)/i.test(e)?t=mt.CHROME:/iP(ad|od|hone)/i.test(e)&&/WebKit/i.test(e)?t=mt.IOS:/Safari/i.test(e)?t=mt.DESKTOP_SAFARI:/Firefox/i.test(e)?t=mt.FIREFOX:(/MSIE/i.test(e)||document.documentMode)&&(t=mt.IE),hn.cachedResult=t,hn.cachedResult}function Ll(t,e){const n=hn(),r=Cd[n];let i=t,s=e,o=i*s;const l=i>s?s/i:i/s;for(;o>r*r;){const a=(r+i)/2,f=(r+s)/2;a<f?(s=f,i=f*l):(s=a*l,i=a),o=i*s}return{width:i,height:s}}function hi(t,e){let n,r;try{if(n=new OffscreenCanvas(t,e),r=n.getContext("2d"),r===null)throw new Error("getContext of OffscreenCanvas returns null")}catch{n=document.createElement("canvas"),r=n.getContext("2d")}return n.width=t,n.height=e,[n,r]}function Pl(t,e){const{width:n,height:r}=Ll(t.width,t.height),[i,s]=hi(n,r);return e&&/jpe?g/.test(e)&&(s.fillStyle="white",s.fillRect(0,0,i.width,i.height)),s.drawImage(t,0,0,i.width,i.height),i}function $r(){return $r.cachedResult!==void 0||($r.cachedResult=["iPad Simulator","iPhone Simulator","iPod Simulator","iPad","iPhone","iPod"].includes(navigator.platform)||navigator.userAgent.includes("Mac")&&typeof document<"u"&&"ontouchend"in document),$r.cachedResult}function ti(t,e={}){return new Promise(function(n,r){let i,s;var o=function(){try{return s=Pl(i,e.fileType||t.type),n([i,s])}catch(a){return r(a)}},l=function(a){try{var f=function(u){try{throw u}catch(d){return r(d)}};try{let u;return ql(t).then(function(d){try{return u=d,Fl(u).then(function(c){try{return i=c,function(){try{return o()}catch(h){return r(h)}}()}catch(h){return f(h)}},f)}catch(c){return f(c)}},f)}catch(u){f(u)}}catch(u){return r(u)}};try{if($r()||[mt.DESKTOP_SAFARI,mt.MOBILE_SAFARI].includes(hn()))throw new Error("Skip createImageBitmap on IOS and Safari");return createImageBitmap(t).then(function(a){try{return i=a,o()}catch{return l()}},l)}catch{l()}})}function ni(t,e,n,r,i=1){return new Promise(function(s,o){let l;if(e==="image/png"){let d,c,h;return d=t.getContext("2d"),{data:c}=d.getImageData(0,0,t.width,t.height),h=kt.encode([c.buffer],t.width,t.height,4096*i),l=new Blob([h],{type:e}),l.name=n,l.lastModified=r,a.call(this)}{let d=function(){return a.call(this)};var f=d;if(e==="image/bmp")return new Promise(c=>Ml.toBlob(t,c)).then(function(c){try{return l=c,l.name=n,l.lastModified=r,d.call(this)}catch(h){return o(h)}}.bind(this),o);{let c=function(){return d.call(this)};var u=c;if(typeof OffscreenCanvas=="function"&&t instanceof OffscreenCanvas)return t.convertToBlob({type:e,quality:i}).then(function(h){try{return l=h,l.name=n,l.lastModified=r,c.call(this)}catch(p){return o(p)}}.bind(this),o);{let h;return h=t.toDataURL(e,i),us(h,n,r).then(function(p){try{return l=p,c.call(this)}catch(y){return o(y)}}.bind(this),o)}}}function a(){return s(l)}})}function Ot(t){t.width=0,t.height=0}function Sn(){return new Promise(function(t,e){let n,r,i,s;return Sn.cachedResult!==void 0?t(Sn.cachedResult):us("data:image/jpeg;base64,/9j/4QAiRXhpZgAATU0AKgAAAAgAAQESAAMAAAABAAYAAAAAAAD/2wCEAAEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAf/AABEIAAEAAgMBEQACEQEDEQH/xABKAAEAAAAAAAAAAAAAAAAAAAALEAEAAAAAAAAAAAAAAAAAAAAAAQEAAAAAAAAAAAAAAAAAAAAAEQEAAAAAAAAAAAAAAAAAAAAA/9oADAMBAAIRAxEAPwA/8H//2Q==","test.jpg",Date.now()).then(function(o){try{return n=o,ti(n).then(function(l){try{return r=l[1],ni(r,n.type,n.name,n.lastModified).then(function(a){try{return i=a,Ot(r),ti(i).then(function(f){try{return s=f[0],Sn.cachedResult=s.width===1&&s.height===2,t(Sn.cachedResult)}catch(u){return e(u)}},e)}catch(f){return e(f)}},e)}catch(a){return e(a)}},e)}catch(l){return e(l)}},e)})}function Ul(t){return new Promise((e,n)=>{const r=new Nl;r.onload=i=>{const s=new DataView(i.target.result);if(s.getUint16(0,!1)!=65496)return e(-2);const o=s.byteLength;let l=2;for(;l<o;){if(s.getUint16(l+2,!1)<=8)return e(-1);const a=s.getUint16(l,!1);if(l+=2,a==65505){if(s.getUint32(l+=2,!1)!=1165519206)return e(-1);const f=s.getUint16(l+=6,!1)==18761;l+=s.getUint32(l+4,f);const u=s.getUint16(l,f);l+=2;for(let d=0;d<u;d++)if(s.getUint16(l+12*d,f)==274)return e(s.getUint16(l+12*d+8,f))}else{if((65280&a)!=65280)break;l+=s.getUint16(l,!1)}}return e(-1)},r.onerror=i=>n(i),r.readAsArrayBuffer(t)})}function zl(t,e){const{width:n}=t,{height:r}=t,{maxWidthOrHeight:i}=e;let s,o=t;return isFinite(i)&&(n>i||r>i)&&([o,s]=hi(n,r),n>r?(o.width=i,o.height=r/n*i):(o.width=n/r*i,o.height=i),s.drawImage(t,0,0,o.width,o.height),Ot(t)),o}function Zl(t,e){const{width:n}=t,{height:r}=t,[i,s]=hi(n,r);switch(e>4&&e<9?(i.width=r,i.height=n):(i.width=n,i.height=r),e){case 2:s.transform(-1,0,0,1,n,0);break;case 3:s.transform(-1,0,0,-1,n,r);break;case 4:s.transform(1,0,0,-1,0,r);break;case 5:s.transform(0,1,1,0,0,0);break;case 6:s.transform(0,1,-1,0,r,0);break;case 7:s.transform(0,-1,-1,0,r,n);break;case 8:s.transform(0,-1,1,0,0,n)}return s.drawImage(t,0,0,n,r),Ot(t),i}function io(t,e,n=0){return new Promise(function(r,i){let s,o,l,a,f,u,d,c,h,p,y,k,_,C,T,S,$,w,x,v;function b(O=5){if(e.signal&&e.signal.aborted)throw e.signal.reason;s+=O,e.onProgress(Math.min(s,100))}function I(O){if(e.signal&&e.signal.aborted)throw e.signal.reason;s=Math.min(Math.max(O,s),100),e.onProgress(s)}return s=n,o=e.maxIteration||10,l=1024*e.maxSizeMB*1024,b(),ti(t,e).then(function(O){try{return[,a]=O,b(),f=zl(a,e),b(),new Promise(function(R,N){var M;if(!(M=e.exifOrientation))return Ul(t).then(function(A){try{return M=A,L.call(this)}catch(E){return N(E)}}.bind(this),N);function L(){return R(M)}return L.call(this)}).then(function(R){try{return u=R,b(),Sn().then(function(N){try{return d=N?f:Zl(f,u),b(),c=e.initialQuality||1,h=e.fileType||t.type,ni(d,h,t.name,t.lastModified,c).then(function(M){try{{let q=function(){if(o--&&(T>l||T>_)){let P,Y;return P=v?.95*x.width:x.width,Y=v?.95*x.height:x.height,[$,w]=hi(P,Y),w.drawImage(x,0,0,P,Y),c*=h==="image/png"?.85:.95,ni($,h,t.name,t.lastModified,c).then(function(j){try{return S=j,Ot(x),x=$,T=S.size,I(Math.min(99,Math.floor((C-T)/(C-l)*100))),q}catch(H){return i(H)}},i)}return[1]},F=function(){return Ot(x),Ot($),Ot(f),Ot(d),Ot(a),I(100),r(S)};var A=q,E=F;if(p=M,b(),y=p.size>l,k=p.size>t.size,!y&&!k)return I(100),r(p);var L;return _=t.size,C=p.size,T=C,x=d,v=!e.alwaysKeepResolution&&y,(L=function(P){for(;P;){if(P.then)return void P.then(L,i);try{if(P.pop){if(P.length)return P.pop()?F.call(this):P;P=q}else P=P.call(this)}catch(Y){return i(Y)}}}.bind(this))(q)}}catch(q){return i(q)}}.bind(this),i)}catch(M){return i(M)}}.bind(this),i)}catch(N){return i(N)}}.bind(this),i)}catch(R){return i(R)}}.bind(this),i)})}const Dd=`
-let scriptImported = false
-self.addEventListener('message', async (e) => {
-  const { file, id, imageCompressionLibUrl, options } = e.data
-  options.onProgress = (progress) => self.postMessage({ progress, id })
-  try {
-    if (!scriptImported) {
-      // console.log('[worker] importScripts', imageCompressionLibUrl)
-      self.importScripts(imageCompressionLibUrl)
-      scriptImported = true
-    }
-    // console.log('[worker] self', self)
-    const compressedFile = await imageCompression(file, options)
-    self.postMessage({ file: compressedFile, id })
-  } catch (e) {
-    // console.error('[worker] error', e)
-    self.postMessage({ error: e.message + '\\n' + e.stack, id })
-  }
-})
-`;let ki;function Sd(t,e){return new Promise((n,r)=>{ki||(ki=function(o){const l=[];return typeof o=="function"?l.push(`(${o})()`):l.push(o),URL.createObjectURL(new Blob(l))}(Dd));const i=new Worker(ki);i.addEventListener("message",function(o){if(e.signal&&e.signal.aborted)i.terminate();else if(o.data.progress===void 0){if(o.data.error)return r(new Error(o.data.error)),void i.terminate();n(o.data.file),i.terminate()}else e.onProgress(o.data.progress)}),i.addEventListener("error",r),e.signal&&e.signal.addEventListener("abort",()=>{r(e.signal.reason),i.terminate()}),i.postMessage({file:t,imageCompressionLibUrl:e.libURL,options:{...e,onProgress:void 0,signal:void 0}})})}function it(t,e){return new Promise(function(n,r){let i,s,o,l,a,f;if(i={...e},o=0,{onProgress:l}=i,i.maxSizeMB=i.maxSizeMB||Number.POSITIVE_INFINITY,a=typeof i.useWebWorker!="boolean"||i.useWebWorker,delete i.useWebWorker,i.onProgress=h=>{o=h,typeof l=="function"&&l(o)},!(t instanceof Blob||t instanceof xd))return r(new Error("The file given is not an instance of Blob or File"));if(!/^image/.test(t.type))return r(new Error("The file given is not an image"));if(f=typeof WorkerGlobalScope<"u"&&self instanceof WorkerGlobalScope,!a||typeof Worker!="function"||f)return io(t,i).then(function(h){try{return s=h,c.call(this)}catch(p){return r(p)}}.bind(this),r);var u=function(){try{return c.call(this)}catch(h){return r(h)}}.bind(this),d=function(h){try{return io(t,i).then(function(p){try{return s=p,u()}catch(y){return r(y)}},r)}catch(p){return r(p)}};try{return i.libURL=i.libURL||"https://cdn.jsdelivr.net/npm/browser-image-compression@2.0.2/dist/browser-image-compression.js",Sd(t,i).then(function(h){try{return s=h,u()}catch{return d()}},d)}catch{d()}function c(){try{s.name=t.name,s.lastModified=t.lastModified}catch{}try{i.preserveExif&&t.type==="image/jpeg"&&(!i.fileType||i.fileType&&i.fileType===t.type)&&(s=Ol(t,s))}catch{}return n(s)}})}it.getDataUrlFromFile=ql,it.getFilefromDataUrl=us,it.loadImage=Fl,it.drawImageInCanvas=Pl,it.drawFileInCanvas=ti,it.canvasToFile=ni,it.getExifOrientation=Ul,it.handleMaxWidthOrHeight=zl,it.followExifOrientation=Zl,it.cleanupCanvasMemory=Ot,it.isAutoOrientationInBrowser=Sn,it.approximateBelowMaximumCanvasSizeOfBrowser=Ll,it.copyExifWithoutOrientation=Ol,it.getBrowserName=hn,it.version="2.0.2";const Rn=Bt({}),Nn=Bt([]),pi=Bt(null);function Id(){return new Promise(t=>{pi.set(t)})}function $d(){pi.update(t=>(t!==null&&t(null),null))}function Od(t){pi.update(e=>(e!==null&&e(t),null))}async function Md(t,e){const n=Qi(),r=await it(t,{maxSizeMB:.8,onProgress:e});if(r.size>=ac)throw Error("File is too large (max 15mb)");Nn.update(i=>(i.push({uuid:n,name:t.name,size:t.size,blob:r}),i)),jl(r,n)}async function jl(t,e){const n=new FileReader;await new Promise((i,s)=>{n.onload=i,n.onerror=s,n.onabort=s,n.readAsDataURL(t)});const r=n.result;Rn.update(i=>(i[e]=r,i))}function Rd(t){let e,n;return{c(){e=ce("svg"),n=ce("path"),m(n,"d","M180-120q-24 0-42-18t-18-42v-600q0-24 18-42t42-18h600q24 0 42 18t18 42v600q0 24-18 42t-42 18H180Zm0-60h600v-600H180v600Zm56-97h489L578-473 446-302l-93-127-117 152Zm-56 97v-600 600Z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 -960 960 960"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#666666")},m(r,i){W(r,e,i),g(e,n)},p:Q,i:Q,o:Q,d(r){r&&V(e)}}}let Nd=class extends be{constructor(e){super(),ye(this,e,null,Rd,ve,{})}};function so(t,e,n){const r=t.slice();return r[7]=e[n],r}function qd(t){let e,n;return e=new Nd({}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p:Q,i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function Fd(t){let e,n,r;return{c(){e=D("img"),m(e,"class","image svelte-n65hdz"),m(e,"data-fit",n=t[0].image.fit),qn(e.src,r=t[2])||m(e,"src",r),m(e,"alt","Uploaded Content")},m(i,s){W(i,e,s)},p(i,s){s&1&&n!==(n=i[0].image.fit)&&m(e,"data-fit",n),s&4&&!qn(e.src,r=i[2])&&m(e,"src",r)},i:Q,o:Q,d(i){i&&V(e)}}}function oo(t){let e,n=[],r=new Map,i=t[0].answers;const s=o=>o[7].id;for(let o=0;o<i.length;o+=1){let l=so(t,i,o),a=s(l);r.set(a,n[o]=lo(a,l))}return{c(){e=D("div");for(let o=0;o<n.length;o+=1)n[o].c();m(e,"class","answers svelte-n65hdz")},m(o,l){W(o,e,l);for(let a=0;a<n.length;a+=1)n[a]&&n[a].m(e,null)},p(o,l){if(l&1){i=o[0].answers;for(let a=0;a<n.length;a+=1)n[a].r();n=_r(n,l,s,1,o,i,r,e,ya,lo,null,so);for(let a=0;a<n.length;a+=1)n[a].a()}},d(o){o&&V(e);for(let l=0;l<n.length;l+=1)n[l].d()}}}function lo(t,e){let n,r,i,s=Q;return{key:t,first:null,c(){n=D("p"),m(n,"class","answer svelte-n65hdz"),m(n,"data-correct",r=e[7].correct),this.first=n},m(o,l){W(o,n,l)},p(o,l){e=o,l&1&&r!==(r=e[7].correct)&&m(n,"data-correct",r)},r(){i=n.getBoundingClientRect()},f(){gr(n),s()},a(){s(),s=mr(n,i,wr,{duration:250})},d(o){o&&V(n)}}}function Ld(t){let e,n,r,i,s,o,l,a=t[1]+1+"",f,u,d=t[0].text+"",c,h,p,y,k;const _=[Fd,qd],C=[];function T($,w){return $[0].image!==null&&$[2]!==null?0:1}r=T(t),i=C[r]=_[r](t);let S=(t[0].ty==rt.Single||t[0].ty==rt.Multiple)&&oo(t);return{c(){e=D("div"),n=D("div"),i.c(),s=U(),o=D("p"),l=D("span"),f=te(a),u=U(),c=te(d),h=U(),S&&S.c(),m(n,"class","image-wrapper svelte-n65hdz"),m(l,"class","question__index svelte-n65hdz"),m(o,"class","text svelte-n65hdz"),m(e,"class","question svelte-n65hdz"),Or(e,"question--active",t[3]!==null&&t[3].id===t[0].id)},m($,w){W($,e,w),g(e,n),C[r].m(n,null),g(e,s),g(e,o),g(o,l),g(l,f),g(o,u),g(o,c),g(e,h),S&&S.m(e,null),p=!0,y||(k=[J(e,"click",t[4]),J(e,"keypress",t[5])],y=!0)},p($,[w]){let x=r;r=T($),r===x?C[r].p($,w):(qe(),G(C[x],1,1,()=>{C[x]=null}),Fe(),i=C[r],i?i.p($,w):(i=C[r]=_[r]($),i.c()),B(i,1),i.m(n,null)),(!p||w&2)&&a!==(a=$[1]+1+"")&&Ie(f,a),(!p||w&1)&&d!==(d=$[0].text+"")&&Ie(c,d),$[0].ty==rt.Single||$[0].ty==rt.Multiple?S?S.p($,w):(S=oo($),S.c(),S.m(e,null)):S&&(S.d(1),S=null),(!p||w&9)&&Or(e,"question--active",$[3]!==null&&$[3].id===$[0].id)},i($){p||(B(i),p=!0)},o($){G(i),p=!1},d($){$&&V(e),C[r].d(),S&&S.d(),y=!1,je(k)}}}function Pd(t,e,n){let r,i;lt(t,Rn,u=>n(6,r=u)),lt(t,sn,u=>n(3,i=u));let{question:s}=e,{index:o}=e,l=null;function a(){sn.set(s)}function f(u){console.log(u.key),(u.key==="Return"||u.key=="Enter")&&a()}return t.$$set=u=>{"question"in u&&n(0,s=u.question),"index"in u&&n(1,o=u.index)},t.$$.update=()=>{if(t.$$.dirty&65&&s.image!==null){let u=r[s.image.uuid];u!==void 0?n(2,l=u):n(2,l=null)}},[s,o,l,i,a,f,r]}class Ud extends be{constructor(e){super(),ye(this,e,Pd,Ld,ve,{question:0,index:1})}}function zd(t){let e,n,r;return{c(){e=ce("svg"),n=ce("path"),r=ce("path"),m(n,"d","M0 0h24v24H0z"),m(n,"fill","none"),m(r,"d","M9 16h6v-6h4l-7-7-7 7h4zm-4 2h14v2H5z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 0 24 24"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(i,s){W(i,e,s),g(e,n),g(e,r)},p:Q,i:Q,o:Q,d(i){i&&V(e)}}}class Hl extends be{constructor(e){super(),ye(this,e,null,zd,ve,{})}}function Zd(t){let e,n,r;return{c(){e=ce("svg"),n=ce("path"),r=ce("path"),m(n,"d","M0 0h24v24H0z"),m(n,"fill","none"),m(r,"d","M19 9h-4V3H9v6H5l7 7 7-7zM5 18v2h14v-2H5z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 0 24 24"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(i,s){W(i,e,s),g(e,n),g(e,r)},p:Q,i:Q,o:Q,d(i){i&&V(e)}}}class jd extends be{constructor(e){super(),ye(this,e,null,Zd,ve,{})}}function Hd(t){let e,n,r;return{c(){e=ce("svg"),n=ce("path"),r=ce("path"),m(n,"d","M0 0h24v24H0z"),m(n,"fill","none"),m(r,"d","M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 0 24 24"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(i,s){W(i,e,s),g(e,n),g(e,r)},p:Q,i:Q,o:Q,d(i){i&&V(e)}}}class Bd extends be{constructor(e){super(),ye(this,e,null,Hd,ve,{})}}function Vd(t){let e,n;return{c(){e=ce("svg"),n=ce("path"),m(n,"d","M396-524 173-747q-9-9-9-21.5t9-21.5q8.311-8 21.156-8Q207-798 215-790l224 223-43 43Zm216 374q-12.75 0-21.375-8.675-8.625-8.676-8.625-21.5 0-12.825 8.625-21.325T612-210h95L523-393l42-43 186 184v-97q0-12.75 8.675-21.375 8.676-8.625 21.5-8.625 12.825 0 21.325 8.625T811-349v169q0 12.75-8.625 21.375T781-150H612Zm-440-21.5q-9-8.5-9-21t9-21.5l537-537h-97q-12.75 0-21.375-8.675-8.625-8.676-8.625-21.5 0-12.825 8.625-21.325T612-811h169q12.75 0 21.375 8.625T811-781v169q0 12.75-8.675 21.375-8.676 8.625-21.5 8.625-12.825 0-21.325-8.625T751-612v-96L214-171q-8.067 8-20.533 8Q181-163 172-171.5Z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 -960 960 960"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(r,i){W(r,e,i),g(e,n)},p:Q,i:Q,o:Q,d(r){r&&V(e)}}}class Bl extends be{constructor(e){super(),ye(this,e,null,Vd,ve,{})}}const Wd=De.object({name:De.string(),text:De.string(),max_players:De.number(),filtering:De.string(),questions:De.array(Su).min(1),images:De.array(De.object({uuid:De.string().uuid(),name:De.string(),size:De.number(),type:De.string(),data:De.array(De.number())}))});async function Gd(t){const e=await t.blob.arrayBuffer();return{uuid:t.uuid,name:t.name,size:t.size,type:t.blob.type,data:Array.from(new Uint8Array(e))}}async function Qd(t,e){const n=await Promise.all(e.map(Gd)),r={...t,images:n},i=JSON.stringify(r);return new Blob([i],{type:"application/json"})}async function Yd(t){const e=new FileReader;await new Promise((s,o)=>{e.onload=s,e.onerror=o,e.onabort=o,e.readAsText(t)});const n=e.result,r=JSON.parse(n),i=Wd.parse(r);for(const{data:s,uuid:o,name:l,type:a,size:f}of i.images){const u=new Uint8Array(s),d=new Blob([u.buffer],{type:a});Nn.update(c=>(c.push({uuid:o,name:l,size:f,blob:d}),c.filter(h=>h.uuid!==o))),jl(d,o)}return i.questions.forEach(s=>{s.id=Qi()}),{name:i.name,text:i.text,max_players:i.max_players,filtering:i.filtering,questions:i.questions}}function Xd(t=void 0){const e=document.createElement("input");return e.type="file",e.hidden=!0,t!==void 0&&(e.accept=t),document.body.appendChild(e),new Promise((n,r)=>{e.onchange=()=>{if(e.files===null)return r();const i=e.files.item(0);n(i)},e.oncancel=()=>n(null),e.click(),document.body.removeChild(e)})}function Kd(t=void 0){const e=document.createElement("input");return e.type="file",e.multiple=!0,e.hidden=!0,t!==void 0&&(e.accept=t),new Promise((n,r)=>{document.body.appendChild(e),e.onchange=()=>{if(e.files===null)return r();n(e.files)},e.oncancel=()=>n(null),e.click(),document.body.removeChild(e)})}function Jd(t,e){const n=document.createElement("a");n.download=t.replace(/[ ^/]/g,"_"),n.href=URL.createObjectURL(e),n.dataset.downloadurl=["application/json",n.download,n.href].join(":"),n.style.display="none",document.body.appendChild(n),n.click(),document.body.removeChild(n)}function e1(t){let e,n;return{c(){e=ce("svg"),n=ce("path"),m(n,"d","M546-80H414q-11 0-19.5-7T384-105l-16-101q-19-7-40-19t-37-25l-93 43q-11 5-22 1.5T159-220L93-337q-6-10-3-21t12-18l86-63q-2-9-2.5-20.5T185-480q0-9 .5-20.5T188-521l-86-63q-9-7-12-18t3-21l66-117q6-11 17-14.5t22 1.5l93 43q16-13 37-25t40-18l16-102q2-11 10.5-18t19.5-7h132q11 0 19.5 7t10.5 18l16 101q19 7 40.5 18.5T669-710l93-43q11-5 22-1.5t17 14.5l66 116q6 10 3.5 21.5T858-584l-86 61q2 10 2.5 21.5t.5 21.5q0 10-.5 21t-2.5 21l86 62q9 7 12 18t-3 21l-66 117q-6 11-17 14.5t-22-1.5l-93-43q-16 13-36.5 25.5T592-206l-16 101q-2 11-10.5 18T546-80Zm-66-270q54 0 92-38t38-92q0-54-38-92t-92-38q-54 0-92 38t-38 92q0 54 38 92t92 38Z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 -960 960 960"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(r,i){W(r,e,i),g(e,n)},p:Q,i:Q,o:Q,d(r){r&&V(e)}}}class cs extends be{constructor(e){super(),ye(this,e,null,e1,ve,{})}}function t1(t){let e,n,r;return{c(){e=ce("svg"),n=ce("path"),r=ce("path"),m(n,"d","M0 0h24v24H0z"),m(n,"fill","none"),m(r,"d","M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 0 24 24"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(i,s){W(i,e,s),g(e,n),g(e,r)},p:Q,i:Q,o:Q,d(i){i&&V(e)}}}class Bn extends be{constructor(e){super(),ye(this,e,null,t1,ve,{})}}function n1(t){let e,n,r,i,s,o,l,a,f,u,d,c,h,p,y,k,_,C,T,S,$,w,x,v,b,I,O,R,N,M,L,A,E,q,F,P,Y,j,H,z,Z,ne,de,me,ke,Te;return i=new Bn({}),{c(){e=D("div"),n=D("div"),r=D("button"),le(i.$$.fragment),s=te(`
-      Close`),o=U(),l=D("label"),a=D("span"),a.textContent="Title",f=U(),u=D("p"),u.innerHTML='Give your quiz a title <span class="optional svelte-hjxphc">Optional</span>',d=U(),c=D("input"),h=U(),p=D("label"),y=D("span"),y.textContent="Description",k=U(),_=D("p"),_.innerHTML='Description of your Quiz <span class="optional svelte-hjxphc">Optional</span>',C=U(),T=D("textarea"),S=U(),$=D("label"),w=D("span"),w.textContent="Max Players",x=U(),v=D("p"),v.textContent="Maximum number of players allowed to join this quiz",b=U(),I=D("input"),O=U(),R=D("label"),N=D("span"),N.textContent="Name Filtering",M=U(),L=D("p"),L.textContent=`Level of filtering on profane/inappropriate naming. Its recommended that
-        you leave this on Medium or High`,A=U(),E=D("select"),q=D("option"),F=te("None: Don't filter names"),P=D("option"),Y=te("Low: Filter out more severe names"),j=D("option"),H=te(`Medium: Filter out anything thats not mild
-        `),z=D("option"),Z=te("High: Filter out as much as possible"),m(r,"class","btn btn--icon btn--surface"),m(a,"class","field__name svelte-hjxphc"),m(u,"class","field__desc svelte-hjxphc"),m(c,"class","input svelte-hjxphc"),m(c,"type","text"),m(c,"maxlength",Hu),m(l,"class","field svelte-hjxphc"),m(y,"class","field__name svelte-hjxphc"),m(_,"class","field__desc svelte-hjxphc"),m(T,"class","input input--desc svelte-hjxphc"),m(T,"name",""),m(T,"id",""),m(T,"cols","30"),m(T,"rows","5"),m(T,"maxlength",Bu),m(p,"class","field svelte-hjxphc"),m(w,"class","field__name svelte-hjxphc"),m(v,"class","field__desc svelte-hjxphc"),m(I,"class","input svelte-hjxphc"),m(I,"type","number"),m(I,"min",lc),m(I,"max",ll),m($,"class","field svelte-hjxphc"),m(N,"class","field__name svelte-hjxphc"),m(L,"class","field__desc svelte-hjxphc"),q.__value=xn.None,q.value=q.__value,P.__value=xn.Low,P.value=P.__value,j.__value=xn.Medium,j.value=j.__value,z.__value=xn.High,z.value=z.__value,m(E,"class","input svelte-hjxphc"),t[1].filtering===void 0&&Le(()=>t[6].call(E)),m(R,"class","field svelte-hjxphc"),m(n,"class","dialog svelte-hjxphc"),m(e,"class","floating-wrapper svelte-hjxphc")},m(ae,fe){W(ae,e,fe),g(e,n),g(n,r),se(i,r,null),g(r,s),g(n,o),g(n,l),g(l,a),g(l,f),g(l,u),g(l,d),g(l,c),Ze(c,t[1].name),g(n,h),g(n,p),g(p,y),g(p,k),g(p,_),g(p,C),g(p,T),Ze(T,t[1].text),g(n,S),g(n,$),g($,w),g($,x),g($,v),g($,b),g($,I),Ze(I,t[1].max_players),g(n,O),g(n,R),g(R,N),g(R,M),g(R,L),g(R,A),g(R,E),g(E,q),g(q,F),g(E,P),g(P,Y),g(E,j),g(j,H),g(E,z),g(z,Z),Fn(E,t[1].filtering,!0),me=!0,ke||(Te=[J(r,"click",t[2]),J(c,"input",t[3]),J(T,"input",t[4]),J(I,"input",t[5]),J(E,"change",t[6])],ke=!0)},p(ae,[fe]){fe&2&&c.value!==ae[1].name&&Ze(c,ae[1].name),fe&2&&Ze(T,ae[1].text),fe&2&&Rt(I.value)!==ae[1].max_players&&Ze(I,ae[1].max_players),fe&2&&Fn(E,ae[1].filtering)},i(ae){me||(B(i.$$.fragment,ae),Le(()=>{me&&(ne||(ne=Ae(n,Me,{duration:200},!0)),ne.run(1))}),Le(()=>{me&&(de||(de=Ae(e,Nt,{duration:200},!0)),de.run(1))}),me=!0)},o(ae){G(i.$$.fragment,ae),ne||(ne=Ae(n,Me,{duration:200},!1)),ne.run(0),de||(de=Ae(e,Nt,{duration:200},!1)),de.run(0),me=!1},d(ae){ae&&V(e),oe(i),ae&&ne&&ne.end(),ae&&de&&de.end(),ke=!1,je(Te)}}}function r1(t,e,n){let r;lt(t,dt,u=>n(1,r=u));let{visible:i}=e;const s=()=>n(0,i=!1);function o(){r.name=this.value,dt.set(r)}function l(){r.text=this.value,dt.set(r)}function a(){r.max_players=Rt(this.value),dt.set(r)}function f(){r.filtering=Bi(this),dt.set(r)}return t.$$set=u=>{"visible"in u&&n(0,i=u.visible)},[i,r,s,o,l,a,f]}class i1 extends be{constructor(e){super(),ye(this,e,r1,n1,ve,{visible:0})}}function ao(t,e,n){const r=t.slice();return r[14]=e[n],r}function uo(t,e,n){const r=t.slice();return r[17]=e[n],r}function co(t){let e,n,r,i,s,o,l,a,f,u,d,c,h,p,y,k,_,C,T,S=t[2].length<1&&t[0].length<1&&fo(),$=t[2],w=[];for(let b=0;b<$.length;b+=1)w[b]=ho(uo(t,$,b));let x=t[0],v=[];for(let b=0;b<x.length;b+=1)v[b]=po(ao(t,x,b));return d=new Hl({}),{c(){e=D("div"),n=D("div"),r=D("div"),S&&S.c(),i=U();for(let b=0;b<w.length;b+=1)w[b].c();s=U();for(let b=0;b<v.length;b+=1)v[b].c();o=U(),l=D("div"),a=D("button"),a.textContent="Close",f=U(),u=D("button"),le(d.$$.fragment),c=te(`
-          Upload Images`),h=U(),p=D("button"),y=te("Delete All Images"),m(r,"class","images svelte-jdx2r3"),m(a,"class","btn btn--surface"),m(u,"class","btn btn--icon btn--surface"),m(p,"class","btn btn--surface"),p.disabled=k=t[2].length===0,m(l,"class","actions btn-row btn-row--fill svelte-jdx2r3"),m(n,"class","dialog svelte-jdx2r3"),m(e,"class","wrapper svelte-jdx2r3")},m(b,I){W(b,e,I),g(e,n),g(n,r),S&&S.m(r,null),g(r,i);for(let O=0;O<w.length;O+=1)w[O]&&w[O].m(r,null);g(r,s);for(let O=0;O<v.length;O+=1)v[O]&&v[O].m(r,null);g(n,o),g(n,l),g(l,a),g(l,f),g(l,u),se(d,u,null),g(u,c),g(l,h),g(l,p),g(p,y),_=!0,C||(T=[J(a,"click",$d),J(u,"click",t[4]),J(p,"click",t[7]),J(n,"drop",t[5]),J(n,"dragover",c1)],C=!0)},p(b,I){if(b[2].length<1&&b[0].length<1?S||(S=fo(),S.c(),S.m(r,i)):S&&(S.d(1),S=null),I&76){$=b[2];let O;for(O=0;O<$.length;O+=1){const R=uo(b,$,O);w[O]?w[O].p(R,I):(w[O]=ho(R),w[O].c(),w[O].m(r,s))}for(;O<w.length;O+=1)w[O].d(1);w.length=$.length}if(I&1){x=b[0];let O;for(O=0;O<x.length;O+=1){const R=ao(b,x,O);v[O]?v[O].p(R,I):(v[O]=po(R),v[O].c(),v[O].m(r,null))}for(;O<v.length;O+=1)v[O].d(1);v.length=x.length}(!_||I&4&&k!==(k=b[2].length===0))&&(p.disabled=k)},i(b){_||(B(d.$$.fragment,b),_=!0)},o(b){G(d.$$.fragment,b),_=!1},d(b){b&&V(e),S&&S.d(),Kn(w,b),Kn(v,b),oe(d),C=!1,je(T)}}}function fo(t){let e;return{c(){e=D("p"),e.textContent="Click upload or drag and drop files here to upload",m(e,"class","images__text svelte-jdx2r3")},m(n,r){W(n,e,r)},d(n){n&&V(e)}}}function s1(t){let e;return{c(){e=D("span"),e.textContent="Preview loading.."},m(n,r){W(n,e,r)},p:Q,d(n){n&&V(e)}}}function o1(t){let e,n;return{c(){e=D("img"),m(e,"class","image svelte-jdx2r3"),qn(e.src,n=t[3][t[17].uuid])||m(e,"src",n),m(e,"alt","Preview")},m(r,i){W(r,e,i)},p(r,i){i&12&&!qn(e.src,n=r[3][r[17].uuid])&&m(e,"src",n)},d(r){r&&V(e)}}}function ho(t){let e,n,r=t[17].name+"",i,s,o,l,a,f=xs(t[17].size)+"",u,d,c,h,p,y,k,_;function C(x,v){return x[3][x[17].uuid]!==void 0?o1:s1}let T=C(t),S=T(t);function $(){return t[8](t[17])}function w(){return t[9](t[17])}return{c(){e=D("div"),n=D("p"),i=te(r),s=U(),o=D("div"),S.c(),l=U(),a=D("p"),u=te(f),d=U(),c=D("div"),h=D("button"),h.textContent="Select",p=U(),y=D("button"),y.textContent="Delete",m(n,"class","file__name svelte-jdx2r3"),m(o,"class","image-wrapper svelte-jdx2r3"),m(h,"class","btn btn--small"),m(y,"class","btn btn--small"),m(c,"class","file__actions svelte-jdx2r3"),m(e,"class","file svelte-jdx2r3")},m(x,v){W(x,e,v),g(e,n),g(n,i),g(e,s),g(e,o),S.m(o,null),g(e,l),g(e,a),g(a,u),g(e,d),g(e,c),g(c,h),g(c,p),g(c,y),k||(_=[J(h,"click",$),J(y,"click",w)],k=!0)},p(x,v){t=x,v&4&&r!==(r=t[17].name+"")&&Ie(i,r),T===(T=C(t))&&S?S.p(t,v):(S.d(1),S=T(t),S&&(S.c(),S.m(o,null))),v&4&&f!==(f=xs(t[17].size)+"")&&Ie(u,f)},d(x){x&&V(e),S.d(),k=!1,je(_)}}}function l1(t){let e,n,r=t[14].progress+"",i;return{c(){e=D("p"),n=te("Progress: "),i=te(r)},m(s,o){W(s,e,o),g(e,n),g(e,i)},p(s,o){o&1&&r!==(r=s[14].progress+"")&&Ie(i,r)},d(s){s&&V(e)}}}function a1(t){let e,n=t[14].error+"",r;return{c(){e=D("p"),r=te(n),m(e,"class","error svelte-jdx2r3")},m(i,s){W(i,e,s),g(e,r)},p(i,s){s&1&&n!==(n=i[14].error+"")&&Ie(r,n)},d(i){i&&V(e)}}}function po(t){let e,n,r=t[14].name+"",i,s,o;function l(u,d){return u[14].error?a1:l1}let a=l(t),f=a(t);return{c(){e=D("div"),n=D("p"),i=te(r),s=U(),f.c(),o=U(),m(e,"class","file svelte-jdx2r3")},m(u,d){W(u,e,d),g(e,n),g(n,i),g(e,s),f.m(e,null),g(e,o)},p(u,d){d&1&&r!==(r=u[14].name+"")&&Ie(i,r),a===(a=l(u))&&f?f.p(u,d):(f.d(1),f=a(u),f&&(f.c(),f.m(e,o)))},d(u){u&&V(e),f.d()}}}function u1(t){let e,n,r=t[1]&&co(t);return{c(){r&&r.c(),e=Lt()},m(i,s){r&&r.m(i,s),W(i,e,s),n=!0},p(i,[s]){i[1]?r?(r.p(i,s),s&2&&B(r,1)):(r=co(i),r.c(),B(r,1),r.m(e.parentNode,e)):r&&(qe(),G(r,1,1,()=>{r=null}),Fe())},i(i){n||(B(r),n=!0)},o(i){G(r),n=!1},d(i){r&&r.d(i),i&&V(e)}}}function c1(t){t.stopPropagation(),t.preventDefault()}function f1(t,e,n){let r,i,s;lt(t,pi,_=>n(1,r=_)),lt(t,Nn,_=>n(2,i=_)),lt(t,Rn,_=>n(3,s=_));let o=[];async function l(){const _=await Kd("image/*");_!==null&&f(_)}function a(_){_.stopPropagation(),_.preventDefault();const C=_.dataTransfer;if(C===null)return;const T=C.files;T!==null&&f(T)}function f(_){for(const C of _)o.push({name:C.name,progress:0,error:null}),n(0,o),Md(C,T=>{c(C.name,T)}).then(()=>u(C)).catch(T=>d(C,T))}function u(_){n(0,o=o.filter(C=>C.name!==_.name))}function d(_,C){n(0,o=o.map(T=>(T.name===_.name&&(T.error=C.message),T)))}function c(_,C){n(0,o=o.map(T=>(T.name===_&&(T.progress=C),T)))}function h(_){Nn.update(C=>C.filter(T=>T.uuid!==_.uuid)),Rn.update(C=>(delete C[_.uuid],C))}async function p(){await Vi("Confirm Deletion","Are you sure you want to delete all the uploaded files?")&&(Nn.set([]),Rn.set({}),n(0,o=[]))}return[o,r,i,s,l,a,h,p,_=>Od(_),_=>h(_)]}class d1 extends be{constructor(e){super(),ye(this,e,f1,u1,ve,{})}}function mo(t){let e,n,r,i,s,o,l,a,f,u,d,c,h,p,y,k,_,C;return{c(){e=D("label"),n=D("span"),n.textContent="Image Fit",r=U(),i=D("p"),i.textContent="Decide how the image should be best fit for the player screens,",s=U(),o=D("p"),o.textContent=`It's recommended that you use "Contain" if the full image content is
-          important to be visible`,l=U(),a=D("select"),f=D("option"),u=te("Contain: Fit the entire image"),d=D("option"),c=te("Cover: Fill the available space"),h=D("option"),p=te(`Fill Width: Fill available width
-          `),y=D("option"),k=te("Fill Height: Fill available height"),m(n,"class","field__name svelte-1c3glip"),m(i,"class","field__desc svelte-1c3glip"),m(o,"class","field__desc svelte-1c3glip"),f.__value=Dn.Contain,f.value=f.__value,d.__value=Dn.Cover,d.value=d.__value,h.__value=Dn.Width,h.value=h.__value,y.__value=Dn.Height,y.value=y.__value,m(a,"class","input svelte-1c3glip"),t[0].image.fit===void 0&&Le(()=>t[3].call(a)),m(e,"class","field svelte-1c3glip")},m(T,S){W(T,e,S),g(e,n),g(e,r),g(e,i),g(e,s),g(e,o),g(e,l),g(e,a),g(a,f),g(f,u),g(a,d),g(d,c),g(a,h),g(h,p),g(a,y),g(y,k),Fn(a,t[0].image.fit,!0),_||(C=J(a,"change",t[3]),_=!0)},p(T,S){S&1&&Fn(a,T[0].image.fit)},d(T){T&&V(e),_=!1,C()}}}function h1(t){let e,n,r,i,s,o,l,a,f;i=new Bn({});let u=t[0].image!==null&&mo(t);return{c(){e=D("div"),n=D("div"),r=D("button"),le(i.$$.fragment),s=te(`
-      Close`),o=U(),u&&u.c(),m(r,"class","btn btn--icon btn--surface"),m(n,"class","dialog svelte-1c3glip"),m(e,"class","floating-wrapper svelte-1c3glip")},m(d,c){W(d,e,c),g(e,n),g(n,r),se(i,r,null),g(r,s),g(n,o),u&&u.m(n,null),l=!0,a||(f=J(r,"click",t[2]),a=!0)},p(d,[c]){d[0].image!==null?u?u.p(d,c):(u=mo(d),u.c(),u.m(n,null)):u&&(u.d(1),u=null)},i(d){l||(B(i.$$.fragment,d),l=!0)},o(d){G(i.$$.fragment,d),l=!1},d(d){d&&V(e),oe(i),u&&u.d(),a=!1,f()}}}function p1(t,e,n){let{question:r}=e,{visible:i}=e;const s=()=>n(1,i=!1);function o(){r.image.fit=Bi(this),n(0,r)}return t.$$set=l=>{"question"in l&&n(0,r=l.question),"visible"in l&&n(1,i=l.visible)},[r,i,s,o]}class m1 extends be{constructor(e){super(),ye(this,e,p1,h1,ve,{question:0,visible:1})}}function g1(t){let e,n,r;return{c(){e=D("button"),e.textContent="Pick Image",m(e,"class","add svelte-s58787")},m(i,s){W(i,e,s),n||(r=J(e,"click",t[3]),n=!0)},p:Q,i:Q,o:Q,d(i){i&&V(e),n=!1,r()}}}function _1(t){let e,n,r,i,s,o,l,a,f,u,d;return a=new cs({}),{c(){e=D("img"),i=U(),s=D("button"),s.textContent="Click to remove",o=U(),l=D("button"),le(a.$$.fragment),m(e,"class","image svelte-s58787"),m(e,"data-fit",n=t[0].image.fit),qn(e.src,r=t[1])||m(e,"src",r),m(e,"alt","Uploaded Content"),m(s,"class","overlay svelte-s58787"),m(l,"class","btn btn--icon btn--icon-only settings svelte-s58787")},m(c,h){W(c,e,h),W(c,i,h),W(c,s,h),W(c,o,h),W(c,l,h),se(a,l,null),f=!0,u||(d=[J(s,"click",t[4]),J(l,"click",t[6])],u=!0)},p(c,h){(!f||h&1&&n!==(n=c[0].image.fit))&&m(e,"data-fit",n),(!f||h&2&&!qn(e.src,r=c[1]))&&m(e,"src",r)},i(c){f||(B(a.$$.fragment,c),f=!0)},o(c){G(a.$$.fragment,c),f=!1},d(c){c&&V(e),c&&V(i),c&&V(s),c&&V(o),c&&V(l),oe(a),u=!1,je(d)}}}function go(t){let e,n,r,i;function s(a){t[7](a)}function o(a){t[8](a)}let l={};return t[0]!==void 0&&(l.question=t[0]),t[2]!==void 0&&(l.visible=t[2]),e=new m1({props:l}),Ke.push(()=>ut(e,"question",s)),Ke.push(()=>ut(e,"visible",o)),{c(){le(e.$$.fragment)},m(a,f){se(e,a,f),i=!0},p(a,f){const u={};!n&&f&1&&(n=!0,u.question=a[0],at(()=>n=!1)),!r&&f&4&&(r=!0,u.visible=a[2],at(()=>r=!1)),e.$set(u)},i(a){i||(B(e.$$.fragment,a),i=!0)},o(a){G(e.$$.fragment,a),i=!1},d(a){oe(e,a)}}}function v1(t){let e,n,r,i,s,o,l;const a=[_1,g1],f=[];function u(c,h){return c[0].image!==null&&c[1]?0:1}n=u(t),r=f[n]=a[n](t);let d=t[2]&&go(t);return o=new d1({}),{c(){e=D("div"),r.c(),i=U(),d&&d.c(),s=U(),le(o.$$.fragment),m(e,"tabindex","0"),m(e,"role","button"),m(e,"class","wrapper svelte-s58787")},m(c,h){W(c,e,h),f[n].m(e,null),W(c,i,h),d&&d.m(c,h),W(c,s,h),se(o,c,h),l=!0},p(c,[h]){let p=n;n=u(c),n===p?f[n].p(c,h):(qe(),G(f[p],1,1,()=>{f[p]=null}),Fe(),r=f[n],r?r.p(c,h):(r=f[n]=a[n](c),r.c()),B(r,1),r.m(e,null)),c[2]?d?(d.p(c,h),h&4&&B(d,1)):(d=go(c),d.c(),B(d,1),d.m(s.parentNode,s)):d&&(qe(),G(d,1,1,()=>{d=null}),Fe())},i(c){l||(B(r),B(d),B(o.$$.fragment,c),l=!0)},o(c){G(r),G(d),G(o.$$.fragment,c),l=!1},d(c){c&&V(e),f[n].d(),c&&V(i),d&&d.d(c),c&&V(s),oe(o,c)}}}function y1(t,e,n){let r;lt(t,Rn,c=>n(5,r=c));let{question:i}=e,s=null,o=!1;async function l(){let c=await Id();c!==null&&n(0,i.image={uuid:c.uuid,fit:Dn.Contain},i)}function a(c){c.stopPropagation(),n(0,i.image=null,i),n(1,s=null)}const f=()=>n(2,o=!0);function u(c){i=c,n(0,i)}function d(c){o=c,n(2,o)}return t.$$set=c=>{"question"in c&&n(0,i=c.question)},t.$$.update=()=>{if(t.$$.dirty&33&&i.image!==null){let c=r[i.image.uuid];c!==void 0?n(1,s=c):n(1,s=null)}},[i,s,o,l,a,r,f,u,d]}class b1 extends be{constructor(e){super(),ye(this,e,y1,v1,ve,{question:0})}}function w1(t){let e,n,r,i,s,o,l,a,f,u,d,c;return{c(){e=D("div"),n=D("input"),s=U(),o=D("select"),l=D("option"),a=te("Minutes"),f=D("option"),u=te("Seconds"),m(n,"class","input svelte-t6875b"),m(n,"type","number"),m(n,"min",r=t[5](t[0])),m(n,"max",i=t[5](t[1])),l.__value=t[4].Minutes,l.value=l.__value,f.__value=t[4].Seconds,f.value=f.__value,m(o,"class","select svelte-t6875b"),t[2]===void 0&&Le(()=>t[9].call(o))},m(h,p){W(h,e,p),g(e,n),Ze(n,t[3]),g(e,s),g(e,o),g(o,l),g(l,a),g(o,f),g(f,u),Fn(o,t[2],!0),d||(c=[J(n,"input",t[8]),J(n,"change",t[6]),J(o,"change",t[9]),J(o,"change",t[6])],d=!0)},p(h,[p]){p&1&&r!==(r=h[5](h[0]))&&m(n,"min",r),p&2&&i!==(i=h[5](h[1]))&&m(n,"max",i),p&8&&Rt(n.value)!==h[3]&&Ze(n,h[3]),p&20&&Fn(o,h[2])},i:Q,o:Q,d(h){h&&V(e),d=!1,je(c)}}}function k1(t,e,n){let{value:r}=e,{min:i}=e,{max:s}=e;var o=(p=>(p[p.Seconds=0]="Seconds",p[p.Minutes=1]="Minutes",p))(o||{});let l=0,a=u(r);function f(p){return l==0?p*1e3:l==1?p*1e3*60:p}function u(p){return l==0?p/1e3:l==1?p/60/1e3:p}function d(){if(a<1)n(3,a=1);else{let p=f(a);p>s?p=s:p<i&&(p=i),n(3,a=u(p)),n(7,r=p)}}function c(){a=Rt(this.value),n(3,a)}function h(){l=Bi(this),n(2,l),n(4,o)}return t.$$set=p=>{"value"in p&&n(7,r=p.value),"min"in p&&n(0,i=p.min),"max"in p&&n(1,s=p.max)},[i,s,l,a,o,u,d,r,c,h]}class _o extends be{constructor(e){super(),ye(this,e,k1,w1,ve,{value:7,min:0,max:1})}}function E1(t){let e,n,r,i,s,o,l,a,f,u,d,c,h,p,y,k,_,C,T,S,$,w,x,v,b,I,O,R,N,M,L,A,E,q,F,P,Y,j,H,z,Z,ne,de,me,ke,Te,ae,fe,Ge,Je,$e,Ee,pe,Se,Ue,Re,xe,Pe,K,ie,Ne;i=new Bn({});function Qe(ge){t[3](ge)}let on={min:Ju,max:ec};t[0].answer_time!==void 0&&(on.value=t[0].answer_time),C=new _o({props:on}),Ke.push(()=>ut(C,"value",Qe));function Ye(ge){t[4](ge)}let It={min:sc,max:oc};return t[0].bonus_score_time!==void 0&&(It.value=t[0].bonus_score_time),I=new _o({props:It}),Ke.push(()=>ut(I,"value",Ye)),{c(){e=D("div"),n=D("div"),r=D("button"),le(i.$$.fragment),s=te(`
-      Close`),o=U(),l=D("div"),a=D("h2"),a.textContent="Timing",f=U(),u=D("p"),u.textContent="Below you can configuring the timing for different events",d=U(),c=D("div"),h=D("div"),p=D("span"),p.textContent="Answer Time",y=U(),k=D("p"),k.textContent="Time the players have to answer the question",_=U(),le(C.$$.fragment),S=U(),$=D("div"),w=D("span"),w.textContent="Bonus Score Time",x=U(),v=D("p"),v.textContent="Time the players must answer within for bonus scores",b=U(),le(I.$$.fragment),R=U(),N=D("div"),M=D("h2"),M.textContent="Scoring",L=U(),A=D("p"),A.textContent=`Score is awarded to players based on how quickly the player answers the
-        question. You can configure the minimum and maximum values for this
-        below`,E=U(),q=D("div"),F=D("label"),P=D("span"),P.textContent="Min Score",Y=U(),j=D("p"),j.textContent="The minimum amount of score to award for this question",H=U(),z=D("input"),ne=U(),de=D("label"),me=D("span"),me.textContent="Max Score",ke=U(),Te=D("p"),Te.textContent="The maximum amount of score to award for this question",ae=U(),fe=D("input"),Je=U(),$e=D("label"),Ee=D("span"),Ee.textContent="Bonus Score",pe=U(),Se=D("p"),Se.textContent="The amount of score to add for being within the bonus time",Ue=U(),Re=D("input"),m(r,"class","btn btn--icon btn--surface"),m(a,"class","section__title svelte-1i7jozy"),m(u,"class","section__desc svelte-1i7jozy"),m(p,"class","field__name svelte-1i7jozy"),m(k,"class","field__desc svelte-1i7jozy"),m(h,"class","field svelte-1i7jozy"),m(w,"class","field__name svelte-1i7jozy"),m(v,"class","field__desc svelte-1i7jozy"),m($,"class","field svelte-1i7jozy"),m(c,"class","field-group svelte-1i7jozy"),m(l,"class","section svelte-1i7jozy"),m(M,"class","section__title svelte-1i7jozy"),m(A,"class","section__desc svelte-1i7jozy"),m(P,"class","field__name svelte-1i7jozy"),m(j,"class","field__desc svelte-1i7jozy"),m(z,"class","input svelte-1i7jozy"),m(z,"type","number"),m(z,"min",Es),m(z,"max",Z=t[0].scoring.max_score),m(F,"class","field svelte-1i7jozy"),m(me,"class","field__name svelte-1i7jozy"),m(Te,"class","field__desc svelte-1i7jozy"),m(fe,"class","input svelte-1i7jozy"),m(fe,"type","number"),m(fe,"min",Ge=t[0].scoring.min_score),m(fe,"max",rc),m(de,"class","field svelte-1i7jozy"),m(Ee,"class","field__name svelte-1i7jozy"),m(Se,"class","field__desc svelte-1i7jozy"),m(Re,"class","input svelte-1i7jozy"),m(Re,"type","number"),m(Re,"min",Es),m(Re,"max",ic),m($e,"class","field svelte-1i7jozy"),m(q,"class","field-group svelte-1i7jozy"),m(N,"class","section svelte-1i7jozy"),m(n,"class","dialog svelte-1i7jozy"),m(e,"class","floating-wrapper svelte-1i7jozy")},m(ge,He){W(ge,e,He),g(e,n),g(n,r),se(i,r,null),g(r,s),g(n,o),g(n,l),g(l,a),g(l,f),g(l,u),g(l,d),g(l,c),g(c,h),g(h,p),g(h,y),g(h,k),g(h,_),se(C,h,null),g(c,S),g(c,$),g($,w),g($,x),g($,v),g($,b),se(I,$,null),g(n,R),g(n,N),g(N,M),g(N,L),g(N,A),g(N,E),g(N,q),g(q,F),g(F,P),g(F,Y),g(F,j),g(F,H),g(F,z),Ze(z,t[0].scoring.min_score),g(q,ne),g(q,de),g(de,me),g(de,ke),g(de,Te),g(de,ae),g(de,fe),Ze(fe,t[0].scoring.max_score),g(q,Je),g(q,$e),g($e,Ee),g($e,pe),g($e,Se),g($e,Ue),g($e,Re),Ze(Re,t[0].scoring.bonus_score),K=!0,ie||(Ne=[J(r,"click",t[2]),J(z,"input",t[5]),J(fe,"input",t[6]),J(Re,"input",t[7])],ie=!0)},p(ge,[He]){const ft={};!T&&He&1&&(T=!0,ft.value=ge[0].answer_time,at(()=>T=!1)),C.$set(ft);const yt={};!O&&He&1&&(O=!0,yt.value=ge[0].bonus_score_time,at(()=>O=!1)),I.$set(yt),(!K||He&1&&Z!==(Z=ge[0].scoring.max_score))&&m(z,"max",Z),He&1&&Rt(z.value)!==ge[0].scoring.min_score&&Ze(z,ge[0].scoring.min_score),(!K||He&1&&Ge!==(Ge=ge[0].scoring.min_score))&&m(fe,"min",Ge),He&1&&Rt(fe.value)!==ge[0].scoring.max_score&&Ze(fe,ge[0].scoring.max_score),He&1&&Rt(Re.value)!==ge[0].scoring.bonus_score&&Ze(Re,ge[0].scoring.bonus_score)},i(ge){K||(B(i.$$.fragment,ge),B(C.$$.fragment,ge),B(I.$$.fragment,ge),Le(()=>{K&&(xe||(xe=Ae(n,Me,{duration:200},!0)),xe.run(1))}),Le(()=>{K&&(Pe||(Pe=Ae(e,Nt,{duration:200},!0)),Pe.run(1))}),K=!0)},o(ge){G(i.$$.fragment,ge),G(C.$$.fragment,ge),G(I.$$.fragment,ge),xe||(xe=Ae(n,Me,{duration:200},!1)),xe.run(0),Pe||(Pe=Ae(e,Nt,{duration:200},!1)),Pe.run(0),K=!1},d(ge){ge&&V(e),oe(i),oe(C),oe(I),ge&&xe&&xe.end(),ge&&Pe&&Pe.end(),ie=!1,je(Ne)}}}function A1(t,e,n){let{question:r}=e,{visible:i}=e;const s=()=>n(1,i=!1);function o(d){t.$$.not_equal(r.answer_time,d)&&(r.answer_time=d,n(0,r))}function l(d){t.$$.not_equal(r.bonus_score_time,d)&&(r.bonus_score_time=d,n(0,r))}function a(){r.scoring.min_score=Rt(this.value),n(0,r)}function f(){r.scoring.max_score=Rt(this.value),n(0,r)}function u(){r.scoring.bonus_score=Rt(this.value),n(0,r)}return t.$$set=d=>{"question"in d&&n(0,r=d.question),"visible"in d&&n(1,i=d.visible)},[r,i,s,o,l,a,f,u]}class T1 extends be{constructor(e){super(),ye(this,e,A1,E1,ve,{question:0,visible:1})}}function C1(t){let e,n,r=!1,i,s,o,l,a,f,u;const d=t[3].default,c=Yl(d,t,t[2],null);return a=na(t[6][0]),{c(){e=D("label"),n=D("input"),i=U(),s=D("div"),s.innerHTML='<span class="mark svelte-1t2uex"></span>',o=U(),c&&c.c(),m(n,"class","input svelte-1t2uex"),m(n,"type","radio"),n.__value=t[1],n.value=n.__value,m(s,"class","wrapper svelte-1t2uex"),m(e,"class","label svelte-1t2uex"),a.p(n)},m(h,p){W(h,e,p),g(e,n),n.checked=n.__value===t[0],g(e,i),g(e,s),g(e,o),c&&c.m(e,null),l=!0,f||(u=[J(n,"change",t[4]),J(n,"change",t[5])],f=!0)},p(h,[p]){(!l||p&2)&&(n.__value=h[1],n.value=n.__value,r=!0),(r||p&1)&&(n.checked=n.__value===h[0]),c&&c.p&&(!l||p&4)&&Kl(c,d,h,h[2],l?Xl(d,h[2],p,null):Jl(h[2]),null)},i(h){l||(B(c,h),l=!0)},o(h){G(c,h),l=!1},d(h){h&&V(e),c&&c.d(h),a.r(),f=!1,je(u)}}}function x1(t,e,n){let{$$slots:r={},$$scope:i}=e,{value:s}=e,{group:o}=e;const l=[[]];function a(u){ca.call(this,t,u)}function f(){o=this.__value,n(0,o)}return t.$$set=u=>{"value"in u&&n(1,s=u.value),"group"in u&&n(0,o=u.group),"$$scope"in u&&n(2,i=u.$$scope)},[o,s,i,r,a,f,l]}class D1 extends be{constructor(e){super(),ye(this,e,x1,C1,ve,{value:1,group:0})}}function vo(t,e,n){const r=t.slice();return r[5]=e[n],r}function S1(t){let e,n=t[5]+"",r,i,s,o=Du[t[5]]+"",l;return{c(){e=D("b"),r=te(n),i=te(" Choice"),s=te(": "),l=te(o)},m(a,f){W(a,e,f),g(e,r),g(e,i),W(a,s,f),W(a,l,f)},p:Q,d(a){a&&V(e),a&&V(s),a&&V(l)}}}function yo(t){let e,n,r,i,s;function o(a){t[4](a)}let l={value:t[5],$$slots:{default:[S1]},$$scope:{ctx:t}};return t[0].ty!==void 0&&(l.group=t[0].ty),n=new D1({props:l}),Ke.push(()=>ut(n,"group",o)),n.$on("change",t[2]),{c(){e=D("div"),le(n.$$.fragment),i=U()},m(a,f){W(a,e,f),se(n,e,null),g(e,i),s=!0},p(a,f){const u={};f&256&&(u.$$scope={dirty:f,ctx:a}),!r&&f&1&&(r=!0,u.group=a[0].ty,at(()=>r=!1)),n.$set(u)},i(a){s||(B(n.$$.fragment,a),s=!0)},o(a){G(n.$$.fragment,a),s=!1},d(a){a&&V(e),oe(n)}}}function I1(t){let e,n,r,i,s,o,l,a,f,u,d,c,h,p,y,k,_;i=new Bn({});let C=Object.values(rt),T=[];for(let $=0;$<C.length;$+=1)T[$]=yo(vo(t,C,$));const S=$=>G(T[$],1,1,()=>{T[$]=null});return{c(){e=D("div"),n=D("div"),r=D("button"),le(i.$$.fragment),s=te(`
-      Close`),o=U(),l=D("div"),a=D("p"),a.textContent="Question Type",f=U(),u=D("p"),u.textContent="The type of question to use",d=U(),c=D("div");for(let $=0;$<T.length;$+=1)T[$].c();m(r,"class","btn btn--icon btn--surface"),m(a,"class","field__name"),m(u,"class","field__desc"),m(c,"class","radio svelte-16is7hd"),m(l,"class","field"),m(n,"class","dialog svelte-16is7hd"),m(e,"class","floating-wrapper svelte-16is7hd")},m($,w){W($,e,w),g(e,n),g(n,r),se(i,r,null),g(r,s),g(n,o),g(n,l),g(l,a),g(l,f),g(l,u),g(l,d),g(l,c);for(let x=0;x<T.length;x+=1)T[x]&&T[x].m(c,null);y=!0,k||(_=J(r,"click",t[3]),k=!0)},p($,[w]){if(w&5){C=Object.values(rt);let x;for(x=0;x<C.length;x+=1){const v=vo($,C,x);T[x]?(T[x].p(v,w),B(T[x],1)):(T[x]=yo(v),T[x].c(),B(T[x],1),T[x].m(c,null))}for(qe(),x=C.length;x<T.length;x+=1)S(x);Fe()}},i($){if(!y){B(i.$$.fragment,$);for(let w=0;w<C.length;w+=1)B(T[w]);Le(()=>{y&&(h||(h=Ae(n,Me,{duration:200},!0)),h.run(1))}),Le(()=>{y&&(p||(p=Ae(e,Nt,{duration:200},!0)),p.run(1))}),y=!0}},o($){G(i.$$.fragment,$),T=T.filter(Boolean);for(let w=0;w<T.length;w+=1)G(T[w]);h||(h=Ae(n,Me,{duration:200},!1)),h.run(0),p||(p=Ae(e,Nt,{duration:200},!1)),p.run(0),y=!1},d($){$&&V(e),oe(i),Kn(T,$),$&&h&&h.end(),$&&p&&p.end(),k=!1,_()}}}function $1(t,e,n){let{question:r}=e,{visible:i}=e;function s(a){console.log(a),n(0,r=ro(r))}const o=()=>n(1,i=!1);function l(a){t.$$.not_equal(r.ty,a)&&(r.ty=a,n(0,r))}return t.$$set=a=>{"question"in a&&n(0,r=a.question),"visible"in a&&n(1,i=a.visible)},t.$$.update=()=>{t.$$.dirty&1&&n(0,r=ro(r))},[r,i,s,o,l]}class O1 extends be{constructor(e){super(),ye(this,e,$1,I1,ve,{question:0,visible:1})}}function M1(t){let e,n,r,i,s,o,l;return{c(){e=D("label"),n=D("input"),r=U(),i=ce("svg"),s=ce("path"),m(n,"class","input svelte-1sqhokz"),m(n,"type","checkbox"),m(s,"d","M4.575 8.977s-.404-.007-.536-.165L.057 5.64l.829-1.227L4.47 7.268 10.921.086l.923 1.103-6.863 7.64c-.134-.003-.406.148-.406.148z"),m(s,"fill","#FFF"),m(s,"fill-rule","evenodd"),m(i,"width","16"),m(i,"height","16"),m(i,"viewBox","0 0 12 9"),m(i,"xmlns","http://www.w3.org/2000/svg"),m(i,"class","mark svelte-1sqhokz"),m(e,"class","wrapper svelte-1sqhokz")},m(a,f){W(a,e,f),g(e,n),n.checked=t[0],g(e,r),g(e,i),g(i,s),o||(l=J(n,"change",t[1]),o=!0)},p(a,[f]){f&1&&(n.checked=a[0])},i:Q,o:Q,d(a){a&&V(e),o=!1,l()}}}function R1(t,e,n){let{value:r}=e;function i(){r=this.checked,n(0,r)}return t.$$set=s=>{"value"in s&&n(0,r=s.value)},[r,i]}class N1 extends be{constructor(e){super(),ye(this,e,R1,M1,ve,{value:0})}}function bo(t,e,n){const r=t.slice();return r[6]=e[n],r[7]=e,r[8]=n,r}function wo(t){let e,n=[],r=new Map,i,s,o=t[0].answers;const l=f=>f[6].id;for(let f=0;f<o.length;f+=1){let u=bo(t,o,f),d=l(u);r.set(d,n[f]=ko(d,u))}let a=t[0].answers.length<ws&&Eo(t);return{c(){e=D("div");for(let f=0;f<n.length;f+=1)n[f].c();i=U(),a&&a.c(),m(e,"class","answers svelte-ok7b4d")},m(f,u){W(f,e,u);for(let d=0;d<n.length;d+=1)n[d]&&n[d].m(e,null);g(e,i),a&&a.m(e,null),s=!0},p(f,u){if(u&5){o=f[0].answers,qe();for(let d=0;d<n.length;d+=1)n[d].r();n=_r(n,u,l,1,f,o,r,e,ii,ko,i,bo);for(let d=0;d<n.length;d+=1)n[d].a();Fe()}f[0].answers.length<ws?a?a.p(f,u):(a=Eo(f),a.c(),a.m(e,null)):a&&(a.d(1),a=null)},i(f){if(!s){for(let u=0;u<o.length;u+=1)B(n[u]);s=!0}},o(f){for(let u=0;u<n.length;u+=1)G(n[u]);s=!1},d(f){f&&V(e);for(let u=0;u<n.length;u+=1)n[u].d();a&&a.d()}}}function ko(t,e){let n,r,i,s,o,l,a,f,u,d,c,h=Q,p,y,k;function _($){e[3]($,e[6])}let C={};e[6].correct!==void 0&&(C.value=e[6].correct),i=new N1({props:C}),Ke.push(()=>ut(i,"value",_));function T(){e[4].call(l,e[7],e[8])}u=new Bn({});function S(){return e[5](e[8])}return{key:t,first:null,c(){n=D("div"),r=D("div"),le(i.$$.fragment),o=U(),l=D("input"),a=U(),f=D("button"),le(u.$$.fragment),m(r,"class","answer__check svelte-ok7b4d"),m(l,"class","answer__input svelte-ok7b4d"),m(l,"type","text"),m(l,"maxlength",Yu),f.disabled=d=e[0].answers.length==1,m(f,"class","btn btn--surface btn--icon-only"),m(n,"class","answer svelte-ok7b4d"),this.first=n},m($,w){W($,n,w),g(n,r),se(i,r,null),g(n,o),g(n,l),Ze(l,e[6].value),g(n,a),g(n,f),se(u,f,null),p=!0,y||(k=[J(l,"input",T),J(f,"click",S)],y=!0)},p($,w){e=$;const x={};!s&&w&1&&(s=!0,x.value=e[6].correct,at(()=>s=!1)),i.$set(x),w&1&&l.value!==e[6].value&&Ze(l,e[6].value),(!p||w&1&&d!==(d=e[0].answers.length==1))&&(f.disabled=d)},r(){c=n.getBoundingClientRect()},f(){gr(n),h()},a(){h(),h=mr(n,c,wr,{duration:200})},i($){p||(B(i.$$.fragment,$),B(u.$$.fragment,$),p=!0)},o($){G(i.$$.fragment,$),G(u.$$.fragment,$),p=!1},d($){$&&V(n),oe(i),oe(u),y=!1,je(k)}}}function Eo(t){let e,n,r;return{c(){e=D("button"),e.textContent="Add Answer",m(e,"class","btn add svelte-ok7b4d")},m(i,s){W(i,e,s),n||(r=J(e,"click",t[1]),n=!0)},p:Q,d(i){i&&V(e),n=!1,r()}}}function q1(t){let e,n,r=(t[0].ty==rt.Single||t[0].ty==rt.Multiple)&&wo(t);return{c(){r&&r.c(),e=Lt()},m(i,s){r&&r.m(i,s),W(i,e,s),n=!0},p(i,[s]){i[0].ty==rt.Single||i[0].ty==rt.Multiple?r?(r.p(i,s),s&1&&B(r,1)):(r=wo(i),r.c(),B(r,1),r.m(e.parentNode,e)):r&&(qe(),G(r,1,1,()=>{r=null}),Fe())},i(i){n||(B(r),n=!0)},o(i){G(r),n=!1},d(i){r&&r.d(i),i&&V(e)}}}function F1(t,e,n){let{question:r}=e;function i(){let f=0;for(const u of r.answers)u.id>=f&&(f=u.id+1);r.answers.push({id:f,value:"",correct:!1}),n(0,r)}function s(f){n(0,r.answers=r.answers.filter((u,d)=>d!=f),r)}function o(f,u){t.$$.not_equal(u.correct,f)&&(u.correct=f,n(0,r))}function l(f,u){f[u].value=this.value,n(0,r)}const a=f=>s(f);return t.$$set=f=>{"question"in f&&n(0,r=f.question)},[r,i,s,o,l,a]}class L1 extends be{constructor(e){super(),ye(this,e,F1,q1,ve,{question:0})}}function P1(t){let e,n;return{c(){e=ce("svg"),n=ce("path"),m(n,"d","M666-435q-9-9-9-21t9-21l100-100H480q-13 0-21.5-8.5T450-607q0-13 8.5-21.5T480-637h286L665-738q-9-9-8.5-20.5T666-779q9-9 21-9t21 9l151 151q5 5 7 10t2 11q0 6-2 11t-7 10L707-434q-9 9-20.5 8.5T666-435ZM252-181 101-332q-5-5-7-10t-2-11q0-6 2-11t7-10l152-152q9-9 20.5-8.5T294-525q9 9 9 21t-9 21L194-383h286q13 0 21.5 8.5T510-353q0 13-8.5 21.5T480-323H194l101 101q9 9 8.5 20.5T294-181q-9 9-21 9t-21-9Z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 -960 960 960"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(r,i){W(r,e,i),g(e,n)},p:Q,i:Q,o:Q,d(r){r&&V(e)}}}class U1 extends be{constructor(e){super(),ye(this,e,null,P1,ve,{})}}function Ao(t){let e,n,r,i,s,o,l;return n=new Bl({}),{c(){e=D("button"),le(n.$$.fragment),r=te(`\r
-      Shuffle Answers`),m(e,"class","btn btn--icon svelte-1tif7ap"),e.disabled=i=t[0].answers.length<=1},m(a,f){W(a,e,f),se(n,e,null),g(e,r),s=!0,o||(l=J(e,"click",t[3]),o=!0)},p(a,f){(!s||f&1&&i!==(i=a[0].answers.length<=1))&&(e.disabled=i)},i(a){s||(B(n.$$.fragment,a),s=!0)},o(a){G(n.$$.fragment,a),s=!1},d(a){a&&V(e),oe(n),o=!1,l()}}}function To(t){let e,n,r,i;function s(a){t[10](a)}function o(a){t[11](a)}let l={};return t[0]!==void 0&&(l.question=t[0]),t[1]!==void 0&&(l.visible=t[1]),e=new T1({props:l}),Ke.push(()=>ut(e,"question",s)),Ke.push(()=>ut(e,"visible",o)),{c(){le(e.$$.fragment)},m(a,f){se(e,a,f),i=!0},p(a,f){const u={};!n&&f&1&&(n=!0,u.question=a[0],at(()=>n=!1)),!r&&f&2&&(r=!0,u.visible=a[1],at(()=>r=!1)),e.$set(u)},i(a){i||(B(e.$$.fragment,a),i=!0)},o(a){G(e.$$.fragment,a),i=!1},d(a){oe(e,a)}}}function Co(t){let e,n,r,i;function s(a){t[12](a)}function o(a){t[13](a)}let l={};return t[0]!==void 0&&(l.question=t[0]),t[2]!==void 0&&(l.visible=t[2]),e=new O1({props:l}),Ke.push(()=>ut(e,"question",s)),Ke.push(()=>ut(e,"visible",o)),{c(){le(e.$$.fragment)},m(a,f){se(e,a,f),i=!0},p(a,f){const u={};!n&&f&1&&(n=!0,u.question=a[0],at(()=>n=!1)),!r&&f&4&&(r=!0,u.visible=a[2],at(()=>r=!1)),e.$set(u)},i(a){i||(B(e.$$.fragment,a),i=!0)},o(a){G(e.$$.fragment,a),i=!1},d(a){oe(e,a)}}}function z1(t){let e,n,r,i,s,o,l,a,f,u,d,c,h,p,y=t[0].ty+"",k,_,C,T,S,$,w,x,v,b,I,O,R,N,M,L;function A(H){t[5](H)}let E={};t[0]!==void 0&&(E.question=t[0]),e=new b1({props:E}),Ke.push(()=>ut(e,"question",A)),o=new cs({});let q=t[0].answers!==void 0&&Ao(t);c=new U1({}),T=new Bn({});function F(H){t[9](H)}let P={};t[0]!==void 0&&(P.question=t[0]),v=new L1({props:P}),Ke.push(()=>ut(v,"question",F));let Y=t[1]&&To(t),j=t[2]&&Co(t);return{c(){le(e.$$.fragment),r=U(),i=D("div"),s=D("button"),le(o.$$.fragment),l=U(),a=D("span"),a.textContent="Settings",f=U(),q&&q.c(),u=U(),d=D("button"),le(c.$$.fragment),h=te(`\r
-    Change Type\r
-\r
-    `),p=D("span"),k=te(y),_=U(),C=D("button"),le(T.$$.fragment),S=te(`\r
-    Delete`),$=U(),w=D("textarea"),x=U(),le(v.$$.fragment),I=U(),Y&&Y.c(),O=U(),j&&j.c(),R=Lt(),m(s,"class","btn btn--icon svelte-1tif7ap"),m(p,"class","qt__type svelte-1tif7ap"),m(d,"class","btn btn--icon qt svelte-1tif7ap"),m(C,"class","btn btn--icon svelte-1tif7ap"),m(i,"class","actions btn-row svelte-1tif7ap"),m(w,"class","text input svelte-1tif7ap"),m(w,"cols","30"),m(w,"rows","2"),m(w,"maxlength",Xu)},m(H,z){se(e,H,z),W(H,r,z),W(H,i,z),g(i,s),se(o,s,null),g(s,l),g(s,a),g(i,f),q&&q.m(i,null),g(i,u),g(i,d),se(c,d,null),g(d,h),g(d,p),g(p,k),g(i,_),g(i,C),se(T,C,null),g(C,S),W(H,$,z),W(H,w,z),Ze(w,t[0].text),W(H,x,z),se(v,H,z),W(H,I,z),Y&&Y.m(H,z),W(H,O,z),j&&j.m(H,z),W(H,R,z),N=!0,M||(L=[J(s,"click",t[6]),J(d,"click",t[7]),J(C,"click",t[4]),J(w,"input",t[8])],M=!0)},p(H,[z]){const Z={};!n&&z&1&&(n=!0,Z.question=H[0],at(()=>n=!1)),e.$set(Z),H[0].answers!==void 0?q?(q.p(H,z),z&1&&B(q,1)):(q=Ao(H),q.c(),B(q,1),q.m(i,u)):q&&(qe(),G(q,1,1,()=>{q=null}),Fe()),(!N||z&1)&&y!==(y=H[0].ty+"")&&Ie(k,y),z&1&&Ze(w,H[0].text);const ne={};!b&&z&1&&(b=!0,ne.question=H[0],at(()=>b=!1)),v.$set(ne),H[1]?Y?(Y.p(H,z),z&2&&B(Y,1)):(Y=To(H),Y.c(),B(Y,1),Y.m(O.parentNode,O)):Y&&(qe(),G(Y,1,1,()=>{Y=null}),Fe()),H[2]?j?(j.p(H,z),z&4&&B(j,1)):(j=Co(H),j.c(),B(j,1),j.m(R.parentNode,R)):j&&(qe(),G(j,1,1,()=>{j=null}),Fe())},i(H){N||(B(e.$$.fragment,H),B(o.$$.fragment,H),B(q),B(c.$$.fragment,H),B(T.$$.fragment,H),B(v.$$.fragment,H),B(Y),B(j),N=!0)},o(H){G(e.$$.fragment,H),G(o.$$.fragment,H),G(q),G(c.$$.fragment,H),G(T.$$.fragment,H),G(v.$$.fragment,H),G(Y),G(j),N=!1},d(H){oe(e,H),H&&V(r),H&&V(i),oe(o),q&&q.d(),oe(c),oe(T),H&&V($),H&&V(w),H&&V(x),oe(v,H),H&&V(I),Y&&Y.d(H),H&&V(O),j&&j.d(H),H&&V(R),M=!1,je(L)}}}function Z1(t,e,n){let{question:r}=e,i=!1,s=!1;function o(){n(0,r.answers=bc(r.answers),r)}async function l(){await Vi("Confirm Deletion","Are you sure you want to delete this question?")&&bd(r)}function a(_){r=_,n(0,r)}const f=()=>n(1,i=!0),u=()=>n(2,s=!0);function d(){r.text=this.value,n(0,r)}function c(_){r=_,n(0,r)}function h(_){r=_,n(0,r)}function p(_){i=_,n(1,i)}function y(_){r=_,n(0,r)}function k(_){s=_,n(2,s)}return t.$$set=_=>{"question"in _&&n(0,r=_.question)},[r,i,s,o,l,a,f,u,d,c,h,p,y,k]}class j1 extends be{constructor(e){super(),ye(this,e,Z1,z1,ve,{question:0})}}function xo(t,e,n){const r=t.slice();return r[19]=e[n],r[20]=e,r[21]=n,r}function Do(t){let e,n,r,i;const s=[B1,H1],o=[];function l(a,f){return a[1]==="Uploading"?0:1}return e=l(t),n=o[e]=s[e](t),{c(){n.c(),r=Lt()},m(a,f){o[e].m(a,f),W(a,r,f),i=!0},p(a,f){let u=e;e=l(a),e===u?o[e].p(a,f):(qe(),G(o[u],1,1,()=>{o[u]=null}),Fe(),n=o[e],n?n.p(a,f):(n=o[e]=s[e](a),n.c()),B(n,1),n.m(r.parentNode,r))},i(a){i||(B(n),i=!0)},o(a){G(n),i=!1},d(a){o[e].d(a),a&&V(r)}}}function H1(t){let e,n;return e=new Yi({props:{text:"Connecting..."}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p:Q,i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function B1(t){let e,n;return e=new Yi({props:{text:`Uploading ${t[4].toFixed(0)}%`}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p(r,i){const s={};i&16&&(s.text=`Uploading ${r[4].toFixed(0)}%`),e.$set(s)},i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function So(t){let e,n,r;function i(o){t[12](o)}let s={};return t[2]!==void 0&&(s.visible=t[2]),e=new i1({props:s}),Ke.push(()=>ut(e,"visible",i)),{c(){le(e.$$.fragment)},m(o,l){se(e,o,l),r=!0},p(o,l){const a={};!n&&l&4&&(n=!0,a.visible=o[2],at(()=>n=!1)),e.$set(a)},i(o){r||(B(e.$$.fragment,o),r=!0)},o(o){G(e.$$.fragment,o),r=!1},d(o){oe(e,o)}}}function Io(t){let e;return{c(){e=D("div"),m(e,"class","shadow-item svelte-1k3dryf")},m(n,r){W(n,e,r)},d(n){n&&V(e)}}}function $o(t,e){let n,r,i,s,o,l,a=Q,f;function u(h){e[14](h,e[19],e[20],e[21])}let d={index:e[21]};e[19]!==void 0&&(d.question=e[19]),r=new Ud({props:d}),Ke.push(()=>ut(r,"question",u));let c=e[19][Zn]&&Io();return{key:t,first:null,c(){n=D("div"),le(r.$$.fragment),s=U(),c&&c.c(),o=U(),Ai(n,"position","relative"),this.first=n},m(h,p){W(h,n,p),se(r,n,null),g(n,s),c&&c.m(n,null),g(n,o),f=!0},p(h,p){e=h;const y={};p&8&&(y.index=e[21]),!i&&p&8&&(i=!0,y.question=e[19],at(()=>i=!1)),r.$set(y),e[19][Zn]?c||(c=Io(),c.c(),c.m(n,o)):c&&(c.d(1),c=null)},r(){l=n.getBoundingClientRect()},f(){gr(n),a()},a(){a(),a=mr(n,l,wr,{duration:200})},i(h){f||(B(r.$$.fragment,h),f=!0)},o(h){G(r.$$.fragment,h),f=!1},d(h){h&&V(n),oe(r),c&&c.d()}}}function V1(t){let e;return{c(){e=D("div"),e.innerHTML=`<h1 class="svelte-1k3dryf">Select a question</h1> 
-          <p>No question selected. Select a question to start editing</p>`,m(e,"class","editor__none svelte-1k3dryf")},m(n,r){W(n,e,r)},p:Q,i:Q,o:Q,d(n){n&&V(e)}}}function W1(t){let e,n,r;function i(o){t[15](o)}let s={};return t[5]!==void 0&&(s.question=t[5]),e=new j1({props:s}),Ke.push(()=>ut(e,"question",i)),{c(){le(e.$$.fragment)},m(o,l){se(e,o,l),r=!0},p(o,l){const a={};!n&&l&32&&(n=!0,a.question=o[5],at(()=>n=!1)),e.$set(a)},i(o){r||(B(e.$$.fragment,o),r=!0)},o(o){G(e.$$.fragment,o),r=!1},d(o){oe(e,o)}}}function G1(t){let e,n,r,i,s,o,l,a,f,u,d,c,h,p,y,k,_,C,T,S,$,w,x,v,b,I,O,R,N,M,L,A,E,q,F,P,Y,j=[],H=new Map,z,Z,ne,de,me,ke,Te,ae,fe,Ge,Je,$e,Ee=t[0]&&Do(t),pe=t[2]&&So(t);o=new ul({}),d=new Hl({}),k=new jd({}),$=new cs({}),I=new li({}),E=new Bl({});let Se=t[3].questions;const Ue=K=>K[19].id;for(let K=0;K<Se.length;K+=1){let ie=xo(t,Se,K),Ne=Ue(ie);H.set(Ne,j[K]=$o(Ne,ie))}de=new Bd({});const Re=[W1,V1],xe=[];function Pe(K,ie){return K[5]!==null?0:1}return ae=Pe(t),fe=xe[ae]=Re[ae](t),{c(){Ee&&Ee.c(),e=U(),pe&&pe.c(),n=U(),r=D("main"),i=D("header"),s=D("button"),le(o.$$.fragment),l=U(),a=D("span"),a.textContent="Back",f=U(),u=D("button"),le(d.$$.fragment),c=U(),h=D("span"),h.textContent="Import",p=U(),y=D("button"),le(k.$$.fragment),_=U(),C=D("span"),C.textContent="Export",T=U(),S=D("button"),le($.$$.fragment),w=U(),x=D("span"),x.textContent="Quiz Settings",v=U(),b=D("button"),le(I.$$.fragment),O=U(),R=D("span"),R.textContent="Play",N=U(),M=D("div"),L=D("div"),A=D("button"),le(E.$$.fragment),q=te(`\r
-        Shuffle`),P=U(),Y=D("section");for(let K=0;K<j.length;K+=1)j[K].c();Z=U(),ne=D("button"),le(de.$$.fragment),ke=U(),Te=D("div"),fe.c(),m(s,"class","btn btn--icon"),m(u,"class","btn btn--icon"),m(y,"class","btn btn--icon"),m(S,"class","btn btn--icon"),m(b,"class","btn btn--icon"),m(i,"class","header btn-row btn-row--fill svelte-1k3dryf"),A.disabled=F=t[3].questions.length<=1,m(A,"class","btn btn--icon"),m(Y,"class","questions svelte-1k3dryf"),ne.disabled=me=t[3].questions.length>=ks,m(ne,"class","btn add btn--icon-only"),m(L,"class","list svelte-1k3dryf"),m(Te,"class","editor svelte-1k3dryf"),m(M,"class","wrapper svelte-1k3dryf"),m(r,"class","main svelte-1k3dryf")},m(K,ie){Ee&&Ee.m(K,ie),W(K,e,ie),pe&&pe.m(K,ie),W(K,n,ie),W(K,r,ie),g(r,i),g(i,s),se(o,s,null),g(s,l),g(s,a),g(i,f),g(i,u),se(d,u,null),g(u,c),g(u,h),g(i,p),g(i,y),se(k,y,null),g(y,_),g(y,C),g(i,T),g(i,S),se($,S,null),g(S,w),g(S,x),g(i,v),g(i,b),se(I,b,null),g(b,O),g(b,R),g(r,N),g(r,M),g(M,L),g(L,A),se(E,A,null),g(A,q),g(L,P),g(L,Y);for(let Ne=0;Ne<j.length;Ne+=1)j[Ne]&&j[Ne].m(Y,null);g(L,Z),g(L,ne),se(de,ne,null),g(M,ke),g(M,Te),xe[ae].m(Te,null),Ge=!0,Je||($e=[J(s,"click",yr),J(u,"click",t[8]),J(y,"click",t[7]),J(S,"click",t[13]),J(b,"click",t[9]),J(A,"click",wd),Ro(z=md.call(null,Y,{items:t[3].questions,flipDurationMs:200,dropTargetStyle:{}})),J(Y,"consider",t[10]),J(Y,"finalize",t[11]),J(ne,"click",vd)],Je=!0)},p(K,[ie]){if(K[0]?Ee?(Ee.p(K,ie),ie&1&&B(Ee,1)):(Ee=Do(K),Ee.c(),B(Ee,1),Ee.m(e.parentNode,e)):Ee&&(qe(),G(Ee,1,1,()=>{Ee=null}),Fe()),K[2]?pe?(pe.p(K,ie),ie&4&&B(pe,1)):(pe=So(K),pe.c(),B(pe,1),pe.m(n.parentNode,n)):pe&&(qe(),G(pe,1,1,()=>{pe=null}),Fe()),(!Ge||ie&8&&F!==(F=K[3].questions.length<=1))&&(A.disabled=F),ie&8){Se=K[3].questions,qe();for(let Qe=0;Qe<j.length;Qe+=1)j[Qe].r();j=_r(j,ie,Ue,1,K,Se,H,Y,ii,$o,null,xo);for(let Qe=0;Qe<j.length;Qe+=1)j[Qe].a();Fe()}z&&wn(z.update)&&ie&8&&z.update.call(null,{items:K[3].questions,flipDurationMs:200,dropTargetStyle:{}}),(!Ge||ie&8&&me!==(me=K[3].questions.length>=ks))&&(ne.disabled=me);let Ne=ae;ae=Pe(K),ae===Ne?xe[ae].p(K,ie):(qe(),G(xe[Ne],1,1,()=>{xe[Ne]=null}),Fe(),fe=xe[ae],fe?fe.p(K,ie):(fe=xe[ae]=Re[ae](K),fe.c()),B(fe,1),fe.m(Te,null))},i(K){if(!Ge){B(Ee),B(pe),B(o.$$.fragment,K),B(d.$$.fragment,K),B(k.$$.fragment,K),B($.$$.fragment,K),B(I.$$.fragment,K),B(E.$$.fragment,K);for(let ie=0;ie<Se.length;ie+=1)B(j[ie]);B(de.$$.fragment,K),B(fe),Ge=!0}},o(K){G(Ee),G(pe),G(o.$$.fragment,K),G(d.$$.fragment,K),G(k.$$.fragment,K),G($.$$.fragment,K),G(I.$$.fragment,K),G(E.$$.fragment,K);for(let ie=0;ie<j.length;ie+=1)G(j[ie]);G(de.$$.fragment,K),G(fe),Ge=!1},d(K){Ee&&Ee.d(K),K&&V(e),pe&&pe.d(K),K&&V(n),K&&V(r),oe(o),oe(d),oe(k),oe($),oe(I),oe(E);for(let ie=0;ie<j.length;ie+=1)j[ie].d();oe(de),xe[ae].d(),Je=!1,je($e)}}}function Q1(t,e,n){let r,i,s,o;lt(t,dt,w=>n(3,r=w)),lt(t,Nn,w=>n(16,i=w)),lt(t,sn,w=>n(5,o=w));let l=!1,a="",f=Ki(0);lt(t,f,w=>n(4,s=w));async function u(){const w=r,x=i;console.debug("Exporting quiz to file",w.name);const v=await Qd(w,x),b=w.name+".quizler";Jd(b,v)}async function d(){const w=await Xd(".quizler");if(w!==null)try{const x=await Yd(w);_d(x),console.debug("Imported quiz file",x)}catch(x){console.error("Error while importing quiz file",x),Qt("Failed to import","Quiz file invalid or corrupted")}}function c(w){if(w.lengthComputable){const x=w.loaded/w.total*100;console.debug(`Uploading content: ${x.toFixed(0)}%`),f.set(x)}}function h(w){for(let x=0;x<w.questions.length;x++){const v=w.questions[x];if(v.text=v.text.trim(),v.text.length<1)return Qt("Empty quesiton",`Question ${x+1} text must not be empty`),!1;for(let b=0;b<v.answers.length;b++){const I=v.answers[b];if(I.value=I.value.trim(),I.value.length<1)return Qt("Empty answer",`Answer number ${b+1} of question ${x+1} must not be blank`),!1}}return!0}function p(){const w=r;if(!h(w))return;n(0,l=!0),n(1,a="Uploading"),w.name=w.name.trim(),w.text=w.text.trim();const x=i;console.debug("Creating quiz"),Iu(w,x,c).then(v=>(n(1,a="Initializing"),console.debug("Initializing game",v),nn({ty:Ht.Initialize,uuid:v}))).then(({id:v,token:b,config:I})=>{Zo({id:v,token:b,config:I,host:!0})}).catch(v=>{console.error("Failed to create",v),v instanceof Error?Qt("Failed to create",v.message):Qt("Failed to create",Oi[v])}).finally(()=>n(0,l=!1))}function y(w){ps(dt,r.questions=w.detail.items,r)}function k(w){ps(dt,r.questions=w.detail.items,r)}let _=!1;function C(w){_=w,n(2,_)}const T=()=>n(2,_=!0);function S(w,x,v,b){v[b]=w,dt.set(r)}function $(w){o=w,sn.set(o)}return[l,a,_,r,s,o,f,u,d,p,y,k,C,T,S,$]}class Y1 extends be{constructor(e){super(),ye(this,e,Q1,G1,ve,{})}}function X1(t){let e,n;return{c(){e=ce("svg"),n=ce("path"),m(n,"d","M128.00106,0 C57.3172926,0 0,57.3066942 0,128.00106 C0,184.555281 36.6761997,232.535542 87.534937,249.460899 C93.9320223,250.645779 96.280588,246.684165 96.280588,243.303333 C96.280588,240.251045 96.1618878,230.167899 96.106777,219.472176 C60.4967585,227.215235 52.9826207,204.369712 52.9826207,204.369712 C47.1599584,189.574598 38.770408,185.640538 38.770408,185.640538 C27.1568785,177.696113 39.6458206,177.859325 39.6458206,177.859325 C52.4993419,178.762293 59.267365,191.04987 59.267365,191.04987 C70.6837675,210.618423 89.2115753,204.961093 96.5158685,201.690482 C97.6647155,193.417512 100.981959,187.77078 104.642583,184.574357 C76.211799,181.33766 46.324819,170.362144 46.324819,121.315702 C46.324819,107.340889 51.3250588,95.9223682 59.5132437,86.9583937 C58.1842268,83.7344152 53.8029229,70.715562 60.7532354,53.0843636 C60.7532354,53.0843636 71.5019501,49.6441813 95.9626412,66.2049595 C106.172967,63.368876 117.123047,61.9465949 128.00106,61.8978432 C138.879073,61.9465949 149.837632,63.368876 160.067033,66.2049595 C184.49805,49.6441813 195.231926,53.0843636 195.231926,53.0843636 C202.199197,70.715562 197.815773,83.7344152 196.486756,86.9583937 C204.694018,95.9223682 209.660343,107.340889 209.660343,121.315702 C209.660343,170.478725 179.716133,181.303747 151.213281,184.472614 C155.80443,188.444828 159.895342,196.234518 159.895342,208.176593 C159.895342,225.303317 159.746968,239.087361 159.746968,243.303333 C159.746968,246.709601 162.05102,250.70089 168.53925,249.443941 C219.370432,232.499507 256,184.536204 256,128.00106 C256,57.3066942 198.691187,0 128.00106,0 Z M47.9405593,182.340212 C47.6586465,182.976105 46.6581745,183.166873 45.7467277,182.730227 C44.8183235,182.312656 44.2968914,181.445722 44.5978808,180.80771 C44.8734344,180.152739 45.876026,179.97045 46.8023103,180.409216 C47.7328342,180.826786 48.2627451,181.702199 47.9405593,182.340212 Z M54.2367892,187.958254 C53.6263318,188.524199 52.4329723,188.261363 51.6232682,187.366874 C50.7860088,186.474504 50.6291553,185.281144 51.2480912,184.70672 C51.8776254,184.140775 53.0349512,184.405731 53.8743302,185.298101 C54.7115892,186.201069 54.8748019,187.38595 54.2367892,187.958254 Z M58.5562413,195.146347 C57.7719732,195.691096 56.4895886,195.180261 55.6968417,194.042013 C54.9125733,192.903764 54.9125733,191.538713 55.713799,190.991845 C56.5086651,190.444977 57.7719732,190.936735 58.5753181,192.066505 C59.3574669,193.22383 59.3574669,194.58888 58.5562413,195.146347 Z M65.8613592,203.471174 C65.1597571,204.244846 63.6654083,204.03712 62.5716717,202.981538 C61.4524999,201.94927 61.1409122,200.484596 61.8446341,199.710926 C62.5547146,198.935137 64.0575422,199.15346 65.1597571,200.200564 C66.2704506,201.230712 66.6095936,202.705984 65.8613592,203.471174 Z M75.3025151,206.281542 C74.9930474,207.284134 73.553809,207.739857 72.1039724,207.313809 C70.6562556,206.875043 69.7087748,205.700761 70.0012857,204.687571 C70.302275,203.678621 71.7478721,203.20382 73.2083069,203.659543 C74.6539041,204.09619 75.6035048,205.261994 75.3025151,206.281542 Z M86.046947,207.473627 C86.0829806,208.529209 84.8535871,209.404622 83.3316829,209.4237 C81.8013,209.457614 80.563428,208.603398 80.5464708,207.564772 C80.5464708,206.498591 81.7483088,205.631657 83.2786917,205.606221 C84.8005962,205.576546 86.046947,206.424403 86.046947,207.473627 Z M96.6021471,207.069023 C96.7844366,208.099171 95.7267341,209.156872 94.215428,209.438785 C92.7295577,209.710099 91.3539086,209.074206 91.1652603,208.052538 C90.9808515,206.996955 92.0576306,205.939253 93.5413813,205.66582 C95.054807,205.402984 96.4092596,206.021919 96.6021471,207.069023 Z"),m(e,"fill","#FFFFFF"),m(e,"width","32px"),m(e,"height","31.25px"),m(e,"viewBox","0 0 256 250"),m(e,"version","1.1"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"preserveAspectRatio","xMidYMid")},m(r,i){W(r,e,i),g(e,n)},p:Q,i:Q,o:Q,d(r){r&&V(e)}}}class K1 extends be{constructor(e){super(),ye(this,e,null,X1,ve,{})}}function J1(t){let e,n,r;return{c(){e=ce("svg"),n=ce("path"),r=ce("path"),m(n,"d","M0 0h24v24H0z"),m(n,"fill","none"),m(r,"d","M3 17.25V21h3.75L17.81 9.94l-3.75-3.75L3 17.25zM20.71 7.04c.39-.39.39-1.02 0-1.41l-2.34-2.34c-.39-.39-1.02-.39-1.41 0l-1.83 1.83 3.75 3.75 1.83-1.83z"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 0 24 24"),m(e,"width","48px"),m(e,"height","48px"),m(e,"fill","#FFFFFF")},m(i,s){W(i,e,s),g(e,n),g(e,r)},p:Q,i:Q,o:Q,d(i){i&&V(e)}}}class eh extends be{constructor(e){super(),ye(this,e,null,J1,ve,{})}}function th(t){let e,n,r,i,s,o,l,a,f;return{c(){e=ce("svg"),n=ce("path"),r=ce("path"),i=ce("path"),s=ce("path"),o=ce("path"),l=ce("path"),a=ce("path"),f=ce("path"),m(n,"d","M209.6,119.1v-3.38q0-16.18,2.49-20.42T225.44,83q5.69-4.19,7.9-7.8a16.8,16.8,0,0,0,2.2-9q0-6.59-4.2-10.35a14.58,14.58,0,0,0-10.16-3.76q-15.74,0-17.65,22.29L179.15,71q2.7-20.1,14.65-30.46T223,30.19q17.61,0,29.56,10.07a32.29,32.29,0,0,1,12,25.69q.06,19.3-19.76,32.34-8.61,5.71-10.43,8.92t-1.8,11.84Zm27.48,34.73-27.93.06L209.08,127,237,127Z"),m(n,"transform","translate(-5.67 -30.19)"),m(r,"d","M98.54,268.74a43,43,0,0,1-9.39,1.07q-15.06,0-20.17-6.46t-5.92-16.26a95.28,95.28,0,0,1-10.63.74q-21.31,0-34-16.68T5.67,190.36q0-23.55,12.27-40.71t36.14-17.17q23.55,0,35.9,17.08t12.35,41.05q0,19.26-7,32.06T78.7,241q0,8.65,9.88,8.65,2.88,0,10-.83ZM54,154q-20.72,0-20.71,36.39,0,36.55,20.63,36.55,20.87,0,20.87-36.06Q74.75,154,54,154Z"),m(r,"transform","translate(-5.67 -30.19)"),m(i,"d","M189.43,163.44v83.15H166.71V232.43q-6.42,15.88-24.53,15.89-9.39,0-14.41-3.5a19.32,19.32,0,0,1-7.21-9.55q-2.18-6-2.18-17.5V163.44h23.8v54.33q0,11.21,9.22,11.2a13,13,0,0,0,8.15-2.67,12,12,0,0,0,4.4-5.73q.95-3,1-10.45V163.44Z"),m(i,"transform","translate(-5.67 -30.19)"),m(s,"d","M234.22,163.44v83.15H209.69V163.44Z"),m(s,"transform","translate(-5.67 -30.19)"),m(o,"d","M311.45,163.44v15.31l-37.1,50H313.1v17.87H247.64V228.72l36.68-48.57H250.69V163.44Z"),m(o,"transform","translate(-5.67 -30.19)"),m(l,"d","M326.6,246.59V134.21h24.53V246.59Z"),m(l,"transform","translate(-5.67 -30.19)"),m(a,"d","M446,209.13H392.47q-.09,2.22-.09,3.87,0,8.9,4.74,13.42A15.79,15.79,0,0,0,408.44,231q12.27,0,14.74-13.09l22.23,1.64q-9.32,28.83-37.8,28.82-12.42,0-21.11-5a39.35,39.35,0,0,1-14.33-14.53q-5.64-9.56-5.64-22.89,0-19.84,11.32-32T407,161.71A37.84,37.84,0,0,1,434.5,173Q446,184.34,446,209.13Zm-53.6-14.24h29.56q-1.15-17.13-14.08-17.13Q394,177.76,392.38,194.89Z"),m(a,"transform","translate(-5.67 -30.19)"),m(f,"d","M485.25,246.59H461.46V163.44h22.48v15.47q5.18-17.28,17.94-17.29c1.21,0,2.69.09,4.45.25v24.21q-7.49,0-11.65,2.3t-6.79,8.28q-2.64,6-2.64,16.09Z"),m(f,"transform","translate(-5.67 -30.19)"),m(e,"xmlns","http://www.w3.org/2000/svg"),m(e,"viewBox","0 0 500.66 239.62"),m(e,"fill","#FFFFFF"),m(e,"width","248"),m(e,"height","135")},m(u,d){W(u,e,d),g(e,n),g(e,r),g(e,i),g(e,s),g(e,o),g(e,l),g(e,a),g(e,f)},p:Q,i:Q,o:Q,d(u){u&&V(e)}}}class nh extends be{constructor(e){super(),ye(this,e,null,th,ve,{})}}function rh(t){let e,n,r,i,s,o,l,a,f,u,d,c,h,p,y,k,_,C,T,S,$,w,x;return i=new nh({}),l=new K1({}),c=new li({}),_=new eh({}),{c(){e=D("main"),n=D("div"),r=D("div"),le(i.$$.fragment),s=U(),o=D("a"),le(l.$$.fragment),a=te(`
-      View on GitHub`),f=U(),u=D("div"),d=D("button"),le(c.$$.fragment),h=U(),p=D("div"),p.innerHTML=`<p class="action__name svelte-1re4mfc">Join a quiz</p> 
-        <p class="action__text svelte-1re4mfc">Enter a game code and hop right in</p>`,y=U(),k=D("button"),le(_.$$.fragment),C=U(),T=D("div"),T.innerHTML=`<p class="action__name svelte-1re4mfc">Create a quiz</p> 
-        <p class="action__text svelte-1re4mfc">Create your own quiz</p>`,m(r,"class","logo svelte-1re4mfc"),m(o,"href","https://github.com/jacobtread/Quizler-v2"),m(o,"target","_blank"),m(o,"rel","noreferrer"),m(o,"title","View on Github"),m(o,"class","btn btn--icon github svelte-1re4mfc"),m(n,"class","left svelte-1re4mfc"),m(p,"class","action__body svelte-1re4mfc"),m(d,"class","action svelte-1re4mfc"),m(d,"aria-label","Join"),m(T,"class","action__body svelte-1re4mfc"),m(k,"class","action svelte-1re4mfc"),m(k,"aria-label","Create"),m(u,"class","actions svelte-1re4mfc"),m(e,"class","main svelte-1re4mfc")},m(v,b){W(v,e,b),g(e,n),g(n,r),se(i,r,null),g(n,s),g(n,o),se(l,o,null),g(o,a),g(e,f),g(e,u),g(u,d),se(c,d,null),g(d,h),g(d,p),g(u,y),g(u,k),se(_,k,null),g(k,C),g(k,T),$=!0,w||(x=[J(d,"click",Ea),J(k,"click",ka)],w=!0)},p:Q,i(v){$||(B(i.$$.fragment,v),B(l.$$.fragment,v),B(c.$$.fragment,v),B(_.$$.fragment,v),Le(()=>{$&&(S||(S=Ae(e,Me,{},!0)),S.run(1))}),$=!0)},o(v){G(i.$$.fragment,v),G(l.$$.fragment,v),G(c.$$.fragment,v),G(_.$$.fragment,v),S||(S=Ae(e,Me,{},!1)),S.run(0),$=!1},d(v){v&&V(e),oe(i),oe(l),oe(c),oe(_),v&&S&&S.end(),w=!1,je(x)}}}class ih extends be{constructor(e){super(),ye(this,e,null,rh,ve,{})}}function sh(t){let e,n;return e=new il({props:{text:"Connecting to server..."}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p:Q,i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function oh(t){let e,n,r,i;const s=[ch,uh,ah,lh],o=[];function l(a,f){return a[1].ty==Yn.Home?0:a[1].ty===Yn.Create?1:a[1].ty===Yn.Connect?2:a[1].ty===Yn.Game?3:-1}return~(e=l(t))&&(n=o[e]=s[e](t)),{c(){n&&n.c(),r=Lt()},m(a,f){~e&&o[e].m(a,f),W(a,r,f),i=!0},p(a,f){let u=e;e=l(a),e===u?~e&&o[e].p(a,f):(n&&(qe(),G(o[u],1,1,()=>{o[u]=null}),Fe()),~e?(n=o[e],n?n.p(a,f):(n=o[e]=s[e](a),n.c()),B(n,1),n.m(r.parentNode,r)):n=null)},i(a){i||(B(n),i=!0)},o(a){G(n),i=!1},d(a){~e&&o[e].d(a),a&&V(r)}}}function lh(t){let e,n;return e=new pf({props:{gameData:t[1].gameData}}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p(r,i){const s={};i&2&&(s.gameData=r[1].gameData),e.$set(s)},i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function ah(t){let e,n;return e=new vc({}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p:Q,i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function uh(t){let e,n;return e=new Y1({}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p:Q,i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function ch(t){let e,n;return e=new ih({}),{c(){le(e.$$.fragment)},m(r,i){se(e,r,i),n=!0},p:Q,i(r){n||(B(e.$$.fragment,r),n=!0)},o(r){G(e.$$.fragment,r),n=!1},d(r){oe(e,r)}}}function fh(t){let e,n,r,i,s;const o=[oh,sh],l=[];function a(f,u){return f[0]?0:1}return e=a(t),n=l[e]=o[e](t),i=new Da({}),{c(){n.c(),r=U(),le(i.$$.fragment)},m(f,u){l[e].m(f,u),W(f,r,u),se(i,f,u),s=!0},p(f,[u]){let d=e;e=a(f),e===d?l[e].p(f,u):(qe(),G(l[d],1,1,()=>{l[d]=null}),Fe(),n=l[e],n?n.p(f,u):(n=l[e]=o[e](f),n.c()),B(n,1),n.m(r.parentNode,r))},i(f){s||(B(n),B(i.$$.fragment,f),s=!0)},o(f){G(n),G(i.$$.fragment,f),s=!1},d(f){l[e].d(f),f&&V(r),oe(i,f)}}}function dh(t,e,n){let r,i;return lt(t,Gi,s=>n(0,r=s)),lt(t,vr,s=>n(1,i=s)),[r,i]}class hh extends be{constructor(e){super(),ye(this,e,dh,fh,ve,{})}}new hh({target:document.getElementById("app")});
