--- backend-v2/public/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Quizler - Quickly and easily create quiz's</title>
-    <meta name="author" content="Jacobtread" />
-    <meta
-      name="description"
-      content="A Free and easy to use real time quiz maker online."
-    />
-    <meta name="theme-color" content="#f66828" />
-    <script type="module" crossorigin src="/assets/index-aa718ec1.js"></script>
-    <link rel="stylesheet" href="/assets/index-004f90fa.css">
-  </head>
-  <body>
-    <div id="app"></div>
-    
-  </body>
-</html>
