+++ backend/src/http.rs
@@ -1,76 +1,41 @@
 use crate::{
-    game::{GameConfig, GetImageMessage},
-    games::{Games, GetGameMessage, PrepareGameMessage},
+    game::GameConfig,
+    games::Games,
     session::Session,
-    types::{ImStr, Image, NameFiltering, Question},
+    types::{GameToken, ImStr, Image, NameFiltering, Question},
 };
-use actix_multipart::{Multipart, MultipartError};
-use actix_web::{
-    get,
-    http::StatusCode,
-    post,
-    web::{self, ServiceConfig},
-    HttpRequest, HttpResponse, Responder, ResponseError,
+use axum::{
+    body::Full,
+    extract::{multipart::MultipartError, Multipart, Path, WebSocketUpgrade},
+    response::{IntoResponse, Response},
+    routing::{get, post},
+    Json, Router,
 };
-use actix_web_actors::ws::{self};
 use bytes::BytesMut;
 use embeddy::Embedded;
 use futures_util::TryStreamExt;
+use hyper::{header::CONTENT_TYPE, http::HeaderValue, Request, StatusCode};
 use log::debug;
 use serde::{Deserialize, Serialize};
 use std::{
     collections::HashMap,
+    convert::Infallible,
     fmt::Display,
-    sync::{
-        atomic::{AtomicU32, Ordering},
-        Arc,
-    },
-    time::Instant,
+    future::{ready, Ready},
+    sync::Arc,
+    task::{Context, Poll},
 };
+use tower::Service;
 use uuid::Uuid;
 
 /// Configuration function for configuring
 /// all the routes
-pub fn configure(cfg: &mut ServiceConfig) {
-    cfg.service(create_quiz);
-    cfg.service(quiz_image);
-    cfg.service(quiz_socket);
-    cfg.service(public);
-}
-
-/// Embedded assets for serving the frontend of the application
-#[derive(Embedded)]
-#[folder = "public"]
-struct Assets;
-
-#[get("/{path:.*}")]
-async fn public(path: web::Path<String>) -> impl Responder {
-    let path = path.into_inner();
-    let (file, content_type) = if let Some(file) = Assets::get(&path) {
-        let path = std::path::Path::new(&path);
-        // Find a matching content type or default to text/plain
-        let content_type = path
-            .extension()
-            .and_then(|ext| {
-                if ext == "js" {
-                    Some(mime::APPLICATION_JAVASCRIPT)
-                } else if ext == "css" {
-                    Some(mime::TEXT_CSS)
-                } else if ext == "html" {
-                    Some(mime::TEXT_HTML)
-                } else {
-                    None
-                }
-            })
-            .unwrap_or(mime::TEXT_PLAIN);
-
-        (file, content_type)
-    } else {
-        // Fallback to the index.html file for all unknown pages
-        let index = Assets::get("index.html").expect("Missing index.html from build");
-        (index, mime::TEXT_HTML)
-    };
-    HttpResponse::Ok().content_type(content_type).body(file)
+pub fn router() -> Router {
+    Router::new()
+        .route("/api/quiz", post(create_quiz))
+        .route("/api/quiz/:token/:image", get(quiz_image))
+        .route("/api/quiz/socket", get(quiz_socket))
+        .fallback_service(Assets)
 }
 
 #[derive(Deserialize)]
@@ -86,6 +51,7 @@ pub struct GameConfigUpload {
 pub enum CreateError {
     MissingConfig,
     InvalidConfig(serde_json::Error),
+    ValidationFailed,
     InvalidImageUuid(uuid::Error),
     MissingImageType(Uuid),
     Multipart(MultipartError),
@@ -99,14 +65,13 @@ struct QuizCreated {
 }
 
 /// Endpoint for creating a new quiz
-#[post("/api/quiz")]
-async fn create_quiz(mut payload: Multipart) -> Result<impl Responder, CreateError> {
+async fn create_quiz(mut payload: Multipart) -> Result<Response, CreateError> {
     // Configuration data
     let mut config: Option<GameConfigUpload> = None;
     // Map of stored uploaded images
     let mut images = HashMap::new();
 
-    while let Some(mut field) = payload.try_next().await? {
+    while let Some(mut field) = payload.next_field().await? {
         /// Cap the upload max size to 15mb
         const MAX_BUFFER_SIZE_BYTES: usize = 1024 * 1024 * 15;
 
@@ -124,7 +89,11 @@ async fn create_quiz(mut payload: Multipart) -> Result<impl Responder, CreateErr
             buffer.extend_from_slice(&chunk);
         }
 
-        let name = field.name();
+        let name = match field.name() {
+            Some(value) => value,
+            // Skip un-named fields
+            None => continue,
+        };
 
         // Handle the config
         if name == "config" {
@@ -137,8 +106,7 @@ async fn create_quiz(mut payload: Multipart) -> Result<impl Responder, CreateErr
         let uuid: Uuid = name.parse().map_err(CreateError::InvalidImageUuid)?;
         let mime = field
             .content_type()
-            .ok_or_else(|| CreateError::MissingImageType(uuid))?
-            .clone();
+            .ok_or(CreateError::MissingImageType(uuid))?;
 
         debug!(
             "Recieved uploaded file (UUID: {}, Mime: {}, Size: {})",
@@ -150,7 +118,7 @@ async fn create_quiz(mut payload: Multipart) -> Result<impl Responder, CreateErr
         images.insert(
             uuid,
             Image {
-                mime,
+                mime: mime.into(),
                 data: buffer.freeze(),
             },
         );
@@ -173,14 +141,16 @@ async fn create_quiz(mut payload: Multipart) -> Result<impl Responder, CreateErr
         images,
     };
 
-    let uuid = Games::get()
-        .send(PrepareGameMessage { config })
-        .await
-        .expect("Games service is not running");
+    // Validate the config is acceptable
+    if !config.validate() {
+        return Err(CreateError::ValidationFailed);
+    }
+
+    let uuid = Games::prepare(config).await;
 
     debug!("Created new prepared game {}", uuid);
 
-    Ok(HttpResponse::Created().json(QuizCreated { uuid }))
+    Ok((StatusCode::CREATED, Json(QuizCreated { uuid })).into_response())
 }
 
 #[derive(Debug)]
@@ -189,43 +159,85 @@ pub enum ImageError {
     UnknownImage,
 }
 
-#[get("/api/quiz/{token}/{image}")]
-async fn quiz_image(path: web::Path<(String, Uuid)>) -> Result<impl Responder, ImageError> {
-    let (token, uuid) = path.into_inner();
-
-    let game = Games::get()
-        .send(GetGameMessage { token })
+async fn quiz_image(Path((token, uuid)): Path<(String, Uuid)>) -> Result<Response, ImageError> {
+    let token: GameToken = token.parse().map_err(|_| ImageError::UnknownGame)?;
+    let game = Games::get_game(&token)
         .await
-        .expect("Games service is not running")
-        .map_err(|_| ImageError::UnknownGame)?;
+        .ok_or(ImageError::UnknownGame)?;
 
     let image = game
-        .send(GetImageMessage { uuid })
+        .read()
         .await
-        .map_err(|_| ImageError::UnknownGame)?
+        .get_image(uuid)
         .ok_or(ImageError::UnknownImage)?;
 
-    Ok(HttpResponse::Ok().content_type(image.mime).body(image.data))
+    let mut res = Full::from(image.data).into_response();
+    res.headers_mut().insert(
+        CONTENT_TYPE,
+        HeaderValue::from_str(&image.mime).expect("Failed to create mime header"),
+    );
+
+    Ok(res)
 }
 
-static SESSION_ID: AtomicU32 = AtomicU32::new(0);
+async fn quiz_socket(ws: WebSocketUpgrade) -> Response {
+    ws.on_upgrade(Session::start)
+}
 
-#[get("/api/quiz/socket")]
-async fn quiz_socket(
-    req: HttpRequest,
-    stream: web::Payload,
-) -> Result<impl Responder, actix_web::Error> {
-    let session_id = SESSION_ID.fetch_add(1, Ordering::AcqRel);
-    debug!("Starting new socket {}", session_id);
-    ws::start(
-        Session {
-            id: session_id,
-            game: None,
-            hb: Instant::now(),
-        },
-        &req,
-        stream,
-    )
+/// Embedded assets for serving the frontend of the application
+#[derive(Embedded, Clone)]
+#[folder = "public"]
+struct Assets;
+
+impl<T> Service<Request<T>> for Assets {
+    type Response = Response;
+    type Error = Infallible;
+    type Future = Ready<Result<Self::Response, Self::Error>>;
+
+    fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+
+    fn call(&mut self, req: Request<T>) -> Self::Future {
+        let path = req.uri().path();
+        // Strip the leading slash in order to match paths correctly
+        let path = match path.strip_prefix('/') {
+            Some(value) => value,
+            None => path,
+        };
+
+        let std_path = std::path::Path::new(path);
+
+        let (file, content_type) = if let Some(file) = Assets::get(path) {
+            // Find a matching content type or default to text/plain
+            let content_type = std_path
+                .extension()
+                .and_then(|ext| {
+                    if ext == "js" {
+                        Some("application/javascript")
+                    } else if ext == "css" {
+                        Some("text/css")
+                    } else if ext == "html" {
+                        Some("text/html")
+                    } else {
+                        None
+                    }
+                })
+                .unwrap_or("text/plain");
+
+            (file, content_type)
+        } else {
+            // Fallback to the index.html file for all unknown pages
+            let index = Assets::get("index.html").expect("Missing index.html from build");
+            (index, "text/html")
+        };
+
+        let mut res = Full::from(file).into_response();
+        res.headers_mut()
+            .insert(CONTENT_TYPE, HeaderValue::from_static(content_type));
+
+        ready(Ok(res))
+    }
 }
 
 impl From<MultipartError> for CreateError {
@@ -246,13 +258,14 @@ impl Display for CreateError {
             CreateError::Multipart(err) => err.fmt(f),
             CreateError::TooLarge => f.write_str("Uploaded content was too large"),
             CreateError::MissingQuestions => f.write_str("Quiz must have atleast 1 question"),
+            CreateError::ValidationFailed => f.write_str("Validation failure incorrect values"),
         }
     }
 }
 
-impl ResponseError for CreateError {
-    fn status_code(&self) -> StatusCode {
-        StatusCode::BAD_REQUEST
+impl IntoResponse for CreateError {
+    fn into_response(self) -> Response {
+        (StatusCode::BAD_REQUEST, self.to_string()).into_response()
     }
 }
 
@@ -265,10 +278,12 @@ impl Display for ImageError {
     }
 }
 
-impl ResponseError for ImageError {
-    fn status_code(&self) -> StatusCode {
+impl IntoResponse for ImageError {
+    fn into_response(self) -> Response {
         match self {
-            ImageError::UnknownGame | ImageError::UnknownImage => StatusCode::BAD_REQUEST,
+            ImageError::UnknownGame | ImageError::UnknownImage => {
+                (StatusCode::BAD_REQUEST).into_response()
+            }
         }
     }
 }
